% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalCycleGrid.R
\name{adjustTemporalCycleDates}
\alias{adjustTemporalCycleDates}
\title{Adjust temporal cycle dates}
\usage{
adjustTemporalCycleDates(grid)
}
\description{
Prepare the dates of an input grid to a suitable format for temporal cycle grids. Internally 
used by \code{\link{temporalCycleGrid}}
}
\author{
J Bedia
}
\keyword{internal}
