% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{array3Dto2Dmat.stations}
\alias{array3Dto2Dmat.stations}
\title{Conversion of a 3D array to a 2D matrix for station data}
\usage{
array3Dto2Dmat.stations(array3D)
}
\arguments{
\item{array3D}{A 3-dimensional array with longitude, latitude and time dimensions}
}
\value{
A 2-dimensional matrix with time in rows and grid-points in columns.
}
\description{
Converts 3D arrays of the form [lon,lat,time] -not strictly in this order-,
 to 2D matrices of the form [time, grid-point], in this order. Mainly for PCA analysis.
}
\seealso{
\code{\link{mat2Dto3Darray}}, which performs the inverse operation

Other internal.helpers: \code{\link{array3Dto2Dmat}},
  \code{\link{bindGrid.member}},
  \code{\link{bindGrid.spatial}},
  \code{\link{bindGrid.time}},
  \code{\link{grid3Dto2Dmatrix}},
  \code{\link{mat2Dto3Darray.stations}},
  \code{\link{redim}}, \code{\link{sortDim.spatial}},
  \code{\link{sortDim.time}}
}
\author{
M. Iturbide
}
\keyword{internal}
