% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-DataFrame.R
\name{as-DataFrame}
\alias{as-DataFrame}
\alias{coerce,sparseMatrix,DataFrame-method}
\alias{coerce,tbl_df,DataFrame-method}
\title{Coerce to \code{DataFrame}}
\description{
Coerce to \code{DataFrame}
}
\examples{
## sparseMatrix ====
load(system.file("extdata", "sparseMatrix.rda", package = "S4Transformer"))
x <- as(sparseMatrix, "DataFrame")

## tbl_df ====
load(system.file("extdata", "tbl_df.rda", package = "S4Transformer"))
x <- as(tbl_df, "DataFrame")
head(x)
}
