% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-SummarizedExperiment.R
\name{as-SummarizedExperiment}
\alias{as-SummarizedExperiment}
\alias{as.SummarizedExperiment}
\title{Coerce to \code{SummarizedExperiment}}
\usage{
as.SummarizedExperiment(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Improved method for reliably coercing objects that extend
\code{RangedSummarizedExperiment} to a standard \code{SummarizedExperiment}, that
doesn't drop \code{\link[SummarizedExperiment:rowData]{rowData()}}.
}
\examples{
load(system.file("extdata", "rse.rda", package = "S4Transformer"))
x <- as.SummarizedExperiment(rse)
}
