% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tbl_df.R
\name{as-tbl_df}
\alias{as-tbl_df}
\alias{as_tibble}
\alias{coerce,data.frame,tbl_df-method}
\alias{coerce,DataFrame,tbl_df-method}
\alias{coerce,GRanges,tbl_df-method}
\title{Coerce to \code{tbl_df} (tibble)}
\description{
Coerce to \code{tbl_df} (tibble)
}
\section{S3 \code{as_tibble()}}{


transformer extends \code{\link[tibble:as_tibble]{as_tibble()}} method support for
these S4 classes:
\itemize{
\item \code{DataFrame}.
\item \code{GRanges}.
}
}

\section{S4 \code{as()}}{


Since \code{tbl_df} is a virtual class that extends \code{tbl} and \code{data.frame}, we
need to define an S4 coercion method that allows us to use
\code{\link[methods:as]{as()}} to coerce an object to a tibble.

See \code{getClass("tbl_df")} for details on how tibble is a virtual class.
}

\examples{
load(system.file("extdata", "rse.rda", package = "transformer"))

df <- SummarizedExperiment::colData(rse)
gr <- SummarizedExperiment::rowRanges(rse)

## DataFrame to tbl_df ====
x <- as(df, "tbl_df")
x <- as_tibble(df)
print(x)

## GRanges to tbl_df ====
x <- as(gr, "tbl_df")
x <- as_tibble(gr)
print(x)
}
