% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame-s4.R
\docType{methods}
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,sparseMatrix-method}
\title{Additional S4 \code{as.data.frame()} methods}
\usage{
\S4method{as.data.frame}{sparseMatrix}(x, row.names = NULL,
  optional = FALSE, ...)
}
\description{
Additional S4 \code{as.data.frame()} methods
}
\examples{
## sparseMatrix ====
i <- c(1, 3:8)
j <- c(2, 9, 6:10)
x <- 7 * (1:7)
sm <- Matrix::sparseMatrix(i = i, j = j, x = x)
dim(sm)
summary(sm)
str(sm)
## Now you can avoid having to coerce with `as.matrix()` first.
df <- as.data.frame(sm)
}
