% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/as.SummarizedExperiment-methods.R
\docType{methods}
\name{as.SummarizedExperiment}
\alias{as.SummarizedExperiment}
\alias{as.SummarizedExperiment,SummarizedExperiment-method}
\alias{as.SummarizedExperiment,RangedSummarizedExperiment-method}
\title{Coerce to summarized experiment}
\usage{
as.SummarizedExperiment(x)

\S4method{as.SummarizedExperiment}{SummarizedExperiment}(x)

\S4method{as.SummarizedExperiment}{RangedSummarizedExperiment}(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Coerce to \code{SummarizedExperiment}.
}
\details{
Improved S4 methods for reliably coercing objects that extend
\code{RangedSummarizedExperiment} to a standard \code{SummarizedExperiment}, that
doesn't drop \code{\link[SummarizedExperiment:rowData]{rowData()}}.

Related S4 coercion method of interest:\preformatted{getMethod(
    f = "coerce",
    signature = signature(
        from = "RangedSummarizedExperiment",
        to = "SummarizedExperiment"
    ),
    where = asNamespace("SummarizedExperiment")
)
}
}
\examples{
suppressPackageStartupMessages(library(SummarizedExperiment))
data(RangedSummarizedExperiment, package = "acidtest")

## RangedSummarizedExperiment ===
## Current S4 coercion drops row metadata.
names(metadata(rowRanges(RangedSummarizedExperiment)))
se <- as(RangedSummarizedExperiment, "SummarizedExperiment")
names(metadata(rowData(se)))
## Our alternate S3 method preserves the metadata.
se <- as.SummarizedExperiment(RangedSummarizedExperiment)
names(metadata(rowData(se)))
}
