% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_df-s3.R
\name{as_tibble}
\alias{as_tibble}
\title{Additional S3 \code{as_tibble()} methods}
\description{
Additional S3 \code{as_tibble()} methods
}
\details{
S4Transform adds method support for these S4 classes:
\itemize{
\item \code{DataFrame}.
\item \code{GRanges}.
}
}
