% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomize.R
\name{atomize}
\alias{atomize}
\title{Atomize}
\usage{
atomize(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Modified object.
Non-atomic columns (e.g. \code{list} or complex S4 classes) will be dropped.
}
\description{
Atomize
}
\note{
Input object must be coercible to \code{data.frame}.
}
\examples{
df <- S4Vectors::DataFrame(a = "a", b = list(a = seq_len(3)))
lapply(df, is.atomic)
x <- atomize(df)
}
