% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attention.R
\name{attention}
\alias{attention}
\title{Attnention mechanism}
\usage{
attention(Q, K, V, mask = NULL)
}
\arguments{
\item{Q}{queries}

\item{K}{keys}

\item{V}{values}

\item{mask}{optional mask}
}
\value{
attention values
}
\description{
Attnention mechanism
}
