% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindGrid.time.R
\name{bindGrid.time}
\alias{bindGrid.time}
\title{Grid binding by time dimension}
\usage{
bindGrid.time(..., spatial.tolerance = 0.001)
}
\arguments{
\item{...}{Input grids to bind by their time dimension. These must be compatible in members and space (see details).}

\item{spatial.tolerance}{numeric. Coordinate differences smaller than \code{spatial.tolerance} will be considered equal 
coordinates. Default to 0.001 --assuming that degrees are being used it seems a reasonable rounding error after interpolation--.
This value is passed to the \code{\link{identical}} function to check for spatial consistency of the input grids.}
}
\description{
Flexible binding of (spatially/member consistent) grids by their time dimension
}
\details{
The function makes a number of checks in order to test the spatiotemporal compatibility of the input multi-member grids.
 The spatial consistency of the input grids is also checked. In order to avoid possible errors from the user, the spatial
  consistency (i.e., equal XY coordinates) of the input grids must be ensured before attempting the creation of the multigrid,
  otherwise giving an error. This can be achieved either through the specification of the same 'lonLim' and 'latLim' argument
  values when loading the grids, or using the \code{\link{interpGrid}} interpolator in conjuntion with the \code{\link{getGrid}}
  method.
}
\author{
M De Felice, J Bedia
}

