% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\name{CFS_Iberia_tas}
\alias{CFS_Iberia_tas}
\title{Multimember grid containing CFSv2 seasonal forecast data of daily mean temperature for
the Iberian Peninsula (DJF, 1983-2002).}
\format{
A multimember grid
}
\source{
\url{http://www.meteo.unican.es/ecoms-udg}
}
\description{
This R data object was obtained from the ECOMS-UDG server 
(\url{http://www.meteo.unican.es/ecoms-udg}, 
log in is requiered, see \code{\link[loadeR]{loginUDG}}) 
by means of function \code{\link[loadeR.ECOMS]{loadECOMS}} 
(package \href{https://github.com/SantanderMetGroup/loadeR.ECOMS}{\code{loadeR.ECOMS}}) 
in the following manner:

\code{years <- 1983:2002}\cr 
\code{season <- c(12,1,2)}\cr 
\code{lonLim <- c(-10, 4)}\cr 
\code{latLim <- c(35, 44)}\cr 

\code{loginUDG("username", "pasword")}\cr 

\code{CFS_Iberia_tas <- loadECOMS(dataset = "CFSv2_seasonal",
var = "tas", years = years, lonLim = lonLim, latLim = latLim, season = season,
time = "DD", aggr.d = "mean", leadMonth = 1, members = 1:9)}
}
\examples{
data(CFS_Iberia_tas)
plotClimatology(climatology(CFS_Iberia_tas, clim.fun = list(FUN = sum)),
                backdrop.theme = "coastline")
}
\references{
Saha, S. \emph{et al.}, 2013. The NCEP Climate Forecast System Version 2. J Clim 130925135638001. doi:10.1175/JCLI-D-12-00823.1
}
