% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkGrid}
\alias{checkGrid}
\title{Check the structure of the grid}
\usage{
checkGrid(..., redim = TRUE)
}
\arguments{
\item{grid}{An input grid (including station data)}
}
\value{
Diagnostic message
}
\description{
Check the structure of the grid
}
\seealso{
Other get.helpers: \code{\link{get2DmatCoordinates}},
  \code{\link{getCoordinates}}, \code{\link{getDim}},
  \code{\link{getGridVerticalLevels}},
  \code{\link{getGrid}}, \code{\link{getRefDates}},
  \code{\link{getSeason}}, \code{\link{getShape}},
  \code{\link{getStationID}},
  \code{\link{getTimeResolution}},
  \code{\link{getVarNames}}, \code{\link{getYearsAsINDEX}},
  \code{\link{typeofGrid}}, \code{\link{which.leap}}
}
\author{
M. Iturbide
}
\concept{get.helpers}
