% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkSeason}
\alias{checkSeason}
\title{Seasonal check}
\usage{
checkSeason(...)
}
\arguments{
\item{...}{Input grids to be compared}
}
\value{
In case of seasonal inconsistency of any of the inputs grids, the function stops the execution 
of the current expression, with an error message.
}
\description{
Check the consistency of seasons across the input grids
}
\examples{
data("EOBS_Iberia_tas")
data("EOBS_Iberia_pr")
grid1 <- subsetGrid(EOBS_Iberia_tas, season = 1)
grid2 <- subsetGrid(EOBS_Iberia_pr, season = 1)
checkSeason(grid1, grid2) # ok, go on
\donttest{
try(checkSeason(grid1, grid2, EOBS_Iberia_tas)) # oops
}
# However note that it is insensitive to the level of temporal aggregation:
agg <- aggregateGrid(EOBS_Iberia_tas, aggr.m = list(FUN = "mean", na.rm = TRUE))
identical(getSeason(agg), getSeason(EOBS_Iberia_tas))
checkSeason(agg, EOBS_Iberia_tas)

#Station data
data("VALUE_Iberia_tas")
data("VALUE_Iberia_pr")
checkSeason(VALUE_Iberia_tas, VALUE_Iberia_pr)
}
\seealso{
\code{\link{getSeason}}

Other check.helpers: \code{\link{checkDim}},
  \code{\link{checkVarNames}}
}
\author{
J Bedia
}
\concept{check.helpers}
\keyword{internal}
