% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkTemporalConsistency}
\alias{checkTemporalConsistency}
\title{Temporal consistency check across input grids}
\usage{
checkTemporalConsistency(...)
}
\arguments{
\item{...}{Input grids to be compared. Either as a list, or comma sepparated}
}
\value{
In case of inconsistency of any of the inputs grids, the function stops the execution 
of the current expression, with an error message.
}
\description{
Check the consistency of temporal aspects across input climate4R objects
}
\details{
The function takes into account the temporal resolution of the input grids via 
the \code{\link{getTimeResolution}} helper, and ensures its consistency. Then, it makes the 
comparison among inputs at that temporal level of resolution. For instance, if the input grids are all 
monthly, the function will intercompare only the consistency of years AND months, but will disregard
lower temporal units (days, hours).
}
\seealso{
\code{\link{checkSeason}}, for a more specific -less strict- type of check

Other check.helpers: 
\code{\link{checkDim}()},
\code{\link{checkGrid}()},
\code{\link{checkSeason}()},
\code{\link{checkVarNames}()}
}
\author{
J Bedia
}
\concept{check.helpers}
\keyword{internal}
