% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkVarNames}
\alias{checkVarNames}
\title{Variable name check}
\usage{
checkVarNames(..., check.order = TRUE)
}
\arguments{
\item{...}{Input grids to be compared}

\item{check.order}{Check if the order is consistent across input grids. This is the default. When set to
\code{TRUE}, the function does not only check that the variable names match, but also that their order is the same
(in case of multigrids)}
}
\value{
In case of inconsistency of any of the inputs grids, the function stops the execution 
of the current expression, with an error message.
}
\description{
Check the consistency of variable (short)names across grids/multigrids
}
\examples{
\donttest{ 
require(climate4R.datasets)
data("NCEP_Iberia_hus850", "NCEP_Iberia_psl", "NCEP_Iberia_ta850")
#' # Testing grids that contain different variables:
try(checkVarNames(NCEP_Iberia_hus850, NCEP_Iberia_psl))
# Testing multigrid contents:
mg <- makeMultiGrid(NCEP_Iberia_hus850, NCEP_Iberia_psl, NCEP_Iberia_ta850)
mg1 <- makeMultiGrid(NCEP_Iberia_hus850, NCEP_Iberia_ta850, NCEP_Iberia_psl)
mg2 <- makeMultiGrid(NCEP_Iberia_ta850, NCEP_Iberia_psl, NCEP_Iberia_hus850)
# The input grids contains the same variables, but their ordering is different:
checkVarNames(mg, mg1, mg2, check.order = FALSE) # OK
# If we do check.order = TRUE (the default), it gives an error:
try(checkVarNames(mg, mg1, mg2, check.order = TRUE)) # Error
}
}
\seealso{
\code{\link{getVarNames}}

Other check.helpers: 
\code{\link{checkDim}()},
\code{\link{checkGrid}()},
\code{\link{checkSeason}()},
\code{\link{checkTemporalConsistency}()}
}
\author{
J Bedia
}
\concept{check.helpers}
\keyword{internal}
