% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatologyVariogram.R
\name{climatologyVariogram}
\alias{climatologyVariogram}
\title{Empirical variogram of a climatology}
\usage{
climatologyVariogram(clim, n.classes = 20, do.log = FALSE)
}
\arguments{
\item{clim}{A climatological grid (see \code{\link{climatology}})}

\item{n.classes}{The number of distance classes used to compute the variogram}

\item{do.log}{Logical. Should the variable be log-transformed prior to variogram calculation? Default to \code{FALSE}.
Otherwise \code{\link{log1p}} is used.}
}
\value{
Plots a semivariogram and invisibly returns the data used to construct it (semivariance, distance class and number
 of pairs of points per distance class).
}
\description{
Calculate (and draw) the empirical semivariogram of a climatological grid
}
\details{
The x-axis of the variogram is tyically truncated at half the maximum distance of the dataset, so this is the default behaviour.

Note that the empirical semivariogram is used under the assumption of normality. Thus, the 
\code{do.log} option may be activated for non-gaussian fields.
}
\examples{
data("EOBS_Iberia_tp")
# We compute the mean annual DJF precipitation
aggr.fun <- list(FUN = "sum")
annual.tp <- aggregateGrid(EOBS_Iberia_tp, aggr.m = aggr.fun, aggr.y = aggr.fun)
# Now the winter precipitation climatology is computed
tp.clim <- climatology(annual.tp)
plotClimatology(tp.clim,
                backdrop.theme = "countries",
                main = "mean DJF precip (1983-2002)")
# Visual assessment of normality                
par(mfrow = c(1,2))
hist(tp.clim$Data, main = "raw field")
hist(log1p(tp.clim$Data), main = "log-transformed")
par(mfrow = c(1,1))
# Log transformation seems advisable
# The empirical variogram:
climatologyVariogram(clim = tp.clim, n.classes = 20, do.log = TRUE)
# The number of paris of points within each distance class is
# indicated by the figures on the line
}
\author{
J Bedia
}
