% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster2plot.R
\name{cluster2plot}
\alias{cluster2plot}
\title{Convert a grid of clusters into a C4R grid for plotting}
\usage{
cluster2plot(cluster, members = 1, var = getVarNames(cluster)[1])
}
\arguments{
\item{cluster}{A grid (gridded or station dataset), multigrid, multimember grid or multimember multigrid object, as 
returned by \code{clusterGrid}, with clusters in the 'time' dimension.}

\item{members}{(optional) Integer value indicating \strong{the position} of the member to be subset. Default: 1.
If input grid has no member dimension, this argument is ignored.}

\item{var}{(optional) Character vector indicating the variable name to be extracted. Default: first variable.
See \link[transformeR]{subsetGrid}.}
}
\value{
A new C4R grid object that contains the clusters from the specified variable and/or member. 
This subset grid of clusters is ready to be plotted with C4R plotting tools, e.g. \code{spatialPlot}, since clusters are intrepreted as variables (see \link[transformeR]{makeMultiGrid}).
}
\description{
This function allows to prepare a grid of clusters (included in 'time' dimension) for plotting. A subset of a variable and/or member is also performed in multimember or multigrid objects.
}
\details{

}
\examples{
#Example 1: 'cluster' is a 3D grid of clusters.
data(NCEP_Iberia_psl, package = "transformeR")
clusters<- clusterGrid(NCEP_Iberia_psl, type="kmeans", centers=10)
mg <- cluster2plot(clusters)
visualizeR::spatialPlot(mg, backdrop.theme = "coastline", rev.colors = TRUE, layout = c(2,ceiling(attr(clusters, "centers")/2)), as.table=TRUE)

#Example 2: 'cluster' is a grid of clusters with two variables (same for members). 
data(NCEP_Iberia_ta850, package = "transformeR")
clusters<- clusterGrid(makeMultiGrid(NCEP_Iberia_psl, NCEP_Iberia_ta850), type="kmeans", centers=10, iter.max=1000)
mg <- cluster2plot(clusters, members=1, var="psl") #To obtain first member and variable called "psl"
visualizeR::spatialPlot(mg, backdrop.theme = "coastline", rev.colors = TRUE, layout = c(2,ceiling(attr(clusters, "centers")/2)), as.table=TRUE)
}
\seealso{
\link[transformeR]{makeMultiGrid}, \link[transformeR]{subsetGrid}, \link[transformeR]{subsetDimension}.
}
\author{
J. A. Fernandez
}
