% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGrid.R
\name{clusterGrid}
\alias{clusterGrid}
\title{Cluster analysis of grids}
\usage{
clusterGrid(grid, type = "kmeans", centers = NULL, newdata = NULL,
  y = NULL, ...)
}
\arguments{
\item{grid}{A grid (gridded or station dataset), multigrid, multimember grid or multimember multigrid object, as 
returned e.g. by \code{loadeR::loadGridData} (or \code{loadeR::loadStationData}), a
multigrid, as returned by \code{makeMultiGrid}, or other types of multimember grids
(possibly multimember grids) as returned e.g. by \code{loadeR.ECOMS::loadECOMS}.}

\item{type}{Clustering algorithm to be used for the cluster analysis. 
Possible values are "\strong{kmeans}" (default), "\strong{hierarchical}", "\strong{som}". 
The core functions are \link[stats]{kmeans}, \link[stats]{hclust}, \link[kohonen]{som}, respectively. See Details.}

\item{centers}{Integer value indicating the number of clusters, \strong{k}, or center points. See Details.}

\item{newdata}{A grid containing the prediction data. It must contain the same variables as in "grid". Clustering Analysis of this 
grid will be performed taking into account the Clustering of 'grid' as reference. Default: NULL.}

\item{y}{A grid containing predictands data. Clustering Analysis of this grid will be performed as a day-by-day correspondence with 
the reference grid ('grid' or 'newdata'), thus time-dimension from 'y' and the reference grid must intersect. Default: NULL.}

\item{...}{Further specific arguments passed to the clustering functions}
}
\value{
A C4R (multimember/multi)grid object that will contain data from: 
\item 'grid' input if none 'newdata' or 'y' are passed.  
\item 'newdata' if input 'newdata' is passed and 'y' is not. 
\item 'y' if this predictands grid is passed. 
The clustering type (cluster.type), number of clusters (centers), centroids of clusters from 'grid' input (centroids) and the days corresponding to each 
cluster (index) are returned as attributes in all clustering algorithms. Then, other algorithm-specific parameters are provided as attributes.
}
\description{
Performs cluster analysis of grids, multigrids or multimember multigrids. Several clustering algorithms are available.
}
\details{
\strong{kmeans}

While using the  K-means algorithm, the number of clusters (argument 'centers') needs to be provided (no default). 
See  \link[stats]{kmeans} for more details in the implementation.

\strong{hierarchical}

While using the hierarchical algorithm (check \link[stats]{hclust} for further information) 
\code{clusterGrid} allows the user either to especify the number of clusters ('centers') or not. 
If the argument 'centers' is not provided, they are automatically set and the tree is cut when the height 
difference between two consecutive divisions (sorted in ascending order) is larger than the interquartile 
range of the heights vector (see \link[stats]{cutree}) . 

\strong{som}

While using the SOM (self-organizing maps) algorithm (check \link[kohonen]{som} for further information), the argument 'centers' is provided as
a two-element vector, indicating the dimensions \code{xdim,ydim} of the grid (see \link[kohonen]{somgrid}).
Otherwise, by default 48 clusters (8x6) with rectangular topology are obtained. 

#'\strong{Lamb}

While using the Lamb Weather Typing algorithm (check \link{LambWT} for further information), the argument 'centers' is not requiered, as it will 
be forced to be equal to 27, according to Jones et al. 2012 (Int J Climatol).
}
\examples{
#Example of K-means clustering: 
data(NCEP_Iberia_psl, package = "transformeR")
clusters<- clusterGrid(NCEP_Iberia_psl, type="kmeans", centers=10, iter.max=1000)

#Example of hierarchical clustering: 
clusters<- clusterGrid(NCEP_Iberia_psl, type="hierarchical")

#Example of som clustering: 
clusters<- clusterGrid(NCEP_Iberia_psl, type="som", centers = c(10,1))

#Example of lamb clustering:
data(NCEP_slp_2001_2010)
clusters <- clusterGrid(grid = NCEP_slp_2001_2010, type = "lamb")

#Example of 'newdata' clustering:
grid <- makeMultiGrid(CMIP5_Iberia_pr, CMIP5_Iberia_tas, CMIP5_Iberia_hus850)
newdata <- makeMultiGrid(CMIP5_Iberia_pr.rcp85, CMIP5_Iberia_tas.rcp85, CMIP5_Iberia_hus850.rcp85)
clusters <- clusterGrid(grid = grid, newdata = newdata, type = "kmeans", centers = 100, iter.max = 10000, nstart = 1)
}
\seealso{
\link[stats]{kmeans}, \link[stats]{hclust}, \link[kohonen]{som}.
}
\author{
J. A. Fernandez
}
