% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGrid.R
\name{clusterGrid_2D}
\alias{clusterGrid_2D}
\title{Cluster analysis of 2D matrix}
\usage{
clusterGrid_2D(grid.2D, type, centers, ...)
}
\arguments{
\item{type}{Clustering algorithm to be used for the cluster analysis. 
Possible values are "\strong{kmeans}" (default), "\strong{hierarchical}", "\strong{som}". 
The core functions are \link[stats]{kmeans}, \link[stats]{hclust}, \link[kohonen]{som}, respectively. See Details.}

\item{centers}{Integer value indicating the number of clusters, \strong{k}, or center points. See Details.}

\item{...}{Further specific arguments passed to the clustering functions}

\item{grid}{A grid (gridded or station dataset), multigrid, multimember grid or multimember multigrid object}
}
\value{
A matrix object that contains the clusters centroids' created using the specified algorithm.
The clustering type, number of clusters, cluster index of each day and other algorithm-specific parameters are 
provided as attributes.
}
\description{
Performs cluster analysis of 3D grids. Several clustering algorithms are available.
}
\author{
J. A. Fernandez
}
\keyword{internal}
