% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGrid.R
\name{clusterGrid_3D}
\alias{clusterGrid_3D}
\title{Cluster analysis of 3D grids}
\usage{
clusterGrid_3D(grid, type, centers, iter.max, nstart, method)
}
\arguments{
\item{grid}{A grid (gridded or station dataset), multigrid, multimember grid or multimember multigrid object, as 
returned e.g. by \code{loadeR::loadGridData} (or \code{loadeR::loadStationData}), a
multigrid, as returned by \code{makeMultiGrid}, or other types of multimember grids
(possibly multimember grids) as returned e.g. by \code{loadeR.ECOMS::loadECOMS}.}

\item{type}{Clustering algorithm to be used for the cluster analysis. 
Possible values are "\strong{kmeans}" (default), "\strong{hierarchical}", "\strong{som}". 
The core functions are \link[stats]{kmeans}, \link[stats]{hclust}, \link[kohonen]{som}, respectively. See Details.}

\item{centers}{Integer value indicating the number of clusters, \strong{k}, or center points. See Details.}

\item{iter.max}{(for the K-means algorithm) Integer value indicating the maximum number of iterations allowed. Default: 10.}

\item{nstart}{(for the K-means algorithm) If centers is a number, how many random sets should be chosen? Default: 1.}

\item{method}{(for the hierarchical algorithm) Agglomeration method to be used, one of "complete" (default), "ward.D", "ward.D2", "single",
"average", "mcquitty", "median" or "centroid".}
}
\value{
A new 3D grid object that contains the clusters created using the specified algorithm.
The clustering type, number of clusters and other algorithm-specific parameters are provided as attributes.
}
\description{
Performs cluster analysis of 3D grids. Several clustering algorithms are available.
}
\keyword{internal}
