% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{CMIP5_Iberia_hus850.rcp85}
\alias{CMIP5_Iberia_hus850.rcp85}
\title{Grid containing CMIP5 data (CNRM) of daily specific humidity at 850mb for the Iberian Peninsula (DJF, 2081-2100).}
\format{A grid}
\source{
\url{http://www.meteo.unican.es/udg-tap}
}
\description{
This R data object was obtained from the UDG server (\url{http://www.meteo.unican.es/udg-tap}, 
log in is requiered, 
see \code{\link[loadeR]{loginUDG}}) by means of function \code{\link[loadeR]{loadGridData}} 
(package \href{https://github.com/SantanderMetGroup/loadeR}{\code{loadeR}}) in the following manner:


\code{years <- 2081:2100}\cr 
\code{season <- c(12,1,2)}\cr 
\code{lonLim <- c(-10, 4)}\cr 
\code{latLim <- c(35, 44)}\cr 


\code{loginUDG("username", "pasword")}\cr 
 
\code{CMIP5_Iberia_hus850.rcp85 <- loadGridData(dataset = "http://meteo.unican.es/tds5/dodsC/cmip5/CNRM-CERFACS/CNRM-CM5/rcp85/day/cnrm-cerfacs_cnrm-cm5_rcp85_r1i1p1.ncml",
                    var = "hus@85000", years = years, lonLim = lonLim, latLim = latLim, season = season,
                     time = "DD", aggr.d = "mean")}\cr
}
\examples{
data("CMIP5_Iberia_hus850.rcp85")
plotClimatology(climatology(CMIP5_Iberia_hus850.rcp85),
                backdrop.theme = "countries", scales = list(draw = TRUE))
}
\seealso{
\code{\link{makeMultiGrid}}, \code{\link[loadeR]{loadGridData}}
}
\keyword{Gridded}
\keyword{projections}
