% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame-methods.R,
%   R/coerce-data.frame-methods.R
\docType{methods}
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,IPosRanges-method}
\alias{as.data.frame,Matrix-method}
\alias{coerce-data.frame}
\alias{coerce,IPosRanges,data.frame-method}
\alias{coerce,Matrix,data.frame-method}
\title{Coerce to data frame}
\usage{
as.data.frame(x, row.names = NULL, optional = FALSE, ...)

\S4method{as.data.frame}{IPosRanges}(x, row.names = NULL,
  optional = FALSE, ...)

\S4method{as.data.frame}{Matrix}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{any \R object.}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link{make.names}}) is optional.  Note that all of \R's
    \pkg{base} package \code{as.data.frame()} methods use
    \code{optional} only for column names treatment, basically with the
    meaning of \code{\link{data.frame}(*, check.names = !optional)}.
    See also the \code{make.names} argument of the \code{matrix} method.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
\code{data.frame}.
}
\description{
Coerce to \code{data.frame}.
}
\examples{
data(sparseMatrix, package = "acidtest")

## Matrix to data.frame ====
x <- as(sparseMatrix, "data.frame")
head(x)
}
