% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-data.table.R
\name{coerce-data.table}
\alias{coerce-data.table}
\alias{as.data.table}
\alias{coerce,data.frame,data.table-method}
\alias{coerce,DataFrame,data.table-method}
\alias{coerce,GRanges,data.table-method}
\title{Coerce to \code{data.table}}
\description{
Coerce to \code{data.table}
}
\section{S3 \code{as.data.table()}}{


transformer extends \code{\link[data.table:as.data.table]{as.data.table()}} method
support for these S4 classes:
\itemize{
\item \code{DataFrame}.
\item \code{GRanges}.
}
}

\section{S4 \code{as()}}{


Since \code{data.table} is a class that extends \code{data.frame}, we need to define an
S4 coercion method that allows us to use \code{\link[methods:as]{as()}} to coerce an
object to a \code{data.table}.

See \code{getClass("data.table")} for details.
}

\examples{
load(system.file("extdata", "rse.rda", package = "transformer"))

df <- SummarizedExperiment::colData(rse)
gr <- SummarizedExperiment::rowRanges(rse)

## DataFrame to data.table ====
x <- as(df, "data.table")
x <- as.data.table(df)
print(x)

## GRanges to data.table ====
x <- as(gr, "data.table")
x <- as.data.table(gr)
print(x)
}
