% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-SummarizedExperiment.R
\name{coerce-SummarizedExperiment}
\alias{coerce-SummarizedExperiment}
\alias{as.SummarizedExperiment}
\title{Coerce to \code{SummarizedExperiment}}
\usage{
as.SummarizedExperiment(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Improved method for reliably coercing objects that extend
\code{RangedSummarizedExperiment} to a standard \code{SummarizedExperiment}, that
doesn't drop \code{\link[SummarizedExperiment:rowData]{rowData()}}.
}
\examples{
data(rse, package = "acidtest")

## SummarizedExperiment ====
x <- as.SummarizedExperiment(rse)
x
}
