% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-tbl_df.R
\name{coerce-tbl_df}
\alias{coerce-tbl_df}
\alias{as_tibble}
\alias{as_tibble.DataFrame}
\alias{as_tibble.GRanges}
\alias{as_tibble.IRanges}
\alias{coerce,data.frame,tbl_df-method}
\alias{coerce,DataFrame,tbl_df-method}
\alias{coerce,GRanges,tbl_df-method}
\alias{coerce,IRanges,tbl_df-method}
\title{Coerce to tibble}
\usage{
\method{as_tibble}{DataFrame}(x, ...,
  rownames = pkgconfig::get_config("tibble::rownames", "rowname"))

\method{as_tibble}{GRanges}(x, ...,
  rownames = pkgconfig::get_config("tibble::rownames", "rowname"))

\method{as_tibble}{IRanges}(x, ...,
  rownames = pkgconfig::get_config("tibble::rownames", "rowname"))
}
\description{
Coerce to \code{tbl_df}.
}
\details{
Our defined methods attempt to improve on the defaults in the tibble package
to ensure that row names are not dropped by default, which is a poor default
for bioinformatics. This is accomplished by setting \code{rownames = "rowname"} by
default instead of \code{rownames = NULL}.
}
\section{S3 \code{as_tibble()}}{


transformer extends \code{\link[tibble:as_tibble]{as_tibble()}} method support for
these S4 classes:
\itemize{
\item \code{DataFrame}.
\item \code{GRanges}.
}
}

\section{S4 \code{as()}}{


Since \code{tbl_df} is a virtual class that extends \code{tbl} and \code{data.frame}, we
need to define an S4 coercion method that allows us to use
\code{\link[methods:as]{as()}} to coerce an object to a tibble.

See \code{getClass("tbl_df")} for details on how tibble is a virtual class.
}

\examples{
data(df, gr, ir, package = "acidtest")

## DataFrame to tbl_df ====
x <- as(df, "tbl_df")
x <- as_tibble(df)
print(x)

## GRanges to tbl_df ====
x <- as(gr, "tbl_df")
x <- as_tibble(gr)
print(x)

## IRanges to tbl_df ====
x <- as(ir, "tbl_df")
x <- as_tibble(ir)
print(x)
}
\seealso{
\code{\link[tibble:as_tibble]{tibble::as_tibble()}}.
}
