% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerceS4ToList.R
\name{coerceS4ToList}
\alias{coerceS4ToList}
\title{Coerce any S4 object to a list}
\usage{
coerceS4ToList(from)
}
\arguments{
\item{from}{The classes between which the coerce methods
    \code{def} and \code{replace} perform coercion.
  }
}
\value{
\code{list}.
}
\description{
Coerce any S4 object to a list
}
\examples{
load(system.file("extdata", "rse.rda", package = "transformer"))

x <- coerceS4ToList(rse)
class(x)
names(x)
}
\seealso{
\code{methods::coerce()}.
}
