% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGrid.R
\name{comb.vars}
\alias{comb.vars}
\title{Variables combination}
\usage{
comb.vars(grid, base, ref, var.names)
}
\arguments{
\item{grid}{A grid containing the variables to be scaled and combined}

\item{base}{Reference baseline grid to be passed to \code{'scaleGrid'}. See \link{scaleGrid} for further information.}

\item{ref}{Reference grid to be passed to \code{'scaleGrid'}. See \link{scaleGrid} for further information.}

\item{var.names}{Variables names}
}
\value{
A list
}
\description{
Scaling and combining data from all variables
}
\author{
J. A. Fernandez
}
\keyword{internal}
