% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prinComp.R
\name{combine.PCs}
\alias{combine.PCs}
\title{Field Combination}
\usage{
combine.PCs(Xsc.list, which)
}
\arguments{
\item{Xsc.list}{A nested list of (scaled and centered) input variables (as returned by \code{\link{prinComp.scale}})}

\item{which}{Integer vector of indices of variables to produced the combined element}
}
\value{
A list like the input \code{Xsc.list} but with an additional element, resulting from the combination of the
input fields
}
\description{
Combine different fields into one single input 2D matrix for PCA (internal of \code{\link{prinComp}})
}
\author{
J Bedia
}
\keyword{internal}
