% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeUncertainty.R
\name{computeUncertainty}
\alias{computeUncertainty}
\title{Calculate the uncertainty of the climate change signal.}
\usage{
computeUncertainty(anomaly, historical = NULL, method = "simple")
}
\arguments{
\item{anomaly}{A grid of the climate change signal/anomaly (time dimension = 1).}

\item{historical}{A grid with annual historical data. Applicable when method = "advanced".
Default is NULL.}

\item{method}{"simple" or "advanced".}
}
\value{
A grid with three values (uncertainty categories); 0 = robust signal, 
1 = no signal or no change, 2 = conflicting signals).
}
\description{
Calculate the uncertainty of the climate change signal.
}
\author{
M. Iturbide
}
