% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2bin.R
\name{convert2bin}
\alias{convert2bin}
\title{Convert grid values to a binary variable}
\usage{
convert2bin(x, threshold = NULL, partial = FALSE, ref.obs = NULL,
  ref.pred = NULL)
}
\arguments{
\item{x}{The input grid to be converted to binary. It can be an object or a matrix/array.}

\item{threshold}{Upon or equal the threshold the grid values turn to 1, whereas below it turns to 0. This can be a single value or a vector. In case it is a vector it will apply a different threshold on each station. If threshold is NULL, then the data is converted to binary by a value that makes the ref.pred to have the same number of 0 values than those in the ref.obs. Default is NULL.}

\item{partial}{If TRUE, only values below the threshold will turn to 0 and the others will remain with their original grid real value. Default is FALSE.}

\item{ref.obs}{Grid of observations. It is used to determine the number of days where there is no event (or value equal to 0).}

\item{ref.pred}{Grid of predictions- It is used to calculate the threshold that will further be used to convert to binary the grid x.}
}
\value{
A new grid object with binary values
}
\description{
Convert a real variable into a binary variable (i.e., 0 or 1) filtering by a threshold.
}
\details{
The function only works for a vector of observations/predictands downloaded from \pkg{loadeR}).
}
\examples{
# Take a look at the data:
head(VALUE_Iberia_pr$Data)
# Convert to complete binary variable:
ybin <- convert2bin(VALUE_Iberia_pr,threshold = 1)
head(ybin$Data)
# Convert to partial binary variable:
ybin2 <- convert2bin(VALUE_Iberia_pr,threshold = 1, partial = TRUE)
head(ybin2$Data)
}
\seealso{
Other downscaling.helpers: \code{\link{filterNA}},
  \code{\link{makeMultiGrid}}
}
\author{
J. Bano-Medina
}
