% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{CORDEX_Iberia_pr}
\alias{CORDEX_Iberia_pr}
\title{Grid containing historical CORDEX data (model KNMI-RACMO22E) of daily precipitation for the Iberian Peninsula (DJF, 1983-2002).}
\format{
A grid
}
\source{
\url{http://www.meteo.unican.es/udg-tap}
}
\usage{
CORDEX_Iberia_pr
}
\description{
This R data object was obtained from the UDG server (\url{http://www.meteo.unican.es/udg-tap}, 
log in is requiered, 
see \code{\link[loadeR]{loginUDG}}) by means of function \code{\link[loadeR]{loadGridData}} 
(package \href{https://github.com/SantanderMetGroup/loadeR}{\code{loadeR}}) in the following manner:


\code{years <- 1983:2002}\cr 
\code{season <- c(12,1,2)}\cr 
\code{lonLim <- c(-10, 4)}\cr 
\code{latLim <- c(35, 44)}\cr 


\code{loginUDG("username", "pasword")}\cr 
 
\code{CORDEX_Iberia_pr <- loadGridData(dataset = "http://www.meteo.unican.es/tds5/dodsC/cordex/EUR-44/KNMI/ICHEC-EC-EARTH_historical_r1i1p1_RACMO22E_v1_day.ncml",
var = "pr", years = years, lonLim = lonLim, latLim = latLim, season = season)}\cr
\code{CORDEX_Iberia_pr$Data <- CORDEX_Iberia_pr$Data * 86400}\cr 
\code{attr(CORDEX_Iberia_pr$Variable, "units") <- "mm"}\cr
}
\keyword{Gridded}
\keyword{observations}
