% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_split.R
\name{data_split}
\alias{data_split}
\title{Split data into two different sets.}
\usage{
data_split(x, y, fraction = 3/4)
}
\arguments{
\item{x}{The input grid object.}

\item{y}{The observations object.}

\item{fraction}{A value between (0,1) that indicates the fraction of the data that will define the train set. The unselected data will be the test set. Default is 3/4.}
}
\value{
A list containing the grids x and y splitted as well as the indices in the original grids.
}
\description{
Split data into two different sets by a specific fraction. Splitting data is normally used to obtain a train and a validation set.
}
\details{
The function internally intersect datasets x and y in such a manner that only the observations present in both sets are then splitted, by \code{\link[transformeR]{getTemporalIntersection}}. It also removes those days were there is no value in at least 1 station by using \code{\link[transformeR]{filterNA}}).
}
\examples{
x <- makeMultiGrid(NCEP_Iberia_hus850, NCEP_Iberia_psl, NCEP_Iberia_ta850)
y <- VALUE_Iberia_pr
# Take a look at the dimensions of the data:
dim(VALUE_Iberia_pr$Data)
# Split the data 
data.splitted <- data_split(x,y,fraction = 0.5)
str(data.splitted)
# Take a look at the dimensions of the pbservations of the train set:
dim(data.splitted$yT$Data)
}
\seealso{
Other subsetting: \code{\link{filterNA}},
  \code{\link{getTemporalIntersection}},
  \code{\link{subsetDimension}}, \code{\link{subsetGrid}},
  \code{\link{subsetMembers}}, \code{\link{subsetRuntime}},
  \code{\link{subsetSeason}}, \code{\link{subsetSpatial}},
  \code{\link{subsetVar}}, \code{\link{subsetYears}}
}
\author{
J. Bano-Medina
}
