% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-methods.R
\docType{methods}
\name{decode}
\alias{decode}
\alias{decode,DataFrame-method}
\alias{decode,Ranges-method}
\alias{decode,SummarizedExperiment-method}
\title{Decode column data that uses run-length encoding}
\usage{
decode(x, ...)

\S4method{decode}{DataFrame}(x)

\S4method{decode}{Ranges}(x)

\S4method{decode}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Columns will be decoded and no longer \code{Rle} class.
}
\description{
The Rle class is a general container for storing an atomic vector
  that is stored in a run-length encoding format. It is based on the
  \code{\link[base]{rle}} function from the base package.
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")

## DataFrame ====
df <- SummarizedExperiment::rowData(RangedSummarizedExperiment)
lapply(df, class)
x <- decode(df)
lapply(x, class)

## SummarizedExperiment ====
## This works on rowData and colData.
x <- decode(RangedSummarizedExperiment)
lapply(rowData(x), class)
}
\seealso{
\code{\link[S4Vectors:decode]{S4Vectors::decode()}}.
}
