% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{relevel}
\alias{relevel.DataFrame}
\alias{relevel.Ranges}
\alias{relevel.SummarizedExperiment}
\alias{left_join}
\title{Deprecated functions}
\usage{
relevel(x, ref, ...)

\method{relevel}{DataFrame}(x, ref = NULL, ...)

\method{relevel}{Ranges}(x, ref = NULL, ...)

\method{relevel}{SummarizedExperiment}(x, ref = NULL, ...)

left_join(...)
}
\arguments{
\item{x}{Object.}

\item{ref}{Unsupported.}

\item{...}{Additional arguments.}
}
\value{
\code{.Deprecated}.
}
\description{
Deprecated functions
}
\examples{
a <- function(...) {
    .Deprecated("b")
    b(...)
}

b <- function(x) {
    x + 1L
}

withCallingHandlers(
    expr = a(1L),
    warning = function(w) {
        print(w)
        invokeRestart("muffleWarning")
    }
)
}
\seealso{
\link[base:Deprecated]{Deprecated}.
}
\keyword{internal}
