% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplevels-S4methods.R
\docType{methods}
\name{droplevels}
\alias{droplevels}
\alias{droplevels,Ranges-method}
\alias{droplevels,SummarizedExperiment-method}
\title{Drop unused levels from factors}
\usage{
\S4method{droplevels}{Ranges}(x)

\S4method{droplevels}{SummarizedExperiment}(x)
}
\value{
Modified object.
}
\description{
The function \code{droplevels} is used to drop unused levels from a
  \code{\link{factor}} or, more commonly, from factors in a data frame.
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")

## Ranges ====
object <- rowRanges(RangedSummarizedExperiment)
droplevels(object)

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
}
