% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyVeri2grid.R
\name{easyVeri.signif}
\alias{easyVeri.signif}
\title{Identify points significantly different from a given ref value}
\usage{
easyVeri.signif(ref.mean, est.mean, sigma, cl = 0.95)
}
\arguments{
\item{ref.mean}{Numeric of length one. Null hypothesis reference mean (e.g. 0.5 for AUC tests, 0 for skill scores)}

\item{est.mean}{easyVerification output matrix with scores/measures. This is the \code{cat[num]} element of the \code{veriApply} output}

\item{sigma}{Standard deviation of the estimated mean. This is the \code{cat[num].sigma} element of the \code{veriApply} output.}

\item{cl}{Confidence level. Default to 0.95}
}
\value{
A easyVerification-like binary matrix (1 significant / 0 non-significant).
 Ready to be passed to \code{\link{easyVeri2grid}} for climatological grid conversion, 
 and possibly to \code{\link{map.stippling}} after that, to depict significant points in a verification map.
}
\description{
Two-tailed hypothesis testing for \pkg{easyVerification} sigma outputs
}
\seealso{
\code{\link{easyVeri2grid}}, for conversion of verification matrices into climatological grids.
 Also, \code{\link{map.stippling}}, to draw significant points on verification maps.

Other easyVerification.bridges: \code{\link{easyVeri2grid}}
}
\author{
J Bedia
}
