% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOF2clim.R
\name{EOF2clim}
\alias{EOF2clim}
\title{PCA object EOFs to climatology grid}
\usage{
EOF2clim(prinCompObj, ind.var, member = 1L, n.eofs = 1L, rescale = TRUE)
}
\arguments{
\item{prinCompObj}{PCA object}

\item{ind.var}{index position of the target variable}

\item{member}{Integer. index position of the member whose EOFs are to be displayed.
Default to 1 (ignored if no members available)}

\item{n.eofs}{Integer. Number of EOFs to be displayed (from 1 to \code{n.eofs}). Default to one -first- EOF.}

\item{rescale}{Logical flag. Default to \code{TRUE}, which rescales the EOF to the original input variable units.}
}
\value{
A climatological grid. Note that EOFs are treated as members
}
\description{
Converts a \code{prinComp} output to a climatology of EOFs.
}
\examples{
\dontrun{
require(climate4R.datasets)
data("NCEP_Iberia_hus850", "NCEP_Iberia_psl", "NCEP_Iberia_ta850")
multigrid <- makeMultiGrid(NCEP_Iberia_hus850, NCEP_Iberia_psl, NCEP_Iberia_ta850)
# In this example, we retain the PCs explaining the 99\\% of the variance
pca <- prinComp(multigrid, v.exp = c(.95,0.90,.90), keep.orig = FALSE)
require(visualizeR)
# Original EOF
EOF2clim(pca, ind.var = "psl") \%>\% spatialPlot(backdrop.theme = "coastline")
# Rescaled EOF
EOF2clim(pca, ind.var = "psl", rescale = FALSE) \%>\% spatialPlot(backdrop.theme = "coastline")
}
}
\author{
J Bedia
}
