% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize.R
\name{factorize}
\alias{factorize}
\title{Factorize}
\usage{
factorize(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Modified object.
All columns will be coerced to \code{factor}.
}
\description{
Factorize
}
\examples{
df <- S4Vectors::DataFrame(a = letters[seq_len(5)], b = seq_len(5))
x <- factorize(df)
}
\seealso{
\code{encode()} for Rle approach.
}
