% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed_forward.R
\name{feed_forward}
\alias{feed_forward}
\title{Feed Forward Layer}
\usage{
feed_forward(x, dff, d_model)
}
\arguments{
\item{x}{inputs}

\item{dff}{dimensions of feed-forward model}

\item{d_model}{dimensions of the model}
}
\value{
output of the feed-forward layer
}
\description{
Feed Forward Layer
}
