% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillGridDates.R
\name{fillGridDates}
\alias{fillGridDates}
\title{Fill missing dates}
\usage{
fillGridDates(grid, tz = "")
}
\arguments{
\item{grid}{grid or station data}

\item{tz}{Otptional. Time zone (by default CET is used, see e.g. \code{\{as.POSIXlt}).}
}
\description{
fill with NA missing dates grids and station data
}
\author{
M Iturbide
}
