% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillGridDates.R
\name{fillGridSpatial}
\alias{fillGridSpatial}
\title{Fill missing dates}
\usage{
fillGridSpatial(grid, lonLim = c(-180, 180), latLim = c(-90, 90))
}
\arguments{
\item{grid}{grid or station data}

\item{tz}{Optional. Time zone (by default CET is used, see e.g. \code{\{as.POSIXlt}).}
}
\value{
A grid filled with NAs in the previously missing date positions
}
\description{
fill with NA missing dates in grids and station datasets
}
\author{
J. Baño-Medina
}
