% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get2DmatCoordinates}
\alias{get2DmatCoordinates}
\title{Get grid coordinates as 2D matrix}
\usage{
get2DmatCoordinates(grid)
}
\arguments{
\item{grid}{An input grid}
}
\value{
A 2D matrix of x-y coordinates (in this order)
}
\description{
Obtain grid coordinates as 2D matrix
}
\seealso{
Other get.helpers: \code{\link{getCoordinates}},
  \code{\link{getDim}},
  \code{\link{getGridVerticalLevels}},
  \code{\link{getGrid}}, \code{\link{getRefDates}},
  \code{\link{getSeason}}, \code{\link{getShape}},
  \code{\link{getTimeResolution}},
  \code{\link{getVarNames}}, \code{\link{getYearsAsINDEX}},
  \code{\link{typeofGrid}}, \code{\link{which.leap}}
}
\author{
J. Bedia
}
\concept{get.helpers}
\keyword{internal}
