% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getCoordinates}
\alias{getCoordinates}
\title{Get geographical coordinates of a climate data object}
\usage{
getCoordinates(obj)
}
\arguments{
\item{obj}{Any object extending the station or grid classes}
}
\value{
A list with x and y components
}
\description{
Returns the coordinates of a climate data object, either stations or grid
}
\seealso{
Other get.helpers: \code{\link{getDim}},
  \code{\link{getGrid}}, \code{\link{getRefDates}},
  \code{\link{getSeason}}, \code{\link{getShape}},
  \code{\link{getYearsAsINDEX}}, \code{\link{which.leap}}
}
\author{
J. Bedia
}
