% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getDim}
\alias{getDim}
\title{Retrieve dimensions attribute}
\usage{
getDim(obj)
}
\arguments{
\item{obj}{A grid or station object}
}
\value{
A character vector with the dimensions attribute of the object's \code{Data} component.
}
\description{
Retrieve dimensions attribute
}
\seealso{
Other get.helpers: \code{\link{getCoordinates}},
  \code{\link{getGrid}}, \code{\link{getRefDates}},
  \code{\link{getSeason}}, \code{\link{getShape}},
  \code{\link{getYearsAsINDEX}}, \code{\link{which.leap}}
}
\author{
J. Bedia
}
\keyword{internal}
