% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillGridDates.R
\name{getFirstChar}
\alias{getFirstChar}
\title{Get first character string before punctuation or space}
\usage{
getFirstChar(x)
}
\arguments{
\item{x}{character string}
}
\value{
First part of a character string before any punctuation or space character
}
\description{
Utility for parsing date elements as character strings
}
\examples{
\dontrun{ 
 getFirstChar("2008 Feb 12")
 getFirstChar("2008-02-12")
 getFirstChar("12:00:00 GMT")
 getFirstChar("Etc/UTC")
 }
}
\seealso{
Other dateutils: 
\code{\link{fillGridDates}()},
\code{\link{isAlphaCharacter}()},
\code{\link{setGridDates.asPOSIXlt}()}
}
\author{
J Bedia
}
\concept{dateutils}
\keyword{internal}
