% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getGridProj}
\alias{getGridProj}
\title{Obtain projection information of a grid}
\usage{
getGridProj(grid)
}
\arguments{
\item{grid}{An input climate4R grid}
}
\value{
A character string with the projection definition
}
\description{
Obtain projection information of a grid
}
\details{
The projection may be defined loosely as e.g., \dQuote{LatLon Projection},
or more formally as a proj4 string. This function just retrieves the projection attribute
as included within the target grid.
Grid projection definition (and/or re-projection operations) can be achieved by the function \code{projectGrid}
from the climate4R package \pkg{geoprocessor} (\url{https://github.com/SantanderMetGroup/geoprocessoR})
}
\seealso{
Other get.helpers: 
\code{\link{get2DmatCoordinates}()},
\code{\link{getCoordinates}()},
\code{\link{getDim}()},
\code{\link{getGridVerticalLevels}()},
\code{\link{getGrid}()},
\code{\link{getRefDates}()},
\code{\link{getSeason}()},
\code{\link{getShape}()},
\code{\link{getStationID}()},
\code{\link{getVarNames}()},
\code{\link{getWT}()},
\code{\link{getYearsAsINDEX}()},
\code{\link{setGridProj}()},
\code{\link{typeofGrid}()},
\code{\link{which.leap}()}
}
\author{
J. Bedia
}
\concept{get.helpers}
