% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getGridUnits}
\alias{getGridUnits}
\title{Get grid units}
\usage{
getGridUnits(grid, var = NULL)
}
\arguments{
\item{grid}{An input grid}

\item{var}{Character vector of variable length. Variable short name(s) whose units are returned.
Only makes sense in the case of multigrids storing several variables.
 Otherwise ignored.}
}
\value{
Returns the \code{"units"} attribute
}
\description{
Get the \code{"units"} attribute of a grid
}
\examples{
\donttest{ 
require(climate4R.datasets)
data(NCEP_Iberia_ta850)
getGridUnits(NCEP_Iberia_ta850)
data(NCEP_Iberia_hus850)
getGridUnits(NCEP_Iberia_hus850)
data(NCEP_Iberia_psl)
getGridUnits(NCEP_Iberia_psl)
mf <- makeMultiGrid(NCEP_Iberia_hus850, NCEP_Iberia_psl, NCEP_Iberia_ta850)
getGridUnits(mf)
getVarNames(mf)
getGridUnits(mf, "hus@850")
getGridUnits(mf, var = c("hus@850", "ta@850"))
}
}
\seealso{
Other get.helpers unit.helpers: 
\code{\link{setGridUnits}()}
}
\author{
J Bedia
}
\concept{get.helpers unit.helpers}
\keyword{internal}
