% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getGridVerticalLevels}
\alias{getGridVerticalLevels}
\title{Get grid vertical levels}
\usage{
getGridVerticalLevels(grid, var.index = NULL)
}
\arguments{
\item{grid}{Input grid}

\item{var.index}{A vector of indices indicating the positions of the variables whose level will be returned.
This can be either an integer vector of positions, or a character vector with the short names of the variables contained
  in the multigrid (see the \code{\link{getVarNames}} helper for more details). Default to \code{NULL} (levels of all variables)}
}
\value{
A (short-)named vector with the vertical levels of the (multi)grid variable(s)
}
\description{
A helper function that returns a vector of the variable(s) vertical levels
}
\examples{
\donttest{ 
require(climate4R.datasets)
data("CFS_Iberia_hus850")
getGridVerticalLevels(CFS_Iberia_hus850)
# Surface variables usually have an undefined vertical level <NA>
data("EOBS_Iberia_tas")
getGridVerticalLevels(EOBS_Iberia_tas)

data("NCEP_Iberia_hus850")
data("NCEP_Iberia_psl")
data("NCEP_Iberia_ta850")
mg <- makeMultiGrid(NCEP_Iberia_hus850, NCEP_Iberia_psl, NCEP_Iberia_ta850)
getGridVerticalLevels(mg)
# Use of var.index to select some variables:
# either by their shortname...
getGridVerticalLevels(mg, var.index = c("hus@850","ta@850"))
# ... or by index position in the multigrid
getGridVerticalLevels(mg, var.index = c(1,3))
}
}
\seealso{
Other get.helpers: 
\code{\link{get2DmatCoordinates}()},
\code{\link{getCoordinates}()},
\code{\link{getDim}()},
\code{\link{getGridProj}()},
\code{\link{getGrid}()},
\code{\link{getRefDates}()},
\code{\link{getSeason}()},
\code{\link{getShape}()},
\code{\link{getStationID}()},
\code{\link{getTimeResolution}()},
\code{\link{getVarNames}()},
\code{\link{getWT}()},
\code{\link{getYearsAsINDEX}()},
\code{\link{setGridProj}()},
\code{\link{typeofGrid}()},
\code{\link{which.leap}()}
}
\author{
J Bedia
}
\concept{get.helpers}
\keyword{internal}
