% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getRefDates}
\alias{getRefDates}
\title{Retrieve reference dates}
\usage{
getRefDates(obj)
}
\arguments{
\item{obj}{A grid or station object}
}
\value{
A character vector of dates
}
\description{
Retrieves the \code{$start} component of the \code{$Dates} element of a grid or station.
 In case of multigrids, the dates from the first grid are returned (i.e.: \code{grid$Dates[[1]]$start}).
}
\author{
J. Bedia
}
\keyword{internal}
