% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getSeason}
\alias{getSeason}
\title{Get season}
\usage{
getSeason(obj)
}
\arguments{
\item{obj}{Any object extending the station or grid classes}
}
\value{
An integer vector with the season
}
\description{
Retrieves the season encompassed by a station or grid object
}
\examples{
data(iberia_ncep_ta850)
getSeason(iberia_ncep_ta850) # Boreal winter (DJF)
}
\seealso{
Other get.helpers: \code{\link{getCoordinates}},
  \code{\link{getDim}}, \code{\link{getGrid}},
  \code{\link{getRefDates}}, \code{\link{getShape}},
  \code{\link{getYearsAsINDEX}}, \code{\link{which.leap}}
}
\author{
J. Bedia
}
