% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getShape}
\alias{getShape}
\title{Retrieve array shape}
\usage{
getShape(obj, dimension = NULL)
}
\arguments{
\item{obj}{A grid or station object}
}
\value{
An integer vector with dim values, labelled with the \code{"dimension"} attribute names
}
\description{
Retrieve array attributes 'dimensions' and 'dim'
}
\examples{
#Station data:
data("VALUE_Iberia_tas")
getShape(VALUE_Iberia_tas)
#Regular grid
data("NCEP_Iberia_hus850")
getShape(NCEP_Iberia_hus850)
}
\seealso{
Other get.helpers: \code{\link{get2DmatCoordinates}},
  \code{\link{getCoordinates}}, \code{\link{getDim}},
  \code{\link{getGridProj}},
  \code{\link{getGridVerticalLevels}},
  \code{\link{getGrid}}, \code{\link{getRefDates}},
  \code{\link{getSeason}}, \code{\link{getStationID}},
  \code{\link{getTimeResolution}},
  \code{\link{getVarNames}}, \code{\link{getYearsAsINDEX}},
  \code{\link{typeofGrid}}, \code{\link{which.leap}}
}
\author{
J. Bedia
}
\concept{get.helpers}
\keyword{internal}
