% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetGrid.R
\name{getTemporalIntersection}
\alias{getTemporalIntersection}
\title{Temporal intersection}
\usage{
getTemporalIntersection(obs, prd, which.return = c("obs", "prd"))
}
\arguments{
\item{obs}{First grid (typically observations, but not necessarily)}

\item{prd}{Second grid (typically predictors in downscaling applications, but not necessarily)}

\item{which.return}{Which subset grid should be returned, \code{obs} or \code{prd}?}
}
\value{
The grid indicated in \code{which.return}, encompassing the overlapping time period with the other one.
}
\description{
Takes two input grids and crops the overlapping part along time dimension
}
\examples{
\donttest{
require(climate4R.datasets)
data("NCEP_Iberia_psl")
range(getRefDates(NCEP_Iberia_psl))
data("EOBS_Iberia_tas")
range(getRefDates(EOBS_Iberia_tas))
# Assume NCEP's sea-level pressure is the predictor, 
# and EOBS observations are the predictand,
# encompassing both datasets different temporal periods:
predictor <- subsetGrid(NCEP_Iberia_psl, years = 1987:2001, season = 1)
getSeason(predictor) # January
range(getYearsAsINDEX(predictor)) # period 1987-2001 
predictand <- EOBS_Iberia_tas
getSeason(predictand) # December-January-February (winter)
range(getYearsAsINDEX(predictand)) # period 1983-2002
# We often want to ensure that their time dimension matches perfectly before downscaling:
try(checkDim(predictor, predictand, dimensions = "time"))
# getTemporalIntersection is the solution:
predictand.adj <- getTemporalIntersection(obs = predictand, prd = predictor, which.return = "obs")
getSeason(predictand.adj) # January 
range(getYearsAsINDEX(predictand.adj)) # 1987-2001 
# In the same vein, it is often required to do the same on the predictor 
predictor.adj <- getTemporalIntersection(obs = predictand, prd = predictor, which.return = "prd")
checkDim(predictor.adj, predictand.adj, dimensions = "time") # perfect
}
}
\seealso{
\code{\link{checkDim}}, \code{\link{checkSeason}}, \code{\link{getYearsAsINDEX}}, \code{\link{getSeason}}, for other time dimension helpers

Other subsetting: 
\code{\link{intersectGrid}()},
\code{\link{intersectGrid.spatial}()},
\code{\link{intersectGrid.time}()},
\code{\link{limitArea}()},
\code{\link{subsetCluster}()},
\code{\link{subsetDimension}()},
\code{\link{subsetGrid}()},
\code{\link{subsetMembers}()},
\code{\link{subsetRuntime}()},
\code{\link{subsetSeason}()},
\code{\link{subsetSpatial}()},
\code{\link{subsetStation}()},
\code{\link{subsetVar}()},
\code{\link{subsetYears}()}
}
\author{
J Bedia
}
\concept{subsetting}
