% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getTimeResolution}
\alias{getTimeResolution}
\title{Get time resolution}
\usage{
getTimeResolution(grid)
}
\arguments{
\item{grid}{a grid or station data}
}
\value{
A character representation of the time resolution. Current possible values are:
\code{"1h", "3h", "6h", "12h", "DD", "MM", "YY"}. If none of these matches, \code{"unknown"} is returned
}
\description{
Get the time resolution of the input grid as a character
}
\examples{
data("EOBS_Iberia_tas")
getTimeResolution(EOBS_Iberia_tas)
monthly.grid <- aggregateGrid(EOBS_Iberia_tas, aggr.m = list(FUN = "mean", na.rm = TRUE))
stopifnot(identical(getTimeResolution(monthly.grid), "MM"))
annual.grid <- aggregateGrid(monthly.grid, aggr.y = list(FUN = "mean", na.rm = TRUE))
stopifnot(identical(getTimeResolution(annual.grid), "YY"))
}
\seealso{
Other get.helpers: \code{\link{get2DmatCoordinates}},
  \code{\link{getCoordinates}}, \code{\link{getDim}},
  \code{\link{getGridProj}},
  \code{\link{getGridVerticalLevels}},
  \code{\link{getGrid}}, \code{\link{getRefDates}},
  \code{\link{getSeason}}, \code{\link{getShape}},
  \code{\link{getStationID}}, \code{\link{getVarNames}},
  \code{\link{getYearsAsINDEX}}, \code{\link{setGridProj}},
  \code{\link{typeofGrid}}, \code{\link{which.leap}}
}
\author{
J Bedia
}
\concept{get.helpers}
\keyword{internal}
