% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeUncertainty.R
\name{getVariabilityThreshold}
\alias{getVariabilityThreshold}
\title{Calculate the variability threshold.}
\usage{
getVariabilityThreshold(grid)
}
\arguments{
\item{grid}{A grid (ideally with annual data).}
}
\value{
A grid.
}
\description{
Subset two or more grids to the common members.
}
\author{
M. Iturbide
}
