% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getVarNames}
\alias{getVarNames}
\title{Get grid variable names}
\usage{
getVarNames(grid, type = c("short", "long"))
}
\arguments{
\item{grid}{A grid or station data}

\item{type}{Character. Should either the \code{"short"} (default) or the \code{"long"} variable name(s) be returned?.
See Details.}
}
\value{
A character vector with the variable name(s) in the indicated \code{type} format.
}
\description{
Get the names of the variables of a (multi)grid
}
\details{
The long name is an optional attribute of a grid (not so the short name), and it is sometimes undefined (particularly in station datasets).
In this case, the function returns \code{NULL}.
}
\examples{
data("CFS_Iberia_pr")
getVarNames(CFS_Iberia_pr)
getVarNames(CFS_Iberia_pr, "long")
## Example with a multigrid
data("CFS_Iberia_tas")
mg <- makeMultiGrid(CFS_Iberia_tas, CFS_Iberia_pr)
getVarNames(mg)
getVarNames(mg, "long")
## Example with station data
data("VALUE_Iberia_tas")
## The long name is an optional attribute, and may be undefined:
getVarNames(VALUE_Iberia_tas, "long")
}
\seealso{
Other get.helpers: \code{\link{get2DmatCoordinates}},
  \code{\link{getCoordinates}}, \code{\link{getDim}},
  \code{\link{getGridVerticalLevels}},
  \code{\link{getGrid}}, \code{\link{getRefDates}},
  \code{\link{getSeason}}, \code{\link{getShape}},
  \code{\link{getStationID}},
  \code{\link{getTimeResolution}},
  \code{\link{getYearsAsINDEX}}, \code{\link{typeofGrid}},
  \code{\link{which.leap}}
}
\author{
J Bedia
}
\keyword{internal}
