% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWindowIndex.R
\name{getWindowIndex}
\alias{getWindowIndex}
\title{Get the index of the window days.}
\usage{
getWindowIndex(y, x, newdata, window, delta.method = FALSE)
}
\arguments{
\item{y}{A grid or station data containing the observed climate data for the training period}

\item{x}{A grid or station data containing the simulated climate data for the training period}

\item{newdata}{A grid containing the simulated climate for the test period.
\code{"scaling"}, \code{"pqm"} and \code{"gpqm"} \code{"variance"},\code{"loci"} and \code{"ptr"}. See details.}

\item{window}{vector of length = 2 (or 1) specifying the time window width used to calibrate and the 
target days (days that are being corrected). If the window length = 1 the window width is no larger than the 
target days. The window is centered on the target day/s (window width >= target days).}

\item{delta.method}{Logical (default is FALSE).}
}
\description{
Get the index of the days that corresponding to the window and the target days centered in it.
}
\author{
M. Iturbide
}
