% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getWT}
\alias{getWT}
\title{Obtain Weather Typing (WT) classification}
\usage{
getWT(grid)
}
\arguments{
\item{grid}{An input climate4R grid.}
}
\value{
The attribute vector \code{"wt.index"}, when present. Otherwise \code{NULL}. See \code{\link{clusterGrid}} for details.
}
\description{
Obtain the classification attribute from \code{\link{clusterGrid}} outputs
}
\seealso{
Other get.helpers: \code{\link{get2DmatCoordinates}},
  \code{\link{getCoordinates}}, \code{\link{getDim}},
  \code{\link{getGridProj}},
  \code{\link{getGridVerticalLevels}},
  \code{\link{getGrid}}, \code{\link{getRefDates}},
  \code{\link{getSeason}}, \code{\link{getShape}},
  \code{\link{getStationID}},
  \code{\link{getTimeResolution}},
  \code{\link{getVarNames}}, \code{\link{getYearsAsINDEX}},
  \code{\link{setGridProj}}, \code{\link{typeofGrid}},
  \code{\link{which.leap}}
}
\author{
J. Bedia
}
\concept{get.helpers}
