% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid2raster.R
\name{grid2raster}
\alias{grid2raster}
\title{Compute a grid climatology}
\usage{
grid2raster(grid, crs = NA)
}
\arguments{
\item{grid}{Input grid or multigrid without time, member or runtime dimensions, e.g. a grid returned by function 
\code{\link{climatology}}.}

\item{crs}{character or object of class CRS. PROJ.4 type description of a Coordinate Reference System (map projection). 
If this argument is missing, and the x coordinates are withing -360 .. 360 and the y coordinates are 
within -90 .. 90, "+proj=longlat +datum=WGS84" is used. Also see under Details if x is a character (filename).}
}
\value{
A raster for grid objects and a RasterStack for multigrids.
}
\description{
Calculates the climatology (i.e., complete temporal aggregation, typically the mean)
of the input grid.
}
\examples{
\dontrun{
#' # Maximum July surface temp forecast climatology
data("tasmax_forecast")
# Aggregate all members before computing the climatology
tx_mean.clim <- climatology(tasmax_forecast, by.member = FALSE)
plotClimatology(tx_mean.clim)
tx_ras <- grid2raster(tx_mean.clim)
library(raster)
spplot(tx_ras)
}
}
\seealso{
\code{\link{climatology}}; \code{\link{raster}}
}
\author{
M. Iturbide
}
