% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid2sp.R
\name{grid2sp}
\alias{grid2sp}
\title{Grid to SpatialGridDataFrame or SpatialPointsDataFrame}
\usage{
grid2sp(grid)
}
\arguments{
\item{grid}{A C4R grid}
}
\value{
A \pkg{sp} object of the class \code{\link[sp]{SpatialGridDataFrame}} or 
\code{\link[sp]{SpatialPointsDataFrame}} (Depending on the input grid).
If the inpurt grid contains members (member dimension) a list of Spatial*
objects is returned (each slot in the list is a member).
}
\description{
Convert a grid to a SpatialGridDataFrame object from package sp
}
\details{
This function was built based on function \code{\link[transformeR]{clim2sgdf}}
}
\examples{
\donttest{
require(climate4R.datasets)
library(sp)
# Climatologies of different members:
data("CFS_Iberia_tas")
x <- grid2sp(climatology(CFS_Iberia_tas))
plot(x[1]) # Plot the first member

# Daily data:
x <- grid2sp(CFS_Iberia_tas)
plot(x[[1]][1]) # Plot the first day of the first member

x <- grid2sp(EOBS_Iberia_tas)
plot(x[1]) # Plot the first day
}
}
\seealso{
\code{\link{sp2grid}}
}
\author{
M. Iturbide
}
