% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid3Dto2Dmatrix.R
\name{grid3Dto2Dmatrix}
\alias{grid3Dto2Dmatrix}
\title{Conversion of a 3D grid to a 2D matrix}
\usage{
grid3Dto2Dmatrix(grid3D)
}
\arguments{
\item{grid3D}{A 3-dimensional grid with longitude, latitude and time dimensions.}
}
\value{
A 2-dimensional matrix with time in rows and grid-points in columns.
}
\description{
Converts 3D grids of the form [time,lat,lon]
 to 2D matrices of the form [time, grid-point].
}
\details{
Global attributes of the output matrix are provided for metadata description and spatiotemporal collocation.
}
\seealso{
Other internal.helpers: \code{\link{array3Dto2Dmat.stations}},
  \code{\link{array3Dto2Dmat}},
  \code{\link{bindGrid.member}},
  \code{\link{bindGrid.spatial}},
  \code{\link{bindGrid.time}},
  \code{\link{mat2Dto3Darray.stations}},
  \code{\link{redim}}, \code{\link{sortDim.spatial}},
  \code{\link{sortDim.time}}
}
\author{
J. Bedia
}
\concept{internal.helpers}
\keyword{internal}
