% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gridDepth}
\alias{gridDepth}
\title{Level depth in a list}
\usage{
gridDepth(this)
}
\arguments{
\item{this}{list}
}
\value{
number of nesting lists
}
\description{
Level depth in a list
}
\author{
M. Iturbide
}
\keyword{internal}
