% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridFromPCs.R
\name{gridFromPCs}
\alias{gridFromPCs}
\title{Reconstruct a Grid from EOFs and principal components}
\usage{
gridFromPCs(prinCompObj, var)
}
\arguments{
\item{prinCompObj}{A EOF analysis object as returned by \code{\link{prinComp}}}

\item{var}{Character string indicating the variable to be re-constructed. In case of
PCA analyses performed on a single variable (either multimember or not), this can be omitted.}
}
\value{
A list similar to the \code{loadGridData} output (package \pkg{loadeR}), but simplified. See details.
}
\description{
Reconstructs a grid of a climatic variable from the outputs of a
principal components analysis
}
\details{
The output of this function returns the minimum required information to use the
 \code{\link{plotMeanGrid}} method, and is intended for comparison and visual analysis
 of the differences between the original grids and the reconstructed ones, for instance
 in determining an optimal number of PCs etc... Hence, the temporal information (i.e.,
  the \code{Dates} object) is lost, and should be retrieved from the original grid/multigrid
  object used to compute the PC/EOF analysis.
}
\examples{
# Multigrid constructor:
multigrid <- makeMultiGrid(NCEP_Iberia_hus850, NCEP_Iberia_ta850, NCEP_Iberia_psl)
# In this example, we retain the first 10 PCs
pca <- prinComp(multigrid, n.eofs = 10)
# We recover the sea-level pressure grid from the its PCs:
names(pca)
psl2 <- gridFromPCs(pca, "psl")
str(psl2)
# The attributes of psl2 indicate that this is a reconstructed grid from 10 PCs, 
# explaining 99\\\% of the variance:
attributes(psl2)
multigrid$Variable$varName
# psl is the 3rd one

# An example of multimember reconstruction from a multimember PC analysis:
data("CFS_Iberia_tas")
# Note that multimember pca analysis takes some time, depending on the computer
pca2 <- prinComp(CFS_Iberia_tas, n.eofs = 10)
tas_recovered <- gridFromPCs(pca2)
plotMeanGrid(tas_recovered, multi.member = TRUE)
# Also note that now the length of the "nPCs" and "explained_variance" matches 
# the number of members:
attributes(tas_recovered)
}
\seealso{
\code{\link{prinComp}}, \code{\link{plotEOF}}
}
\author{
J. Bedia
}
