% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleGrid.R
\name{gridScale.}
\alias{gridScale.}
\title{Local scaling internal}
\usage{
gridScale.(grid, base, ref, clim.fun, by.member, type, parallel,
  max.ncores, ncores, spatial.frame)
}
\description{
Internal workhorse for local scaling
}
\author{
J Bedia
}
\keyword{internal}
