% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleGrid.R
\name{gridScale.type}
\alias{gridScale.type}
\title{Local scaling type internal}
\usage{
gridScale.type(
  clim,
  n.times,
  ind.time,
  Xc,
  Xref,
  type,
  lapply_fun,
  base.std,
  ref.std
)
}
\description{
Internal for center/standardize/ratio local scaling
}
\author{
J Bedia
}
\keyword{internal}
