% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridUncertainty.R
\name{gridUncertainty}
\alias{gridUncertainty}
\title{Calculate the uncertainty of the climate change signal.}
\usage{
gridUncertainty(historical, anomaly)
}
\arguments{
\item{historical}{A grid with annual historical data}

\item{anomaly}{A grid (with time dimension = 1).}
}
\value{
A grid with three values (uncertainty categories); 0 = robust signal, 
1 = no signal or no change, 2 = conflicting signals).
}
\description{
Calculate the uncertainty of the climate change signal.
}
\author{
M. Iturbide
}
