% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectGrid.R
\name{intersectGrid.spatial}
\alias{intersectGrid.spatial}
\title{Temporal intersection of multiple grids}
\usage{
intersectGrid.spatial(..., which.return = 1)
}
\arguments{
\item{...}{Input grids}

\item{which.return}{Integer of the index to specify which grids in "..." are to be returned.}
}
\value{
The grids indicated in \code{which.return}, encompassing the overlapping time period.
}
\description{
Takes multiple input grids and crops the overlapping part along time longitude and latitude dimensions
}
\seealso{
\code{\link{checkDim}}, \code{\link{checkSeason}}, \code{\link{getYearsAsINDEX}}, \code{\link{getSeason}}, for other time dimension helpers

Other subsetting: \code{\link{getTemporalIntersection}},
  \code{\link{intersectGrid.time}},
  \code{\link{intersectGrid}}, \code{\link{limitArea}},
  \code{\link{subsetCluster}},
  \code{\link{subsetDimension}}, \code{\link{subsetGrid}},
  \code{\link{subsetMembers}}, \code{\link{subsetRuntime}},
  \code{\link{subsetSeason}}, \code{\link{subsetSpatial}},
  \code{\link{subsetStation}}, \code{\link{subsetVar}},
  \code{\link{subsetYears}}
}
\author{
M Iturbide
}
\concept{subsetting}
