% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectMembers.R
\name{intersectMembers}
\alias{intersectMembers}
\title{Subset two or more grids to the common members.}
\usage{
intersectMembers(...)
}
\arguments{
\item{...}{Grids.}
}
\value{
input grids are updated in the global environment.
}
\description{
Subset two or more grids to the common members.
}
\author{
M. Iturbide
}
