% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillGrid.R
\name{isAlphaCharacter}
\alias{isAlphaCharacter}
\title{Is alphabetic character}
\usage{
isAlphaCharacter(x)
}
\arguments{
\item{x}{Character string}
}
\value{
Logical
}
\description{
Is alphabetic character? Internal utility for parsing dates
}
\seealso{
Other dateutils: 
\code{\link{fillGridDates}()},
\code{\link{getFirstChar}()},
\code{\link{setGridDates.asPOSIXlt}()}
}
\author{
J Bedia
}
\concept{dateutils}
\keyword{internal}
