% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isGrid}
\alias{isGrid}
\title{Check if object is a grid}
\usage{
isGrid(grid)
}
\arguments{
\item{grid}{Input object.}
}
\value{
Logical.
}
\description{
Check if object is a grid
}
\author{
M. Iturbide
}
\keyword{internal}
