% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isMultigrid}
\alias{isMultigrid}
\title{Check if object is a multigrid}
\usage{
isMultigrid(grid)
}
\arguments{
\item{grid}{Input object.}
}
\value{
Logical.
}
\description{
Check if object is a multigrid
}
\author{
M. Iturbide
}
\keyword{internal}
