% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-join-S4methods.R
\docType{methods}
\name{join}
\alias{join}
\alias{left_join}
\alias{left_join,data.frame-method}
\alias{left_join,DataFrame-method}
\title{Join operations supporting S4 class objects}
\usage{
left_join(x, y, ...)

\S4method{left_join}{data.frame}(x, y, ...)

\S4method{left_join}{DataFrame}(x, y, by, copy = FALSE, suffix = NULL,
  ...)
}
\arguments{
\item{x, y}{\code{DataFrame}.
Data frames, or objects, to be coerced into a single object.}

\item{by}{\code{character}.
Column names to use for merge operation.
Note that this is always required, unlike the dplyr \code{tbl_df} method, where
this is optional.}

\item{copy, suffix, ...}{Disabled for \code{DataFrame} method.}
}
\description{
Join operations supporting S4 class objects
}
\note{
Updated 2019-08-14.
}
\examples{
DataFrame <- S4Vectors::DataFrame

## DataFrame ====
df1 <- DataFrame(
    id = as.factor(seq(4L)),
    genotype = as.factor(rep(x = c("wt", "ko"), each = 2L))
)
print(df1)
df2 <- DataFrame(
    id = as.factor(seq(4L)),
    treatment = as.factor(rep(x = c("control", "expt"), times = 2L))
)
print(df2)
left_join(df1, df2, by = "id")
}
\seealso{
\itemize{
\item \code{help(topic = "join", package = "dplyr")}.
\item \code{help(topic = "merge", package = "S4Vectors")}.
}
}
