% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambWT.R
\name{lambWT}
\alias{lambWT}
\title{Calculation of Lamb Weather types (WT).}
\usage{
lambWT(grid, center.point = c(-5, 55))
}
\arguments{
\item{grid}{A grid (gridded or station dataset), or multimember grid object of MSLP values.}

\item{center.point}{A two value vector that must include lon and lat from a location that will work as center point for the Lamb WT.
See details.}
}
\value{
The Lamb WT circulation index (and members, if applicable) with:
\itemize{
\item index: vector with the corresponding weather type from each point of the series, that is defined as follows:

purely anticyclonic = 1

directional anticyclonic from NE to N = 2 to 9

purely directional from NE to N = 10 to 17

purely cyclonic = 18

directional cyclonic from NE to N = 19 to 26

light indeterminate flow N = 27  .
\item pattern: Array with the spatial pattern of the 27 weather types obtained.
\item dates and coordinates.
\item further arguments related to the Lamb WT index.
}
}
\description{
Calculate automated Lamb WT as defined in Trigo daCamara 2000 and Jones et al. 2012, 
both Int J Climatol
}
\details{
According to Jones et al. 2012 (Int J Climatol), Lamb WT is only applied on North Atlantic domain. 
The input grid units must be Pa, not hPa/mbar. If it is not in Pa, the units will be converted automatically.
A center location point must be specified by the user. Then, the function calculates from left to right and from first to 16st 
the rest of the location point from the grid specified by Jones et al. 2012:
 
  \tabular{ccccccccccccc}{
    \tab  \tab  \tab    \tab  \tab  \tab 01 \tab  \tab  \tab 02 \tab  \tab  \tab    \cr
    \tab  \tab  \tab 03 \tab  \tab  \tab 04 \tab  \tab  \tab 05 \tab  \tab  \tab 06 \cr
    \tab  \tab  \tab 07 \tab  \tab  \tab 08 \tab  \tab  \tab 09 \tab  \tab  \tab 10 \cr
    \tab  \tab  \tab 11 \tab  \tab  \tab 12 \tab  \tab  \tab 13 \tab  \tab  \tab 14 \cr
    \tab  \tab  \tab    \tab  \tab  \tab 15 \tab  \tab  \tab 16 \tab  \tab  \tab    
}  

where the north-south distance is 5º and the west-east distance is 10º.
}
\examples{
\dontrun{
data(NCEP_slp_2001_2010)
lamb.wt <- lambWT(grid = NCEP_slp_2001_2010)
}
}
