% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_norm.R
\name{layer_norm}
\alias{layer_norm}
\title{Layer Normalization}
\usage{
layer_norm(x, epsilon = 1e-06)
}
\arguments{
\item{x}{inputs}

\item{epsilon}{scale}
}
\value{
outputs of layer normalization
}
\description{
Layer Normalization
}
