% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limitArea.R
\name{limitArea}
\alias{limitArea}
\title{Limit the area in an object returned by function \code{\link{getGrid}}}
\usage{
limitArea(coordinates, lonLim, latLim)
}
\arguments{
\item{coordinates}{A list as returned by function \code{\link{getGrid}}.}

\item{lonLim}{Vector of length = 2, with minimum and maximum longitude coordinates, in decimal degrees,
of the bounding box defining the subset. For single-point subsets, a numeric value with the
longitude coordinate. If \code{NULL} (default), no subsetting is performed on the longitude dimension}

\item{latLim}{Same as \code{lonLim} argument, but for latitude.}
}
\value{
same as the input.
}
\description{
Limit the area in an object returned by function \code{\link{getGrid}}
}
\seealso{
Other subsetting: 
\code{\link{getTemporalIntersection}()},
\code{\link{intersectGrid.spatial}()},
\code{\link{intersectGrid.time}()},
\code{\link{intersectGrid}()},
\code{\link{subsetCluster}()},
\code{\link{subsetDimension}()},
\code{\link{subsetGrid}()},
\code{\link{subsetMembers}()},
\code{\link{subsetRuntime}()},
\code{\link{subsetSeason}()},
\code{\link{subsetSpatial}()},
\code{\link{subsetStation}()},
\code{\link{subsetVar}()},
\code{\link{subsetYears}()}
}
\author{
M. Iturbide
}
\concept{subsetting}
