% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localScaling.R
\name{localScaling.}
\alias{localScaling.}
\title{Local scaling internal}
\usage{
localScaling.(grid, base, ref, clim.fun, by.member, type, parallel, max.ncores,
  ncores, scale)
}
\description{
Internal workhorse for local scaling
}
\author{
J Bedia
}
\keyword{internal}
