% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localScaling.R
\name{localScaling}
\alias{localScaling}
\title{Grid local scaling}
\usage{
localScaling(grid, base = NULL, ref = NULL, clim.fun = list(FUN = "mean",
  na.rm = TRUE), by.member = TRUE, time.frame = c("none", "monthly",
  "daily"), parallel = FALSE, max.ncores = 16, ncores = NULL)
}
\arguments{
\item{grid}{Input grid to be rescaled}

\item{base}{Reference baseline whose climatology will be subtracted to the input grid. If \code{NULL} (the default),
the climatology is directly computed from the input \code{grid} via \code{\link{climatology}}, either member-by-member
or using the ensemble mean climatology, as specified by the \code{by.member} flag. \code{base} must NOT be climatology
 (the base climatology is internally calculated via the argument \code{clim.fun}).}

\item{ref}{Reference grid. After subtracting to grid its climatology (as defined by \code{grid.clim}), the mean
climatology of this additional grid is added. Default to NULL, so no reference grid is used.}

\item{clim.fun}{Function to compute the climatology. Default to mean. 
See \code{\link{climatology}} for details on function definition.}

\item{by.member}{Logical. In case of multimember grids, should the climatology be computed sepparately
for each member (\code{by.member=TRUE}), or a single climatology calculated from the ensemble mean
 (\code{by.member=FALSE})?. Default to \code{TRUE}. Argument passed to \code{\link{climatology}}.}

\item{time.frame}{Character indicating the time frame to perform the scaling. Possible values are
\code{"none"}, which considers the climatological mean of the whole period given in 
\code{base} and/or \code{ref}, \code{"monthly"}, that performs the calculation on a monthly basis
and \code{"daily"}, for a julian day-based approach.}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\value{
A locally scaled grid
}
\description{
Scale a grid (or compute anomalies)
}
\details{
The reference grid (\code{ref}) is used to correct the input grid, as follows:

\deqn{grid' = grid - base_clim + ref_clim}

, where \emph{base_clim} corresponds to the baseline climatology (by default the grid climatology), thus yielding an anomaly
w.r.t. the base, and \emph{ref_clim} is a reference climatological value added to the previously calculated anomaly. Thus, 
 if both \code{base} and \code{ref} are set to \code{NULL}, the output grid corresponds to the anomaly field of the 
 input \code{grid} w.r.t. its own mean. The way \emph{base_clim} and \emph{ref_clim_grid} are computed in case of multimember grids is controlled
 by the argument \code{by.member}. By default the climatological mean is used, but this can be changed
  by the user through the argument \code{clim.fun}, that is passed to \code{\link{climatology}}.


The \code{ref} usually corresponds to the control (historical, 20C3M...) run of the GCM in the training period in climate change applications,
or the hindcast data for the training period in s2d applications. Note that by default \code{ref = NULL}. In this 
case it will be assumed to be the \code{pred} grid. This can be used for instance when train and test correspond
to the same model.
}
\examples{
## ANOMALIES
data("NCEP_Iberia_psl")
# Define average aggregation function
f = list(FUN = "mean", na.rm = TRUE)
psl <- aggregateGrid(NCEP_Iberia_psl, aggr.y = f) # get interannual DJF series 

## When 'base' and 'ref' are not supplied,
## the input grid climatology (by default the mean) is subtracted, thus yielding anomalies:
psl.anom <- localScaling(psl)
# spatial aggregation of the output (for plotting the time series)
psl.anom.iberia <- aggregateGrid(psl.anom, aggr.lat = f, aggr.lon = f)
plot(getYearsAsINDEX(psl.anom.iberia), psl.anom.iberia$Data, ty = 'b',
     ylab = "MSLP anomalies (Pa)", xlab = "year",
     main = "NCEP Reanalysis - Mean SLP anomalies (DJF)")
grid()
abline(h = 0, lty = 2, col = "blue")
# In the particular case of multimember grids, the anomalies are computed for each member
# by subtracting either their own mean (by.member = TRUE) or the 
# multimember mean climatology (by.member = FALSE)
data("CFS_Iberia_tas")
a <- localScaling(CFS_Iberia_tas, by.member = FALSE)
aa <- aggregateGrid(a, aggr.lat = f, aggr.lon = f, aggr.m = f, aggr.y = f)
# Example, member 4 time series
plot(as.Date(getRefDates(aa)), aa$Data[4,], ty = "o", xlab = "Date", 
     ylab = "Tmean anomaly (degC)", main = "Tmean DJF anomalies, Member 4")
abline(h = 0, lty = 2, col = "blue")
grid()
b <- localScaling(CFS_Iberia_tas, by.member = TRUE) # almost identical in this case
bb <- aggregateGrid(b, aggr.lat = f, aggr.lon = f, aggr.m = f, aggr.y = f)
lines(as.Date(getRefDates(aa)), bb$Data[4,], col = "red", ty = "l")
legend("bottomright", c("by.member = FALSE", "by.member = TRUE"), lty = 1, col = c(1,2))

# In this example, the anomalies are calculated using a different period specifying a "base".
# Note that "base" could be also a grid of a different dataset, for instance a reanalysis
data(EOBS_Iberia_tas)
grid <- subsetGrid(EOBS_Iberia_tas, years = 1999:2000)
base <- subsetGrid(EOBS_Iberia_tas, years = 1998)
lc <- localScaling(grid = grid, base = base)
plot(lc$Data[,,15,15], ty = 'l', ylim = c(-10,10),
     xlab = "time", ylab = "Anomaly (degC)",
     main = "Anomalies w.r.t. 1998")
grid()
abline(h = 0, lty = 2, col = "grey30")
# The anomalies are calculated on a monthly basis using the 'time.frame' argument:
lc.m <- localScaling(grid = grid, base = base, time.frame = "monthly")
lines(lc.m$Data[,,15,15], col = "red")
lc.d <- localScaling(grid = grid, base = base, time.frame = "daily")
lines(lc.d$Data[,,15,15], col = "blue")
legend("topleft", c("none", "monthly","daily"), title = "'time.frame'",
       lty = 1, col = c(1,2,4), bty = "n")

# An example in which the reference climatology is added to the anomalies:
# (not quite meaningful, though)
grid <- subsetGrid(EOBS_Iberia_tas, years = 1999)
base <- subsetGrid(EOBS_Iberia_tas, years = 1998)
ref <- subsetGrid(EOBS_Iberia_tas, years = 2000)
lc.ref <- localScaling(grid = grid, base = base, ref = ref)
lc.ref.m <- localScaling(grid = grid, base = base, ref = ref, time.frame = "monthly")
plot(grid$Data[,15,15], ty = "l", ylim = c(-7.5,10))
lines(lc.ref$Data[,,15,15], col = "blue")
lines(lc.ref.m$Data[,,15,15], col = "red")
plotClimatology(climatology(grid))
plotClimatology(climatology(lc.ref))
}
\author{
J. Bedia
}
