% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localScaling.R
\name{localScaling.type}
\alias{localScaling.type}
\title{Local scaling type internal}
\usage{
localScaling.type(clim, n.times, ind.time, Xc, Xref, type, lapply_fun,
  base.std)
}
\description{
Internal for additive/ratio local scaling
}
\author{
J Bedia
}
\keyword{internal}
