% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClimatology.R
\name{map.stippling}
\alias{map.stippling}
\title{Climatological map stippling}
\usage{
map.stippling(clim, draw.below.threshold = 0.05, ...)
}
\arguments{
\item{clim}{A (verification) climatology, e.g.: as produced by \code{\link{easyVeri2grid}}.
This often contains p-values, but not necessarily.}

\item{draw.below.threshold}{Confidence level or threshold below which stippling points are drawn. Default to \code{0.05}.}

\item{...}{Further optional arguments passed to \code{sp.layout} argument in \code{spplot}.}
}
\description{
Create a points panel layout to add to \code{plotClimatology}.
Typically needed to stipple significant points in climatologies, or other types of coordinates
}
\details{
The function generates a \code{"sp.points"} layout list. Further formatting arguments can be passed here.
 For further details and examples see the help of \code{\link[sp]{spplot}}.
}
\seealso{
\code{\link{plotClimatology}}
}
\author{
J. Bedia
}
