% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchRowNameColumn.R
\name{matchRowNameColumn}
\alias{matchRowNameColumn}
\title{Match row name column}
\usage{
matchRowNameColumn(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{character(1)} or \code{NULL}.
\itemize{
\item data.table: \code{"rn"}.
\item tibble: \code{"rowname"}.
}
}
\description{
Automatically detect row names column, if defined.
}
\examples{
data(dt, tbl, package = "acidtest")

## data.table ====
matchRowNameColumn(dt)

## tbl_df ====
matchRowNameColumn(tbl)
}
