% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchStations.R
\name{matchStations}
\alias{matchStations}
\title{Order the stations according to a particular order}
\usage{
matchStations(obj, ref)
}
\arguments{
\item{obj}{A station object to be ordered, normally predictions.}

\item{ref}{A station object used as reference to order, normally observations.}
}
\value{
A new grid object with binary values.
}
\description{
Convert a real variable into a binary variable (i.e., 0 or 1) filtering by a threshold.
}
\author{
J. Bano-Medina
}
