% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridUncertainty.R
\name{modelAgreement}
\alias{modelAgreement}
\title{Calculate model agreement on the sign of change.}
\usage{
modelAgreement(x, th = 80)
}
\arguments{
\item{x}{numeric of anomalies}

\item{th}{percentage value (Default is 80).}
}
\value{
binary numeric (0 = no agreement, 1 = agreement).
}
\description{
Calculate if signal exists in the anomalies relative to the historical variability.
}
\author{
M. Iturbide
}
