% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_head.R
\name{multi_head}
\alias{multi_head}
\title{Multi-Headed Attention}
\usage{
multi_head(Q, K, V, d_model, num_heads, mask = NULL)
}
\arguments{
\item{Q}{queries}

\item{K}{keys}

\item{V}{values}

\item{d_model}{dimensions of the model}

\item{num_heads}{number of heads}

\item{mask}{optional mask}
}
\value{
multi-headed attention outputs
}
\description{
Multi-Headed Attention
}
