% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutate-methods.R
\docType{methods}
\name{mutateAll}
\alias{mutateAll}
\alias{mutateAt}
\alias{mutateIf}
\alias{transmuteAt}
\alias{transmuteIf}
\alias{mutate}
\alias{mutateAll,DataFrame-method}
\alias{mutateAt,DataFrame-method}
\alias{mutateIf,DataFrame-method}
\alias{transmuteAt,DataFrame-method}
\alias{transmuteIf,DataFrame-method}
\title{Mutate multiple columns}
\usage{
mutateAll(object, ...)

mutateAt(object, ...)

mutateIf(object, ...)

transmuteAt(object, ...)

transmuteIf(object, ...)

\S4method{mutateAll}{DataFrame}(object, funs, ...)

\S4method{mutateAt}{DataFrame}(object, vars, funs, ...)

\S4method{mutateIf}{DataFrame}(object, predicate, funs, ...)

\S4method{transmuteAt}{DataFrame}(object, vars, funs, ...)

\S4method{transmuteIf}{DataFrame}(object, predicate, funs, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional argument.}

\item{funs, predicate}{\code{function}.
Refer to dplyr documentation for details.}

\item{vars}{\code{character}.
Column names.}
}
\value{
Modified object.
}
\description{
Mutate multiple columns
}
\note{
Updated 2019-08-15.
}
\section{\code{data.frame} methods}{


Since we are defining S4 methods in this package, we are providing
passthrough support to dplyr for \code{data.frame} class objects. Our generic
methods pass through to dplyr mutate functions, which are optimized for
\code{tbl_df} class.

Refer to \code{help(topic = "mutate_all", package = "dplyr")} for details.
}

\examples{
data(mtcars, package = "datasets")

## DataFrame ====
x <- as(mtcars, "DataFrame")
mutateAll(x, funs = log, base = 2L)
mutateAt(x, vars = c("mpg", "cyl"), log, base = 2L)
mutateIf(x, predicate = is.double, funs = as.integer)
transmuteAt(x, vars = c("mpg", "cyl"), log, base = 2L)
transmuteIf(x, predicate = is.double, funs = as.integer)
}
