% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\name{NCEP_Iberia_hus850}
\alias{NCEP_Iberia_hus850}
\title{Grid containing NCEP-ENCAR reanalysis1 data of daily specific humidity at 850mb for the 
Iberian Peninsula (DJF, 1983-2002).}
\format{A grid}
\source{
\url{http://www.meteo.unican.es/ecoms-udg}
}
\description{
This R data object was obtained from the UDG server (\url{http://www.meteo.unican.es/udg-tap}, 
log in is requiered, 
see \code{\link[loadeR]{loginUDG}}) by means of function \code{\link[loadeR]{loadGridData}} 
(package \href{https://github.com/SantanderMetGroup/loadeR}{\code{loadeR}}) in the following manner:

\code{years <- 1983:2002}\cr 
\code{season <- c(12,1,2)}\cr 
\code{lonLim <- c(-10, 4)}\cr 
\code{latLim <- c(35, 44)}\cr 


\code{loginUDG("username", "pasword")}\cr 


\code{NCEP_Iberia_hus850 <- loadGridData(
dataset = "http://meteo.unican.es/tds5/dodsC/ncepReanalysis1/ncepReanalysis1_4xDaily.ncml", 
var = "shum@850", years = years, lonLim = lonLim, latLim = latLim, 
season = season, time = "DD", aggr.d = "mean")}
}
\examples{
data(NCEP_Iberia_hus850)
plotClimatology(climatology(NCEP_Iberia_hus850),
                backdrop.theme = "countries", scales = list(draw = TRUE))
}
