% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{NCEP_Iberia_ta850}
\alias{NCEP_Iberia_ta850}
\title{Grid containing NCEP-ENCAR reanalysis1 data of daily air temperature at 850mb for the 
Iberian Peninsula (DJF, 1983-2002).}
\format{
A grid
}
\source{
\url{http://www.meteo.unican.es/ecoms-udg}
}
\usage{
NCEP_Iberia_ta850
}
\description{
This R data object was obtained from the UDG server (\url{http://www.meteo.unican.es/udg-tap}, 
log in is requiered, 
see \code{\link[loadeR]{loginUDG}}) by means of function \code{\link[loadeR]{loadGridData}} 
(package \href{https://github.com/SantanderMetGroup/loadeR}{\code{loadeR}}) in the following manner:

\code{years <- 1983:2002}\cr 
\code{season <- c(12,1,2)}\cr 
\code{lonLim <- c(-10, 4)}\cr 
\code{latLim <- c(35, 44)}\cr 


\code{loginUDG("username", "pasword")}\cr 


\code{NCEP_Iberia_ta850 <- loadGridData(
dataset = "http://meteo.unican.es/tds5/dodsC/ncepReanalysis1/ncepReanalysis1_4xDaily.ncml", 
var = "air@850", years = years, lonLim = lonLim, latLim = latLim, 
season = season, time = "DD", aggr.d = "mean")}
}
\keyword{datasets}
