% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\name{NCEP_Iberia_tas}
\alias{NCEP_Iberia_tas}
\title{Grid containing NCEP reanalysis data of daily mean temperature for the 
Iberian Peninsula (DJF, 1983-2002).}
\format{A grid}
\source{
\url{http://www.meteo.unican.es/ecoms-udg}
}
\description{
This R data object was obtained from the ECOMS-UDG server 
(\url{http://www.meteo.unican.es/ecoms-udg}, 
log in is requiered, see \code{\link[loadeR]{loginUDG}}) 
by means of function \code{\link[loadeR.ECOMS]{loadECOMS}} 
(package \href{https://github.com/SantanderMetGroup/loadeR.ECOMS}{\code{loadeR.ECOMS}}) 
in the following manner:

\code{years <- 1983:2002}\cr 
\code{season <- c(12,1,2)}\cr 
\code{lonLim <- c(-10, 4)}\cr 
\code{latLim <- c(35, 44)}\cr 


\code{loginUDG("username", "pasword")}\cr 


\code{NCEP_Iberia_tas <- loadECOMS(dataset = "NCEP_reanalysis1",
var = "tas", years = years, lonLim = lonLim, latLim = latLim, season = season,
time = "DD", aggr.d = "mean")}
}
\examples{
data(NCEP_Iberia_tas)
plotClimatology(climatology(NCEP_Iberia_tas),
                backdrop.theme = "countries", scales = list(draw = TRUE))
}
