% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overGrid.R
\name{overGrid}
\alias{overGrid}
\title{Spatial overlay for grids and Spatial* objects}
\usage{
overGrid(grid, layer, subset = FALSE)
}
\arguments{
\item{grid}{Input grid or station data.}

\item{layer}{SpatialPolygons object from which the geometries or attributes are 
queried (type \code{help(package = "sp")}.}

\item{subset}{Logical (default is FALSE). If TRUE, spatial subsetting is performed over
the imput grid using the bounding coordinates of argument 'layer'.}
}
\value{
A grid or station data.

A grid
}
\description{
Application of function \code{over} from \pkg{sp} to grids or
station data.
}
\details{
All grid locations outside layer are filled with NAs.
}
\seealso{
Other subsetting: \code{\link{getTemporalIntersection}},
  \code{\link{intersectGrid.spatial}},
  \code{\link{intersectGrid.time}},
  \code{\link{intersectGrid}},
  \code{\link{subsetDimension}}, \code{\link{subsetGrid}},
  \code{\link{subsetMembers}}, \code{\link{subsetRuntime}},
  \code{\link{subsetSeason}}, \code{\link{subsetSpatial}},
  \code{\link{subsetStation}}, \code{\link{subsetVar}},
  \code{\link{subsetYears}}
}
\author{
M. Iturbide
}
