% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern2grid.R
\name{pattern2grid}
\alias{pattern2grid}
\title{Transform a climate4R pattern into a climate4R grid.}
\usage{
pattern2grid(
  pattern,
  index.code,
  members = 1,
  season = 1,
  coords = NULL,
  dates = NULL
)
}
\arguments{
\item{pattern}{A climate4R pattern, obtained from \code{circIndexGrid}.}

\item{index.code}{Circulation index to be extracted. Only ONE circulation index can be extracted at a time. See \code{circIndexShow()} for details.}

\item{members}{In a model ensemble, position of the desired member. Default: 1 (no member dimension).}

\item{season}{Season to be extracted. Default: 1.}

\item{coords}{List of xyCoords, as in C4R grids.}

\item{dates}{List of (start and end) dates, as in C4R grids.}
}
\value{
A typical C4R grid.
}
\description{
Transform a climate4R pattern into a climate4R grid.
}
\examples{
\dontrun{ 
data(NCEP_hgt500_2001_2010)
cpc <- indicesCPC(grid=NCEP_hgt500_2001_2010, index.code = c("NAO", "EA","PNA"), season=1)
cpc2plot <- pattern2grid(cpc, index.code="NAO", coords =NCEP_hgt500_2001_2010$xyCoords,
                        dates=NCEP_hgt500_2001_2010$Dates)
spatialPlot(cpc2plot, backdrop.theme="coastline")
}
}
