% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectGrid.R
\name{projectGrid}
\alias{projectGrid}
\title{grid datum definition and transformation}
\usage{
projectGrid(grid, original.CRS = "", new.CRS = "")
}
\arguments{
\item{grid}{a grid or multigrid (including station data).}

\item{original.CRS}{character as passed to function \code{\link{CRS}}. It defines the original projection. If the data contains the 
projection information, a warning is returned and the projection in redefined.}

\item{new.CRS}{character as passed to function \code{\link{CRS}}.}
}
\description{
Defines and/or transforms the projection of a grid (or station data) by means of a \code{\link[sp]{CRS}} object.
}
\details{
This function uses \code{\link{spTransform}},  \code{\link{CRS}} and \code{\link{proj4string}} from package \pkg{sp}
}
\examples{
data("VALUE_Iberia_pr")
plot(getCoordinates(VALUE_Iberia_pr))
grid <- projectGrid(VALUE_Iberia_pr,
                    original.CRS = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0",
                    new.CRS = "+init=epsg:28992")
plot(getCoordinates(grid))

data("EOBS_Iberia_pr")
plot(get2DmatCoordinates(EOBS_Iberia_pr))
grid <- projectGrid(EOBS_Iberia_pr,
                    original.CRS = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0",
                    new.CRS = "+init=epsg:28992")
plot(get2DmatCoordinates(grid))
require(visualizeR)
spatialPlot(climatology(grid))
}
\seealso{
\code{\link{spTransform}}, \code{\link{proj4string}}.
}
\author{
M. Iturbide
}
