% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/redim.R
\name{redim}
\alias{redim}
\title{Complete missing dimensions of Grid or Station objects}
\usage{
redim(obj, member = TRUE, runtime = FALSE, var = FALSE, station = FALSE,
  drop = FALSE)
}
\arguments{
\item{obj}{A grid or station data}

\item{member}{logical. Add member dimension (default = TRUE)}

\item{runtime}{logical. Add runtime dimension (default = FALSE)}

\item{var}{logical. Add var dimension (default = FALSE)}

\item{station}{logical. Only if obj is station data. If TRUE, the "station"
dimension is not replaced by fake "lat" and "lon" dimensions (default is FALSE).}

\item{drop}{logical. Drop dimensions of length = 1 (default = FALSE)}
}
\value{
The same object with all the dimensions (i.e. member, time, station)
}
\description{
Complete all dimensions of the Data array
}
\details{
The function down not handle multigrids (i.e. \code{"var"} dimension).
 Thus, multigrids need to be subsetted along \code{"var"} prior to redimensioning.
}
\author{
M. Iturbide, J. Bedia
}
\keyword{internal}

