% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refineGrid.R
\name{refineGrid}
\alias{refineGrid}
\title{Horizontal grid refining}
\usage{
refineGrid(grid, times = 5)
}
\arguments{
\item{grid}{a grid or multigrid to be aggregated.}

\item{times}{Refinement times. For instance, if the resolution of \code{grid}
is 1 degree and \code{times = 2}, the resolution of the output grid will be 0.5 
degrees.}
}
\value{
A grid or multigrid.
}
\description{
Increasing the grid coordinate density without altering the original pixel values.
}
\examples{
\dontrun{
require(climate4R.datasets)
data("EOBS_Iberia_pr")
library(visualizeR)
spatialPlot(climatology(EOBS_Iberia_pr))
newgrid <- refineGrid(redim(EOBS_Iberia_pr), 
                       times = 2)
spatialPlot(climatology(newgrid))
}
}
\author{
M. Iturbide
}
