% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel.R
\name{relevel}
\alias{relevel}
\alias{relevel.DataFrame}
\alias{relevel.GRanges}
\title{Reorder factor levels}
\usage{
\method{relevel}{DataFrame}(object)

\method{relevel}{GRanges}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Modified object.
Factor levels will be readjusted (i.e. superfluous levels are dropped).
}
\description{
Reorder factor levels
}
\examples{
data(rse, package = "acidtest")

## DataFrame
df <- SummarizedExperiment::rowData(rse)
x <- relevel(df)
summary(x)

## GRanges
gr <- SummarizedExperiment::rowRanges(rse)
x <- relevel(gr)
summary(x)
}
