% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{reorderStation}
\alias{reorderStation}
\title{Reorder loc dimension in irregular grids}
\usage{
reorderStation(grid, axis = c("x", "y"))
}
\arguments{
\item{grid}{Input irregular grid to be subset (possibly a multimember/multigrid).}

\item{lon}{Logical to order data according to longitudes.}

\item{lat}{Logical to order data according to latitudes.}
}
\value{
An irregular grid (or multigrid).
}
\description{
Retrieves a grid that is with the 'loc' dimension reordered.
 Multimember multigrids are supported. Subroutine of \code{\link{subsetGrid}}.
}
\author{
M. Iturbide
}
\keyword{internal}
