% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_means.R
\name{row_means}
\alias{row_means}
\title{Row Means}
\usage{
row_means(x)
}
\arguments{
\item{x}{matrix}
}
\value{
vector with the mean of each of row of the input matrix
}
\description{
Row Means
}
\examples{
row_means(t(matrix(1:5)))
}
