% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_vars.R
\name{row_vars}
\alias{row_vars}
\title{Row Variances}
\usage{
row_vars(x)
}
\arguments{
\item{x}{matrix}
}
\value{
vector with the variance of each of row of the input matrix
}
\description{
Row Variances
}
\examples{
row_vars(t(matrix(1:5)))
}
