% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{S4Transformer-package}
\alias{S4Transformer}
\alias{S4Transformer-package}
\title{S4Transformer}
\description{
Improved Bioconductor S4 coercion method support for tibble and sparse matrix
data classes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/steinbaugh/S4Transformer/}
  \item Report bugs at \url{https://github.com/steinbaugh/S4Transformer/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

}
