% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{selectPar.pplyFun}
\alias{selectPar.pplyFun}
\title{Select an appropriate *pply function}
\usage{
selectPar.pplyFun(
  parallel.pars,
  .pplyFUN = c("apply", "lapply", "sapply", "mapply")
)
}
\arguments{
\item{parallel.pars}{The output from \code{\link{parallelCheck}}, passing the parallelization options.}

\item{.pplyFUN}{*pply function to be used. Current options are \code{"apply"}, \code{"lapply"}, \code{"sapply"} and \code{"mapply"}.}
}
\value{
A function
}
\description{
Selects an appropriate *pply function depending on the parallel choice
}
\details{
The output function will be either \code{parallel::parLapply} or \code{lapply} for \code{.pplyFUN = "lapply"},
depending of whether parallelization is enabled or not. Same for \code{apply} and \code{parallel::parApply} and so on.
}
\section{Warning}{


From the \code{\link[parallel]{makeCluster}} help: 
\dQuote{It is good practice to shut down the workers by calling \code{stopCluster}}. It is therefore recommended the
following line of code after using this function: \code{on.exit(parallel::stopCluster(parallel.pars$cl))}
}

\seealso{
Other parallel.helpers: 
\code{\link{parallelCheck}()}
}
\author{
J. Bedia
}
\concept{parallel.helpers}
\keyword{internal}
