% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{setGridProj}
\alias{setGridProj}
\title{Set grid projection}
\usage{
setGridProj(grid, proj = NULL)
}
\arguments{
\item{grid}{An input grid}

\item{proj}{Character string. A string specifying the projection attribute value. See details.}
}
\value{
Returns (invisible) the same input grid with the new \code{"projection"} 
attribute in \code{"$xyCoords"} list element.
}
\description{
Set the \code{"projection"} attribute of a climate4R grid
}
\details{
The projection string format is free, although usually a valid epsg code or proj4 string is introduced here.
}
\examples{
\donttest{ 
require(climate4R.datasets)
data(NCEP_Iberia_hus850)
getGridProj(NCEP_Iberia_hus850)
# We introduce the standard definition of WGS84 lat-lon projection
NCEP_Iberia_hus850 <- setGridProj(NCEP_Iberia_hus850,
                                  proj= "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs")
getGridProj(NCEP_Iberia_hus850)                 
}                  
}
\seealso{
Other get.helpers: 
\code{\link{get2DmatCoordinates}()},
\code{\link{getCoordinates}()},
\code{\link{getDim}()},
\code{\link{getGridProj}()},
\code{\link{getGridVerticalLevels}()},
\code{\link{getGrid}()},
\code{\link{getRefDates}()},
\code{\link{getSeason}()},
\code{\link{getShape}()},
\code{\link{getStationID}()},
\code{\link{getTimeResolution}()},
\code{\link{getVarNames}()},
\code{\link{getWT}()},
\code{\link{getYearsAsINDEX}()},
\code{\link{typeofGrid}()},
\code{\link{which.leap}()}
}
\author{
J Bedia
}
\concept{get.helpers}
