% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTimeResolution.R
\name{setTimeResolution}
\alias{setTimeResolution}
\title{Set Time Resolution}
\usage{
setTimeResolution(grid, time_resolution)
}
\arguments{
\item{grid}{grid or station data.}

\item{time_resolution}{time resolution to be set in the grid.}
}
\value{
A C4R grid object that will contain the attribute \code{time_resolution} in grid$Variable.
}
\description{
Set the time resolution for a C4R grid object. Several time resolutions are available.
}
\details{
The available time resolutions are "1h", "3h", "6h", "12h", "DD", "MM", "YY" or "unknown".
}
\author{
O.Mirones
}
