% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdf2clim.R
\name{sgdf2clim}
\alias{sgdf2clim}
\title{Conversion from SpatialGridDataFrame-class objects to climate4R climatology grids}
\usage{
sgdf2clim(
  sp,
  member = FALSE,
  varName = NULL,
  level = NULL,
  dates = list(start = NULL, end = NULL),
  season = NULL,
  attr.list = NULL
)
}
\arguments{
\item{sp}{A SpatialGridDataFrame-class object, as returned by function \code{gdalwarp} (package \pkg{gdalUtils}).
(SpatialPointsDataFrame not implemented yet).}

\item{varName}{Character with the variable name for the resulting grid.}

\item{level}{Character with the level for the resulting grid.}

\item{dates}{Named list of the form: list(start = NULL, end = NULL), for the resulting grid. Dates as POISXct vectors.}

\item{attr.list}{Key-value list of additional attributes to be included in the output grid as metadata.}

\item{seasons}{Integer vector defining the grid season}
}
\value{
A C4R climatological grid (i.e.: time dimension is a singleton), with the metadata specified by the input arguments.
}
\description{
Convert a SpatialGridDataFrame (package \pkg{sp}) to a C4R climatological grid
}
\seealso{
\code{\link[transformeR]{clim2sgdf}}, that performs the inverse operation
}
\author{
M. Iturbide, J. Bedia
}
\keyword{internal}
