% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeUncertainty.R
\name{signal}
\alias{signal}
\title{Calculate if signal exists in the anomalies.}
\usage{
signal(historical, anomaly)
}
\arguments{
\item{historical}{A grid with historical data to calculate the variability threshols (ideally annual data)}

\item{anomaly}{A grid (with time dimension = 1).}
}
\value{
A binary grid (0 = no signal, 1 = signal).
}
\description{
Calculate if signal exists in the anomalies relative to the historical variability.
}
\author{
M. Iturbide
}
