% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridUncertainty.R
\name{signalAgreement}
\alias{signalAgreement}
\title{Calculate signal agreement.}
\usage{
signalAgreement(x, th = 66, condition = "GT")
}
\arguments{
\item{x}{binary numeric}

\item{th}{threshold}

\item{condition}{character (choices: "GT", "GE", "LT", "LE")}
}
\value{
binary numeric (0 = no agreement, 1 = agreement).
}
\description{
Calculate signal agreement.
}
\author{
M. Iturbide
}
