% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindGrid.R
\name{sortDim.time}
\alias{sortDim.time}
\title{Grid sorting by time}
\usage{
sortDim.time(grid)
}
\arguments{
\item{grid}{Input grid}
}
\value{
A grid fo similar characteristics, but with time dimension re-arranged in ascending ordered if needed.
}
\description{
Returns a grid with time sorted in ascending order. For internal use mainly, 
the function is useful for instance when subsetting along time and binding with \code{\link{bindGrid.time}}, 
so the user does not need to worry about the ordering of the input.
}
\seealso{
Other internal.helpers: \code{\link{array3Dto2Dmat.stations}},
  \code{\link{array3Dto2Dmat}},
  \code{\link{bindGrid.member}},
  \code{\link{bindGrid.spatial}},
  \code{\link{bindGrid.time}}, \code{\link{bindGrid}},
  \code{\link{grid3Dto2Dmatrix}},
  \code{\link{mat2Dto3Darray.stations}},
  \code{\link{redim}}, \code{\link{sortDim.spatial}}
}
\author{
J Bedia
}
\concept{internal.helpers}
\keyword{internal}
