% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp2grid.R
\name{sp2grid}
\alias{sp2grid}
\title{Convert a SpatialGridDataFrame to a C4R grid}
\usage{
sp2grid(
  sp,
  member = FALSE,
  varName = NULL,
  level = NULL,
  dates = list(start = NULL, end = NULL),
  metadata = NULL
)
}
\arguments{
\item{sp}{A SpatialGridDataFrame, as returned by function \code{gdalwarp} (package \pkg{gdalUtils}).
(SpatialPointsDataFrame-s not implemented yet).}

\item{varName}{Character with the variable name for the resulting grid.}

\item{level}{Character with the level for the resulting grid.}

\item{dates}{Named list of the form: list(start = NULL, end = NULL), for the resulting grid.}

\item{metadata}{Named list of characters with furher metadata for the resulting grid.}
}
\value{
A C4R grid, with the metadata specified by the input arguments.
}
\description{
Convert a SpatialGridDataFrame (package \pkg{sp}) to a C4R grid
}
\seealso{
\code{\link[transformeR]{clim2sgdf}}
}
\author{
M. Iturbide
}
\keyword{internal}
