% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateGrid.R
\name{spatialAggregation}
\alias{spatialAggregation}
\title{Spatial Aggregation}
\usage{
spatialAggregation(grid, aggr.fun, weight.by.lat, parallel, max.ncores, ncores)
}
\arguments{
\item{grid}{Input grid.}

\item{aggr.fun}{Aggregation function}

\item{weight.by.lat}{Logical flag}
}
\value{
A spatially averaged time series grid (possibly multimember)
}
\description{
Spatial aggregation for rectangular domains.
}
\author{
J Bedia
}
\keyword{internal}
