% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetGrid.R
\name{subsetCluster}
\alias{subsetCluster}
\title{Cluster subsets from a multimember grid}
\usage{
subsetCluster(grid, cluster)
}
\arguments{
\item{grid}{Input multimember grid to be subset (possibly a multimember multigrid).}

\item{cluster}{An integer indicating \strong{the cluster} to be subset.}
}
\value{
A grid (or multigrid) that is a logical subset of the input grid along its 'time' dimension based on the cluster index.
}
\description{
Retrieves a grid that is a logical subset of a multimember grid along its 'time' dimension based on the cluster index.
 Multimember multigrids are supported. Subroutine of \code{\link{subsetGrid}}.
}
\details{
The variable name will be added an extension refering to the cluster extracted.
}
\seealso{
Other subsetting: \code{\link{getTemporalIntersection}},
  \code{\link{intersectGrid.spatial}},
  \code{\link{intersectGrid.time}},
  \code{\link{intersectGrid}}, \code{\link{limitArea}},
  \code{\link{subsetDimension}}, \code{\link{subsetGrid}},
  \code{\link{subsetMembers}}, \code{\link{subsetRuntime}},
  \code{\link{subsetSeason}}, \code{\link{subsetSpatial}},
  \code{\link{subsetStation}}, \code{\link{subsetVar}},
  \code{\link{subsetYears}}
}
\author{
J. A. Fernandez
}
\concept{subsetting}
\keyword{internal}
