% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetDecadalForecast.R
\name{subsetDecadalForecast}
\alias{subsetDecadalForecast}
\title{Subset decadal forecast by initialization years}
\usage{
subsetDecadalForecast(data, ly = c(1, 2))
}
\arguments{
\item{data}{grid obtained from the output of function loadDecadalForecast.}

\item{ly}{Vector of integers indicating the lead years of initialization. 
Default is ly = c(1,2) (the two years before).}
}
\value{
A grid object with an additional dimension corresponding to the selected lead years.
}
\description{
Subset decadal forecast by initialization years
}
\author{
M. Iturbide
}
\keyword{internal}
