% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetGrid.R
\name{subsetDimension}
\alias{subsetDimension}
\title{Arbitrary grid subsetting along one of its dimensions}
\usage{
subsetDimension(grid, dimension = NULL, indices = NULL)
}
\arguments{
\item{grid}{The input grid to be subset. This is either a grid, as returned e.g. by \code{loadGridData} from package \pkg{loadeR} or a
multigrid, as returned by \code{makeMultiGrid}, or other types of multimember grids
(possibly multimember multigrids) as returned e.g. by \code{loadECOMS}, from package \pkg{loadeR.ECOMS}.}

\item{dimension}{Character vector indicating the dimension along which the positions indicated by the \code{indices} paraneter.
Unlike subsetGrid, only one dimension at a time is accepted here}

\item{indices}{An integer vector indicating \strong{the positions} of the dimension to be extracted.
In case of subsetting several dimensions at at time, this corresponds to a list of indices.}
}
\value{
A new grid object that is a logical subset of the input grid along the specified dimension.
}
\description{
Create a new grid/multigrid that is a subset of the input grid 
along the selected dimension.
}
\details{
The attribute \code{subset} will be added taking the value of the \code{dimension} parameter.
}
\examples{
\donttest{
require(climate4R.datasets)
# Example - Member subset
data("CFS_Iberia_tas")
# Selection of members 3 and 7
sub <- subsetDimension(CFS_Iberia_tas,
                       dimension = "member",
                       indices = 1:2)
require(visualizeR)
spatialPlot(climatology(sub), backdrop.theme = "coastline")
}
}
\seealso{
Other subsetting: 
\code{\link{getTemporalIntersection}()},
\code{\link{intersectGrid}()},
\code{\link{intersectGrid.spatial}()},
\code{\link{intersectGrid.time}()},
\code{\link{limitArea}()},
\code{\link{subsetCluster}()},
\code{\link{subsetGrid}()},
\code{\link{subsetMembers}()},
\code{\link{subsetRuntime}()},
\code{\link{subsetSeason}()},
\code{\link{subsetSpatial}()},
\code{\link{subsetStation}()},
\code{\link{subsetVar}()},
\code{\link{subsetYears}()}
}
\author{
J. Bedia and S. Herrera
}
\concept{subsetting}
\keyword{internal}
