% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetGrid.R
\name{subsetRuntime}
\alias{subsetRuntime}
\title{Runtime subsets from a grid}
\usage{
subsetRuntime(grid, runtime)
}
\arguments{
\item{grid}{Input multiruntime grid to be subset.}

\item{runtime}{An integer vector indicating \strong{the position} of the runtimes to be subset.}
}
\value{
A grid (or multigrid) that is a logical subset of the input grid along its 'runtime' dimension.
}
\description{
Retrieves a grid that is a logical subset of a multiruntime grid along its 'runtime' dimension.
 Multiruntime multigrids are supported. Subroutine of \code{\link{subsetGrid}}.
}
\details{
An attribute 'subset' with value 'subsetRuntime' is added to the Runtime slot of the output subset.
}
\seealso{
Other subsetting: 
\code{\link{getTemporalIntersection}()},
\code{\link{intersectGrid.spatial}()},
\code{\link{intersectGrid.time}()},
\code{\link{intersectGrid}()},
\code{\link{limitArea}()},
\code{\link{subsetCluster}()},
\code{\link{subsetDimension}()},
\code{\link{subsetGrid}()},
\code{\link{subsetMembers}()},
\code{\link{subsetSeason}()},
\code{\link{subsetSpatial}()},
\code{\link{subsetStation}()},
\code{\link{subsetVar}()},
\code{\link{subsetYears}()}
}
\author{
M. Iturbide
}
\concept{subsetting}
\keyword{internal}
