% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetGrid.R
\name{subsetVar}
\alias{subsetVar}
\title{Extract a grid from a multigrid object}
\usage{
subsetVar(grid, var)
}
\arguments{
\item{var}{Character vector indicating the variables(s) to be extracted}

\item{multiGrid}{Input multigrid to be subset}

\item{drop}{Logical (default is TRUE). Drop or keep dimensions of length 1.}
}
\value{
Either a (multimember)grid or (multimember)multigrid if one ore more variables
are selected respectively.
}
\description{
Extracts a grid from a multigrid object. Multimember multigrids are supported. Subroutine of subsetGrid.
}
\details{
Argument \code{var} is insensitive to the order of the variables, i.e.: variables
will be always returned in the same order they are in the original multigrid.

An attribute 'subset' with value 'subsetVar' is added to the Variable slot of the output subset.
}
\seealso{
Other subsetting: \code{\link{convert2bin}},
  \code{\link{filterNA}},
  \code{\link{getTemporalIntersection}},
  \code{\link{subsetDimension}}, \code{\link{subsetGrid}},
  \code{\link{subsetMembers}}, \code{\link{subsetRuntime}},
  \code{\link{subsetSeason}}, \code{\link{subsetSpatial}},
  \code{\link{subsetYears}}, \code{\link{train2test}}
}
\author{
J. Bedia
}
\keyword{internal}
