% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train2test.R
\name{train2test}
\alias{train2test}
\title{Splits a dataset into a train and a test dataset}
\usage{
train2test(grid, splitting.size = c(3/4, 1/4), only.index = FALSE)
}
\arguments{
\item{grid}{The dataset to be splitted.}

\item{splitting.size}{A numeric vector of two coordinates indicating the fraction that goes to the train dataset (first coordinate) and to the test dataset (second coordinate). Default is c(3/4,1/4).}

\item{only.index}{If TRUE, returns a list containing both the index of the train data and the test data. If FALSE, returns the both indexes and the new train and test datasets generated. Default is FALSE.}
}
\value{
Depends on the parameter "only.index". A list containing: two grid objects (train and test datasets), and two index vectors (train and test index)
}
\description{
A dataset is divided in order to generate two datasets: train and test. The amount of data that goes for each dataset is also controlled.
}
\details{
It uses the "base" function \code{\link{floor}} to round the fraction of the observations going to a certain dataset. The function also takes care of adjusting dates and other relevant metadata (via the internal \code{\link{subsetDimension}}). 

Appends the attribute \code{subset = "train2test"} in the \code{Variable} element.
}
\examples{
data("VALUE_Iberia_tp")
# Take a look at the dimensions:
getShape(VALUE_Iberia_tp)
new.datasets <- train2test(VALUE_Iberia_tp)
# Take a look at the dimensions of the train and test datasets:
getShape(new.datasets$grid.train)
getShape(new.datasets$grid.test)
# Take a look at the dimensions of the train and test datasets with a different splitting size:
new.datasets2 <- train2test(VALUE_Iberia_tp, spliting.size = c(1/2,1/2))
getShape(new.datasets2$grid.train)
getShape(new.datasets2$grid.test)
}
\seealso{
Other subsetting: \code{\link{convert2bin}},
  \code{\link{filterNA}},
  \code{\link{getTemporalIntersection}},
  \code{\link{subsetDimension}}, \code{\link{subsetGrid}},
  \code{\link{subsetMembers}}, \code{\link{subsetRuntime}},
  \code{\link{subsetSeason}}, \code{\link{subsetSpatial}},
  \code{\link{subsetVar}}, \code{\link{subsetYears}}
}
\author{
J. Bano-Medina
}
