% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{transformer-package}
\alias{transformer}
\alias{transformer-package}
\title{transformer}
\description{
Additional S3 and S4 coercion methods for easy interconversion of
Bioconductor and tidyverse data classes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://transformer.acidgenomics.com/}
  \item \url{https://github.com/acidgenomics/transformer/}
  \item Report bugs at \url{https://github.com/acidgenomics/transformer/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (0000-0002-2403-2221)

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
