% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\name{transformer}
\alias{transformer}
\title{Transformer}
\usage{
transformer(x, d_model, num_heads, dff, mask = NULL)
}
\arguments{
\item{x}{inputs}

\item{d_model}{dimensions of the model}

\item{num_heads}{number of heads}

\item{dff}{dimensions of feed-forward model}

\item{mask}{optional mask}
}
\value{
output of the transformer layer
}
\description{
Transformer
}
\examples{
x <- matrix(rnorm(50 * 512), 50, 512)
d_model <- 512
num_heads <- 8
dff <- 2048

output <- transformer(x, d_model, num_heads, dff)
}
