% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{typeofGrid}
\alias{typeofGrid}
\title{Type of grid}
\usage{
typeofGrid(grid)
}
\arguments{
\item{grid}{An input grid}
}
\value{
A character vector decribing the type of spatial reference. Three disjoint values possible: \code{"station"}, \code{"rotated_grid"} and \code{"regular_grid"}
}
\description{
Describe the type of grid according to the spatial representation of the data
}
\seealso{
Other get.helpers: \code{\link{get2DmatCoordinates}},
  \code{\link{getCoordinates}}, \code{\link{getDim}},
  \code{\link{getGridVerticalLevels}},
  \code{\link{getGrid}}, \code{\link{getRefDates}},
  \code{\link{getSeason}}, \code{\link{getShape}},
  \code{\link{getTimeResolution}},
  \code{\link{getVarNames}}, \code{\link{getYearsAsINDEX}},
  \code{\link{which.leap}}
}
\author{
J Bedia
}
\concept{get.helpers}
\keyword{internal}
