% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upscaleGrid.R
\name{upscaleData}
\alias{upscaleData}
\title{Grid data upscaling for each grid-box}
\usage{
upscaleData(
  grid,
  times = 5,
  aggr.fun = list(FUN = "mean", na.rm = TRUE),
  weight.by.lat = TRUE,
  parallel = FALSE,
  max.ncores = 16,
  ncores = NULL
)
}
\arguments{
\item{grid}{a grid or multigrid.}

\item{times}{numeric. An odd number.}

\item{aggr.fun}{Spatial aggregation function. A list indicating the name of the
aggregation function in first place, and other optional arguments to be passed to the aggregation function.
To be on the safe side, the function in \code{FUN} should be always indicated as a character string.}

\item{weight.by.lat}{Logical. Should latitudinal averages be weighted by the cosine of latitude?.
Default to \code{TRUE}. Ignored if a function different from \code{"mean"} is applied.}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\value{
A grid or multigrid of the same resolution.
}
\description{
Applies an aggregation function (typically the mean) to each grid-box by considering surrounding grid-boxes.
}
\section{Parallel Processing}{


Parallel processing is enabled using the \pkg{parallel} package. 
Parallelization is undertaken by a FORK-type parallel socket cluster formed by \code{ncores}.
If \code{ncores} is not specified (default), \code{ncores} will be one less than the autodetected number of cores.
The maximum number of cores used for parallel processing can be set with the \code{max.ncores} argument, 
although this will be reset to the auto-detected number of cores minus 1 if this number is exceeded. Note that not all 
code, but just some critical loops within the function are parallelized.

In practice, parallelization does not always result in smaller execution times, due to the parallel overhead.
However, parallel computing may potentially provide a significant speedup for the 
particular case of large multimember datasets or large grids.
 
Parallel computing is currently not available for Windows machines.
}

\author{
M. Iturbide
}
