% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upscaleGrid.R
\name{upscaleGrid}
\alias{upscaleGrid}
\title{Flexible grid aggregation along selected dimensions}
\usage{
upscaleGrid(grid, times = 2, aggr.fun = list(FUN = max, na.rm = TRUE))
}
\arguments{
\item{grid}{a grid or multigrid to be aggregated.}

\item{times}{Degradation times. For instance, if the resolution of \code{grid}
is 1 degree and \code{times = 2}, the resolution of the output grid will be 2 
degrees.}

\item{aggr.fun}{Upscale function. Default retains the maximum value.}
}
\value{
A grid or multigrid aggregated along the chosen dimension(s).
}
\description{
Aggregates a grid along the target dimensions using user-defined functions.
}
\details{
The development of a more eficient and fast version of the function is planned.
}
\examples{
data("EOBS_Iberia_pr")
library(visualizeR)
spatialPlot(climatology(EOBS_Iberia_pr))
newgrid <- upscaleGrid(redim(EOBS_Iberia_pr), 
                       times = 2,
                       aggr.fun = list(FUN = "min", na.rm = TRUE))
spatialPlot(climatology(newgrid))
}
\author{
M. Iturbide
}
