% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{VALUE_Iberia_pr}
\alias{VALUE_Iberia_pr}
\title{Station data containing VALUE (ECA-D) daily data of precipitation for the Iberian Peninsula (DJF, 1983-2002).}
\format{Station data}
\source{
\url{http://www.meteo.unican.es/udg-tap}

Subset of VALUE (ECA-D) station data. Full dataset is accessible 
for download in \url{http://meteo.unican.es/work/downscaler/data/VALUE_ECA_86_v2.tar.gz}.
}
\description{
This R data object was obtained by means of function \code{\link[loadeR]{loadStationData}} 
(package \href{https://github.com/SantanderMetGroup/loadeR}{\code{loadeR}}) in the following manner:


\code{years <- 1983:2002}\cr 
\code{season <- c(12,1,2)}\cr 
\code{lonLim <- c(-10, 4)}\cr 
\code{latLim <- c(35, 44)}\cr 

\code{VALUE_Iberia_pr <- loadStationData(dataset = "data/VALUE_ECA_86_v2.zip",
var = "precip", years = years, lonLim = lonLim, latLim = latLim, season = season)}\cr

Station data from the VALUE (ECA-D) dataset containing daily precipitation for stations in the Iberian Peninsula.
}
\examples{
data("VALUE_Iberia_pr")
plotClimatology(climatology(VALUE_Iberia_pr),
                backdrop.theme = "countries", scales = list(draw = TRUE))
}
\seealso{
\code{\link{makeMultiGrid}}, \code{\link[loadeR]{loadGridData}}
}
\keyword{Station}
\keyword{data}
