% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeUncertainty.R
\name{vth.aux}
\alias{vth.aux}
\title{Compute variability threshold}
\usage{
vth.aux(x, period = 20)
}
\arguments{
\item{x}{numeric}

\item{period}{value of the number of years to consider as a period (default is 20)}
}
\value{
numeric. A single value of the variability threshold as defined in the IPCC AR6 WGI (Atlas chapter)
}
\description{
Compute variability threshold
}
\author{
M. Iturbide
}
