% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{which.leap}
\alias{which.leap}
\title{Leap years}
\usage{
which.leap(years)
}
\arguments{
\item{years}{a integer vector of (gregorian) years}
}
\value{
a vector of indices of the position of leap years
}
\description{
Return the indices of leap years in a vector of years
}
\examples{
leap.years <- which.leap(1885:1937)
(1885:1937)[leap.years]
}
\references{
\url{https://en.wikipedia.org/wiki/Leap_year#Algorithm}
}
\seealso{
Other get.helpers: \code{\link{get2DmatCoordinates}},
  \code{\link{getCoordinates}}, \code{\link{getDim}},
  \code{\link{getGridProj}},
  \code{\link{getGridVerticalLevels}},
  \code{\link{getGrid}}, \code{\link{getRefDates}},
  \code{\link{getSeason}}, \code{\link{getShape}},
  \code{\link{getStationID}},
  \code{\link{getTimeResolution}},
  \code{\link{getVarNames}}, \code{\link{getYearsAsINDEX}},
  \code{\link{typeofGrid}}
}
\author{
J. Bedia
}
\concept{get.helpers}
\keyword{internal}
