% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomize-methods.R
\docType{methods}
\name{atomize}
\alias{atomize}
\alias{atomize,data.frame-method}
\alias{atomize,DataFrame-method}
\alias{atomize,GRanges-method}
\title{Atomize}
\usage{
\S4method{atomize}{data.frame}(object)

\S4method{atomize}{DataFrame}(object)

\S4method{atomize}{GRanges}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Modified object.
Non-atomic columns (e.g. \code{list} or complex S4 classes) will be dropped.
}
\description{
Ensure all columns in the object are \code{atomic}.
}
\examples{
df <- DataFrame(a = "a", b = list(a = seq_len(3)))
lapply(df, is.atomic)
x <- atomize(df)
}
