% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindGrid.member.R
\name{bindGrid.member}
\alias{bindGrid.member}
\title{Grid binding by member dimension}
\usage{
bindGrid.member(..., spatial.tolerance = 0.001)
}
\arguments{
\item{...}{Input grids to bind by their member dimension. These must be compatible in time and space (see details).}

\item{spatial.tolerance}{numeric. Coordinate differences smaller than \code{spatial.tolerance} will be considered equal 
coordinates. Default to 0.001 --assuming that degrees are being used it seems a reasonable rounding error after interpolation--.
This value is passed to the \code{\link{identical}} function to check for spatial consistency of the input grids.}
}
\description{
Flexible binding of (spatiotemporally consistent) grids by their member dimension
 useful to handle sets of predictors as a single block.
}
\details{
The function can be useful for handling loading large domains and multimembers, that are difficult or impossible to load
 at once with the loading functions. The task of loading can be efficiently parallelized by splitting the request by
 ensemble member subsets, and then binded with this function. 

The function makes a number of checks in order to test the spatiotemporal compatibility of the input multi-member grids.
 The spatial consistency of the input grids is also checked. In order to avoid possible errors from the user, the spatial
  consistency (i.e., equal XY coordinates) of the input grids must be ensured before attempting the creation of the multigrid,
  otherwise giving an error. This can be achieved either through the specification of the same 'lonLim' and 'latLim' argument
  values when loading the grids, or using the \code{\link{interpGrid}} interpolator in conjuntion with the \code{\link{getGrid}}
  method.
}
\examples{
data("tasmin_forecast")
# We first diaggregate in various grids with different members
members1_2 <- subsetGrid(tasmin_forecast, members = 1:2)
members3_4 <- subsetGrid(tasmin_forecast, members = 3:4)
member7 <- subsetGrid(tasmin_forecast, members = 7)
member8 <- subsetGrid(tasmin_forecast, members = 8)
# The function is insensitive to the number of members per input grid
bindedGrid <- bindGrid.member(members1_2, members3_4, member7, member8)
plotClimatology(climatology(bindedGrid), backdrop.theme = "coastline")
}
\author{
J Bedia
}

