% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindGrid.R
\name{bindGrid.member}
\alias{bindGrid.member}
\title{Grid binding by member dimension}
\usage{
bindGrid.member(..., tol, attr.)
}
\arguments{
\item{...}{Input grids to bind by their member dimension. These must be compatible in time and space (see details).
For flexibility, they can be introduced as a list or directly as consecutive arguments.}

\item{tol}{numeric. Coordinate differences smaller than \code{spatial.tolerance} will be considered equal 
coordinates. Default to 0.001 --assuming that degrees are being used it seems a reasonable rounding error after interpolation--.
This value is passed to the \code{\link{identical}} function to check for spatial consistency of the input grids.}

\item{attr.}{Character string used to update the \code{"dataset"} attribute of the output grid. This
is used only when binding grids along the \code{"member"} dimension, in case multi-model combination is being undertaken,
(i.e., different models are being joined as members of a single multimodel dataset)
so the new dataset attribute reflects this. Note that the default behaviour is retaining the \code{"dataset"} 
attribute of the first grid given, resulting in a misleading metadata information if this is the case. Default to \code{NULL},
and ignored.}
}
\description{
Flexible binding of (spatiotemporally consistent) grids by their member dimension
 useful to handle sets of predictors as a single block.
}
\examples{
data("CFS_Iberia_tas")
# We first diaggregate in various grids with different members
members1_2 <- subsetGrid(CFS_Iberia_tas, members = 1:2)
members3_4 <- subsetGrid(CFS_Iberia_tas, members = 3:4)
member7 <- subsetGrid(CFS_Iberia_tas, members = 7)
member8 <- subsetGrid(CFS_Iberia_tas, members = 8)
# The function is insensitive to the number of members per input grid
bindedGrid <- bindGrid.member(members1_2, members3_4, member7, member8)
plotClimatology(climatology(bindedGrid), backdrop.theme = "coastline")
}
\seealso{
Other internal.helpers: \code{\link{array3Dto2Dmat.stations}},
  \code{\link{array3Dto2Dmat}},
  \code{\link{bindGrid.spatial}},
  \code{\link{bindGrid.time}},
  \code{\link{grid3Dto2Dmatrix}},
  \code{\link{mat2Dto3Darray.stations}},
  \code{\link{redim}}, \code{\link{sortDim.spatial}},
  \code{\link{sortDim.time}}
}
\author{
J Bedia
}
