% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindGrid.R
\name{bindGrid.spatial}
\alias{bindGrid.spatial}
\title{Grid binding by spatial dimension}
\usage{
bindGrid.spatial(..., dimn, tol)
}
\arguments{
\item{...}{Input grids to bind by their time dimension. These must be compatible in members and space (see details).
For flexibility, they can be introduced as a list or directly as consecutive arguments.}

\item{dimn}{latitude ("lat") or longitude ("lon")}

\item{tol}{numeric. Coordinate differences smaller than \code{spatial.tolerance} will be considered equal 
coordinates. Default to 0.001 --assuming that degrees are being used it seems a reasonable rounding error after interpolation--.
This value is passed to the \code{\link{identical}} function to check for spatial consistency of the input grids.}
}
\description{
Flexible binding of (spatially/member consistent) grids by their longitude or latitude dimension
}
\seealso{
Other internal.helpers: 
\code{\link{array3Dto2Dmat}()},
\code{\link{array3Dto2Dmat.stations}()},
\code{\link{bindGrid.member}()},
\code{\link{bindGrid.time}()},
\code{\link{grid3Dto2Dmatrix}()},
\code{\link{mat2Dto3Darray.stations}()},
\code{\link{redim}()},
\code{\link{sortDim.spatial}()},
\code{\link{sortDim.time}()}
}
\author{
M Iturbide
}
\concept{internal.helpers}
