% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkDim}
\alias{checkDim}
\title{Spatial check of input grids}
\usage{
checkDim(..., dimensions = c("var", "member", "time", "lat", "lon"))
}
\arguments{
\item{...}{Input grids to be compared}

\item{dimension}{Character vector. Dimensions to check for consistency. Several dimensions can be tested at once.
Possible values are: \code{"var"}, \code{"member"}, \code{"time"}, \code{"lat"} and \code{"lon"}.}
}
\value{
In case of spatial inconsistency of any of the inputs grids, 
the function stops the execution of the current expression, with an error message.
}
\description{
Checks that the input grids have the same lon/lat shape

The function simply compares the size of the dimension across input grids for 
structural consistency of their respective data arrays. It does not make any additional checks on the metadata,
so equal array shapes do not entail full temporal/spatial consistency.
}
\examples{
data("VALUE_Iberia_tas")
data("VALUE_Iberia_pr")
checkDim(VALUE_Iberia_tas, VALUE_Iberia_pr) # ok, go on
data("EOBS_Iberia_tas")
data("EOBS_Iberia_pr")
checkDim(EOBS_Iberia_tas, EOBS_Iberia_pr) # ok, go on
data("NCEP_Iberia_psl")
try(checkDim(EOBS_Iberia_tas, EOBS_Iberia_pr, NCEP_Iberia_psl)) # inconsistent dimensions
data("CFS_Iberia_tas")
try(checkDim(NCEP_Iberia_psl, CFS_Iberia_tas, dimensions = "member")) 
checkDim(VALUE_Iberia_tas, VALUE_Iberia_pr) # ok, go on
}
\seealso{
\code{\link{getShape}}

Other check.helpers: \code{\link{checkSeason}},
  \code{\link{checkTemporalConsistency}},
  \code{\link{checkVarNames}}
}
\author{
J Bedia
}
\concept{check.helpers}
\keyword{internal}
