% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkGrid}
\alias{checkGrid}
\title{Check the structure of the grid}
\usage{
checkGrid(...)
}
\arguments{
\item{...}{Input grids (including station data)}
}
\value{
Diagnostic message
}
\description{
Check the structure of the grid
}
\seealso{
Other check.helpers: 
\code{\link{checkDim}()},
\code{\link{checkSeason}()},
\code{\link{checkTemporalConsistency}()},
\code{\link{checkVarNames}()}
}
\author{
M. Iturbide
}
\concept{check.helpers}
