% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgdf2clim.R
\name{clim2sgdf}
\alias{clim2sgdf}
\title{Climatology to SpatialGridDataFrame or SpatialPointsDataFrame}
\usage{
clim2sgdf(clim, set.min = NULL, set.max = NULL)
}
\arguments{
\item{clim}{A climatological grid, as returned by function \code{\link{climatology}}}

\item{set.min}{Minimum value, as passed by \code{\link[visualizeR]{spatialPlot}}}

\item{set.max}{Maximum value, as passed by \code{\link[visualizeR]{spatialPlot}}}
}
\value{
A \pkg{sp} object of the class \code{\link[sp]{SpatialGridDataFrame}}
}
\description{
Convert a climatological climate4R grid to a SpatialGridDataFrame object from package sp. See details.
}
\details{
This function is intended for internal usage by \code{\link[visualizeR]{spatialPlot}},
that accepts all possible arguments that can be passed to \code{\link[sp]{spplot}} for plotting. 
However, it may be useful for advanced \pkg{sp} users in different contexts
(e.g. for reprojecting via \code{\link[sp]{spTransform}} etc.).

Note that the function can be only applied to \dQuote{climatological} climate4R grids, i.e., grids
 whose time dimension is a singleton, typically after the application of function 
 \code{\link{climatology}}, although not necessarily so.
}
\examples{
\donttest{
data("CFS_Iberia_tas")
# Climatology is computed:
clim <- climatology(CFS_Iberia_tas, by.member = TRUE)
sgdf <- clim2sgdf(clim, NULL, NULL)
class(sgdf)
sp::spplot(sgdf)
}
}
\seealso{
\code{\link{climatology}}, \code{\link[visualizeR]{spatialPlot}}

\code{sgdf2clim}, to perform the reverse operation
}
\author{
J. Bedia
}
\keyword{internal}
