% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotClimatology.R
\name{clim2sgdf}
\alias{clim2sgdf}
\title{Climatology to SpatialGridDataFrame}
\usage{
clim2sgdf(clim)
}
\arguments{
\item{clim}{A climatological grid, as returned by function \code{\link{climatology}}}
}
\value{
A \pkg{sp} object of the class \code{\link[sp]{SpatialGridDataFrame}}
}
\description{
Convert a climatological grid to a SpatialGridDataFrame object from package sp
}
\details{
This function is intended for internal usage by \code{\link{plotClimatology}},
that accepts all possible arguments that can be passed to \code{\link[sp]{spplot}} for plotting.
However, it may be useful for advanced \pkg{sp} users in different contexts
(e.g. for reprojecting via \code{\link[sp]{spTransform}} etc.)
}
\examples{
\donttest{
data(tasmax_forecast)
# Climatology is computed:
clim <- climatology(tasmax_forecast, by.member = TRUE)
sgdf <- clim2sgdf(clim)
class(sgdf)
spplot(sgdf)
}
}
\author{
J. Bedia
}
\seealso{
\code{\link{climatology}}, \code{\link{plotClimatology}}
}
\keyword{internal}

