% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGrid.R
\name{clusterGrid}
\alias{clusterGrid}
\title{Cluster analysis}
\usage{
clusterGrid(
  grid,
  type = "kmeans",
  centers = NULL,
  newdata = NULL,
  y = NULL,
  ...
)
}
\arguments{
\item{grid}{A grid (gridded or station dataset), multigrid, multimember grid or multimember multigrid object}

\item{type}{Clustering algorithm. Possible values are \code{"kmeans"} (default), \code{"hierarchical"}, \code{"som"} and \code{"lamb"}. See Details.}

\item{centers}{Integer value indicating the number of clusters, \strong{k}, or center points. See Details.}

\item{newdata}{Optional grid containing the data for prediction. It must contain the same variables as the input \code{grid} taken as reference.}

\item{y}{Optional predictand grid data. Clustering Analysis of this grid will be performed as a day-by-day correspondence with 
the reference grid (\code{grid} or \code{newdata}). Thus time dimension from \code{y} and the reference grid must intersect. For weather typing.}

\item{...}{Further specific arguments passed to the different clustering functions.}
}
\value{
A C4R (multimember/multi) grid object that will contain the data from: 
\itemize{
\item The input \code{grid}, if neither \code{newdata} nor \code{y} are indicated.  
\item The \code{newdata} grid if this is specified, and \code{y = NULL}. 
\item The \code{y} grid, if this is specified.
}
The clustering type (cluster.type), number of clusters (centers), centroids of clusters from 'grid' input (centroids) and the days corresponding to each 
cluster (index) are returned as attributes in all clustering algorithms. Then, other algorithm-specific parameters are provided as attributes.
}
\description{
Cluster analysis of climate data. Several clustering algorithms are available.
}
\details{
\strong{kmeans}

While using the K-means algorithm, the number of clusters (argument \code{centers}) needs to be provided (with no default). 
See the \code{\link[stats]{kmeans}} documentation for further details and optional arguments of the method.

\strong{hierarchical}

While using the hierarchical algorithm (check \code{\link[stats]{hclust}} for further information) 
\code{clusterGrid} allows the user either to especify a predefined number of clusters (\code{centers}) or not. 
If the argument \code{centers} is not provided, they are automatically set and the tree is cut when the height 
difference between two consecutive divisions (sorted in ascending order) is larger than the interquartile 
range of the heights vector, as determined by \code{\link[stats]{cutree}}. 

\strong{som}

While using the SOM (self-organizing maps) algorithm (check \code{\link[kohonen]{som}} for further information), the argument \code{centers} is provided as
a two-element vector, indicating the dimensions \code{xdim,ydim} of the grid (see \code{\link[kohonen]{somgrid}}).
By default, a rectangular topology (8x6) of 48 clusters is obtained. 

\strong{Lamb}

The Lamb Weather Typing algorithm (Lamb 1972) is implemented in the function \code{\link{lambWT}}. The argument \code{centers} is not used.
A default of 26 types is calculated, following Trigo and daCamara (2000)
}
\examples{
\donttest{
require(climate4R.datasets)
#Example of K-means clustering: 
data(NCEP_Iberia_psl)
clusters<- clusterGrid(NCEP_Iberia_psl, type="kmeans", centers=10, iter.max=1000)

#Example of hierarchical clustering: 
clusters<- clusterGrid(NCEP_Iberia_psl, type="hierarchical")

#Example of som clustering: 
clusters<- clusterGrid(NCEP_Iberia_psl, type="som", centers = c(10,1))

#Example of lamb clustering:
data(NCEP_slp_2001_2010)
clusters <- clusterGrid(grid = NCEP_slp_2001_2010, type = "lamb")
}
}
\references{
Lamb, H., 1972. British Isles weather types and a register of the daily sequence of circulation patterns.

Trigo, R.M., DaCamara, C.C., 2000. Circulation weather types and their influence on the precipitation regime in Portugal. 
Int. J. Climatol. 23. https://doi.org/10.1002/1097-0088(20001115)20:13%3C1559::AID-JOC555%3E3.0.CO;2-5
}
\seealso{
\code{\link[stats]{kmeans}}, \code{\link[stats]{hclust}}, \code{\link[kohonen]{som}}.
}
\author{
J. A. Fernández
}
