% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGrid.R
\name{clusterGrid}
\alias{clusterGrid}
\title{Cluster analysis of grids}
\usage{
clusterGrid(grid, type = "kmeans", centers = NULL, iter.max = 10,
  nstart = 1, method = "complete")
}
\arguments{
\item{grid}{A grid (gridded or station dataset), multigrid, multimember grid or multimember multigrid object, as 
returned e.g. by \code{loadeR::loadGridData} (or \code{loadeR::loadStationData}), a
multigrid, as returned by \code{makeMultiGrid}, or other types of multimember grids
(possibly multimember grids) as returned e.g. by \code{loadeR.ECOMS::loadECOMS}.}

\item{type}{Clustering algorithm to be used for the cluster analysis. 
Possible values are "\strong{kmeans}" (default), "\strong{hierarchical}", "\strong{som}". 
The core functions are \link[stats]{kmeans}, \link[stats]{hclust}, \link[kohonen]{som}, respectively. See Details.}

\item{centers}{Integer value indicating the number of clusters, \strong{k}, or center points. See Details.}

\item{iter.max}{(for the K-means algorithm) Integer value indicating the maximum number of iterations allowed. Default: 10.}

\item{nstart}{(for the K-means algorithm) If centers is a number, how many random sets should be chosen? Default: 1.}

\item{method}{(for the hierarchical algorithm) Agglomeration method to be used, one of "complete" (default), "ward.D", "ward.D2", "single",
"average", "mcquitty", "median" or "centroid".}
}
\value{
A new C4R grid object that contains the clusters created using the specified algorithm. Clusters are included in the dimension 'time'.
The clustering type, number of clusters and other algorithm-specific parameters are provided as attributes.
}
\description{
Performs cluster analysis of grids, multigrids or multimember multigrids. Several clustering algorithms are available.
}
\details{
\strong{kmeans}

While using the  K-means algorithm, the number of clusters (argument 'centers') needs to be provided (no default). 
See  \link[stats]{kmeans} for more details in the implementation.

\strong{hierarchical}

While using the hierarchical algorithm (check \link[stats]{hclust} for further information) 
\code{clusterGrid} allows the user either to especify the number of clusters ('centers') or not. 
If the argument 'centers' is not provided, they are automatically set and the tree is cut when the height 
difference between two consecutive divisions (sorted in ascending order) is larger than the intercuartile 
range of the heights vector (see \link[stats]{cutree}) . 

\strong{som}

While using the SOM (self-organizing maps) algorithm (check \link[kohonen]{som} for further information), the argument 'centers' is provided as
a two-element vector, indicating the dimensions \code{xdim,ydim} of the grid (see \link[kohonen]{somgrid}).
Otherwise, by default 48 clusters (8x6) with rectangular topology are obtained.
}
\examples{
#Example of K-means clustering: 
data(NCEP_Iberia_psl, package = "transformeR")
clusters<- clusterGrid(NCEP_Iberia_psl, type="kmeans", centers=10, iter.max=1000)

#Example of hierarchical clustering: 
clusters<- clusterGrid(NCEP_Iberia_psl, type="hierarchical")

#Example of som clustering: 
clusters<- clusterGrid(NCEP_Iberia_psl, type="som", centers = c(10,1))

#Example of K-means clustering of several variables:
data(NCEP_Iberia_ta850, package = "transformeR")
clusters<- clusterGrid(makeMultiGrid(NCEP_Iberia_psl, NCEP_Iberia_ta850), type="kmeans", centers=10, iter.max=1000)
}
\seealso{
\link[stats]{kmeans}, \link[stats]{hclust}, \link[kohonen]{som}.
}
\author{
J. A. Fernandez
}
