% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-DataFrame-S4methods.R
\name{coerce-DataFrame}
\alias{coerce-DataFrame}
\alias{coerce,data.table,DataFrame-method}
\alias{coerce,Matrix,DataFrame-method}
\alias{coerce,tbl_df,DataFrame-method}
\title{Coerce to data frame}
\description{
Coerce to \code{DataFrame}.
}
\examples{
data(dt, sparse, tbl, package = "acidtest")

## data.table to DataFrame ====
x <- as(dt, "DataFrame")
head(x)

## Matrix to DataFrame ====
x <- as(sparse, "DataFrame")

## tbl_df to DataFrame ====
x <- as(tbl, "DataFrame")
head(x)
}
