% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-DataFrame.R
\name{coerce-DataFrame}
\alias{coerce-DataFrame}
\alias{as.DataFrame}
\alias{coerce,sparseMatrix,DataFrame-method}
\alias{coerce,tbl_df,DataFrame-method}
\title{Coerce to \code{DataFrame}}
\usage{
as.DataFrame(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Coerce to \code{DataFrame}
}
\examples{
## sparseMatrix to DataFrame ====
data(sparse, package = "acidtest")
stopifnot(is(sparse, "sparseMatrix"))
x <- as(sparse, "DataFrame")

## tbl_df to DataFrame ====
data(tbl, package = "acidtest")
stopifnot(is(tbl, "tbl_df"))
x <- as(tbl_df, "DataFrame")
head(x)
}
