% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-DataFrame.R
\name{coerce-DataFrame}
\alias{coerce-DataFrame}
\alias{as.DataFrame}
\alias{coerce,data.table,DataFrame-method}
\alias{coerce,sparseMatrix,DataFrame-method}
\alias{coerce,tbl_df,DataFrame-method}
\title{Coerce to \code{DataFrame}}
\usage{
as.DataFrame(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Coerce to \code{DataFrame}
}
\examples{
data(dt, sparse, tbl, package = "acidtest")

## data.table to DataFrame ====
stopifnot(is(dt, "data.table"))
x <- as(dt, "DataFrame")
head(x)

## sparseMatrix to DataFrame ====
stopifnot(is(sparse, "sparseMatrix"))
x <- as(sparse, "DataFrame")

## tbl_df to DataFrame ====
stopifnot(is(tbl, "tbl_df"))
x <- as(tbl, "DataFrame")
head(x)
}
