% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-DataFrame-methods.R
\name{coerce-DataFrame}
\alias{coerce-DataFrame}
\alias{coerce,data.table,DataFrame-method}
\alias{coerce,Matrix,DataFrame-method}
\alias{coerce,tbl_df,DataFrame-method}
\title{Coerce to data frame}
\value{
\code{DataFrame}.
}
\description{
Coerce to \code{DataFrame}.
}
\examples{
data(data.table, sparseMatrix, tbl_df, package = "acidtest")

## data.table to DataFrame ====
x <- as(data.table, "DataFrame")
head(x)

## Matrix to DataFrame ====
x <- as(sparseMatrix, "DataFrame")

## tbl_df to DataFrame ====
x <- as(tbl_df, "DataFrame")
head(x)
}
