% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-SummarizedExperiment-S3methods.R
\name{coerce-SummarizedExperiment}
\alias{coerce-SummarizedExperiment}
\alias{as.SummarizedExperiment}
\alias{as.SummarizedExperiment.RangedSummarizedExperiment}
\alias{as.SummarizedExperiment.SummarizedExperiment}
\title{Coerce to summarized experiment}
\usage{
as.SummarizedExperiment(x)

\method{as.SummarizedExperiment}{RangedSummarizedExperiment}(x)

\method{as.SummarizedExperiment}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{Object.}
}
\description{
Coerce to \code{SummarizedExperiment}.
}
\details{
Improved S3 methods for reliably coercing objects that extend
\code{RangedSummarizedExperiment} to a standard \code{SummarizedExperiment}, that
doesn't drop \code{\link[SummarizedExperiment:rowData]{rowData()}}.

We can't set S4 coercion methods here because RangedSummarizedExperiment and
SummarizedExperiment aren't local, modifiable.

Related S4 coercion method of interest:\preformatted{getMethod(
    f = "coerce",
    signature = signature(
        from = "RangedSummarizedExperiment",
        to = "SummarizedExperiment"
    ),
    where = asNamespace("SummarizedExperiment")
)
}
}
\examples{
suppressPackageStartupMessages(library(SummarizedExperiment))
data(rse, package = "acidtest")

## RangedSummarizedExperiment ===
## Current S4 coercion drops row metadata.
names(metadata(rowRanges(rse)))
se <- as(rse, "SummarizedExperiment")
names(metadata(rowData(se)))
## Our alternate S3 method preserves the metadata.
se <- as.SummarizedExperiment(rse)
names(metadata(rowData(se)))
}
