% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tibble-methods.R, R/coerce-tbl_df-methods.R
\name{as_tibble}
\alias{as_tibble}
\alias{as_tibble.DataFrame}
\alias{as_tibble.IPosRanges}
\alias{as_tibble.GenomicRanges}
\alias{coerce-tbl_df}
\alias{coerce,data.frame,tbl_df-method}
\alias{coerce,DataFrame,tbl_df-method}
\alias{coerce,GenomicRanges,tbl_df-method}
\alias{coerce,IPosRanges,tbl_df-method}
\title{Coerce to tibble}
\usage{
\method{as_tibble}{DataFrame}(x, ...,
  rownames = pkgconfig::get_config("tibble::rownames", "rowname"))

\method{as_tibble}{IPosRanges}(x, ...,
  rownames = pkgconfig::get_config("tibble::rownames", "rowname"))

\method{as_tibble}{GenomicRanges}(x, ...,
  rownames = pkgconfig::get_config("tibble::rownames", "rowname"))
}
\arguments{
\item{x}{A data frame, list, matrix, or other object that could reasonably be
coerced to a tibble.}

\item{...}{Other arguments passed on to individual methods.}

\item{rownames}{How to treat existing row names of a data frame or matrix:
\itemize{
\item \code{NULL}: remove row names. This is the default.
\item \code{NA}: keep row names.
\item A string: the name of a new column. Existing rownames are transferred
into this column and the \code{row.names} attribute is deleted.
Read more in \link{rownames}.
}}
}
\value{
\code{tbl_df}.
}
\description{
Coerce to \code{tbl_df}.
}
\details{
Our defined methods attempt to improve on the defaults in the tibble package
to ensure that row names are not dropped by default, which is a poor default
for bioinformatics. This is accomplished by setting \code{rownames = "rowname"} by
default instead of \code{rownames = NULL}.
}
\section{S3 \code{as_tibble()}}{


transformer extends \code{\link[tibble:as_tibble]{as_tibble()}} method support for
these S4 classes:
\itemize{
\item \code{DataFrame}.
\item \code{GRanges}.
}
}

\section{S4 \code{as()}}{


Since \code{tbl_df} is a virtual class that extends \code{tbl} and \code{data.frame}, we
need to define an S4 coercion method that allows us to use
\code{\link[methods:as]{as()}} to coerce an object to a tibble.

See \code{getClass("tbl_df")} for details on how tibble is a virtual class.
}

\examples{
data(DataFrame, GRanges, IRanges, package = "acidtest")

## DataFrame to tbl_df ====
x <- as(DataFrame, "tbl_df")
x <- as_tibble(DataFrame)
print(x)

## GenomicRanges to tbl_df ====
x <- as(GRanges, "tbl_df")
x <- as_tibble(GRanges)
print(x)

## IRanges to tbl_df ====
x <- as(IRanges, "tbl_df")
x <- as_tibble(IRanges)
print(x)
}
\seealso{
\code{\link[tibble:as_tibble]{tibble::as_tibble()}}.
}
