% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prinComp.R
\name{combine.PCs}
\alias{combine.PCs}
\title{Field Combination}
\usage{
combine.PCs(Xsc.list)
}
\arguments{
\item{Xsc.list}{A nested list of (scaled and centered) input variables (as returned by \code{\link{prinComp.scale}})}
}
\value{
A list of members with 2D combined scaled and centered matrices, resulting from the combination of the
input fields. A special attribute \code{"combined_variables"} indicated the variables combined
}
\description{
Combine different fields into one single input 2D matrix for PCA (internal of \code{\link{prinComp}})
}
\author{
J Bedia
}
\keyword{internal}
