% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2bin.R
\name{convert2bin}
\alias{convert2bin}
\title{Convert grid values to a binary variable}
\usage{
convert2bin(x, threshold = 0.5, partial = FALSE)
}
\arguments{
\item{threshold}{Upon or equal the threshold the grid values turn to 1, whereas below it turns to 0. Default is 0.5.}

\item{partial}{If TRUE, only values below the threshold will turn to 0 and the others will remain with their original grid real value. Default is FALSE.}

\item{grid}{The input grid to be converted to binary. It can be an object or a matrix/array.}
}
\value{
A new grid object with binary values
}
\description{
Convert a real variable into a binary variable (i.e., 0 or 1) filtering by a threshold.
}
\details{
The function only works for a vector of observations/predictands downloaded from \code{\link{loadeR}}).

Appends the attribute \code{subset = "convert2bin"} in the \code{Variable} element.
}
\examples{
data("VALUE_Iberia_tp")
# Take a look at the data:
head(VALUE_Iberia_tp$Data)
# Convert to complete binary variable:
bin.total <- convert2bin(VALUE_Iberia_tp,threshold = 1)
head(bin.total$Data)
# Convert to partial binary variable:
bin.partial <- convert2bin(VALUE_Iberia_tp,threshold = 1, partial = TRUE)
head(bin.partial$Data)
}
\seealso{
Other subsetting: \code{\link{filterNA}},
  \code{\link{getTemporalIntersection}},
  \code{\link{subsetDimension}}, \code{\link{subsetGrid}},
  \code{\link{subsetMembers}}, \code{\link{subsetRuntime}},
  \code{\link{subsetSeason}}, \code{\link{subsetSpatial}},
  \code{\link{subsetVar}}, \code{\link{subsetYears}},
  \code{\link{train2test}}
}
\author{
J. Bano-Medina
}
