% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-methods.R
\docType{methods}
\name{decode}
\alias{decode}
\alias{decode,DataFrame-method}
\alias{decode,GRanges-method}
\title{Decode column data that uses run-length encoding}
\usage{
\S4method{decode}{DataFrame}(x)

\S4method{decode}{GRanges}(x)
}
\arguments{
\item{x}{Object.}
}
\value{
Modified object.
Columns will be decoded and no longer \code{Rle} class.
}
\description{
The Rle class is a general container for storing an atomic vector
  that is stored in a run-length encoding format. It is based on the
  \code{\link[base]{rle}} function from the base package.
}
\examples{
load(system.file("extdata", "rse.rda", package = "transformer"))
mcols <- S4Vectors::mcols(rse)
lapply(mcols, class)
x <- decode(mcols)
lapply(x, class)
}
\seealso{
\code{S4Vectors::decode()}.
}
