% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplevels-S3methods.R
\name{droplevels}
\alias{droplevels}
\alias{droplevels.SummarizedExperiment}
\title{Drop unused levels from factors}
\usage{
\method{droplevels}{SummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Factor levels will be readjusted (i.e. superfluous levels are dropped).
}
\description{
Drop unused levels from factors
}
\note{
Additional useful methods for S4 objects are defined in S4Vectors.

Updated 2019-08-15.
}
\examples{
data(RangedSummarizedExperiment, package = "acidtest")

## DataFrame ====
object <- RangedSummarizedExperiment
object <- droplevels(object)
}
