% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdas.R
\docType{data}
\name{EOBS_Iberia_tas}
\alias{EOBS_Iberia_tas}
\title{Grid containing E-OBS daily data of mean temperature for the Iberian Peninsula (DJF, 1983-2002).}
\format{A grid}
\source{
\url{http://www.meteo.unican.es/udg-tap}
}
\description{
This R data object was obtained from the UDG server (\url{http://www.meteo.unican.es/udg-tap}, 
log in is requiered, 
see \code{\link[loadeR]{loginUDG}}) by means of function \code{\link[loadeR]{loadGridData}} 
(package \href{https://github.com/SantanderMetGroup/loadeR}{\code{loadeR}}) in the following manner:


\code{years <- 1983:2002}\cr 
\code{season <- c(12,1,2)}\cr 
\code{lonLim <- c(-10, 4)}\cr 
\code{latLim <- c(35, 44)}\cr 


\code{loginUDG("username", "pasword")}\cr 
 
\code{EOBS_Iberia_tas <- loadGridData(dataset = "http://meteo.unican.es/tds5/dodsC/eobs/e-obs_v14_0.50regular.ncml", 
                                      var = "tg", 
                                      years = years, 
                                      lonLim = lonLim, 
                                      latLim = latLim, 
                                      season = season)}\cr
}
\examples{
data("EOBS_Iberia_tas")
plotClimatology(climatology(EOBS_Iberia_tas),
                backdrop.theme = "countries", scales = list(draw = TRUE))
}
\seealso{
\code{\link{makeMultiGrid}}, \code{\link[loadeR]{loadGridData}}
}
\keyword{Gridded}
\keyword{observations}
