% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEOF.R
\name{EOF2clim}
\alias{EOF2clim}
\title{PCA object EOFs to climatology grid}
\usage{
EOF2clim(prinCompObj, ind.var, member, n.eofs)
}
\arguments{
\item{prinCompObj}{PCA object}

\item{ind.var}{index position of the target variable}

\item{member}{index position of the member whose EOFs are to be displayed}

\item{n.eofs}{Number of EOFs to be displayed (from 1 to n.eofs)}
}
\value{
A climatological grid. Note that EOFs are treated as members
}
\description{
Converts a \code{prinComp} output to a climatology of EOFs. Internal of \code{plotEOF}.
}
\author{
J Bedia
}
\keyword{internal}
