% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillGrid.R
\name{fillGridDates}
\alias{fillGridDates}
\title{Fill missing dates}
\usage{
fillGridDates(grid, tz = "")
}
\arguments{
\item{grid}{grid or station data}

\item{tz}{Optional. Time zone. See Details.}
}
\value{
A grid filled with NAs in the previously missing date positions
}
\description{
fill with NA missing dates in grids and station datasets
}
\details{
The function attempts to recover the time zone of the input grid when this is correctly defined.
Otherwise, the function will leave it as unknown. See \code{\link{timezones}} for more details.
}
\seealso{
Other dateutils: 
\code{\link{getFirstChar}()},
\code{\link{isAlphaCharacter}()},
\code{\link{setGridDates.asPOSIXlt}()}
}
\author{
M Iturbide
}
\concept{dateutils}
\keyword{internal}
