% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterNA.R
\name{filterNA}
\alias{filterNA}
\title{Remove missing values from grids}
\usage{
filterNA(grid)
}
\arguments{
\item{grid}{The input grid to be filtered.}
}
\value{
A new grid object without missing values
}
\description{
Removes the dates for all grid points whenever there exists missing values such as NaN or NA.
}
\details{
The function also takes care of adjusting dates and other relevant metadata (via the internal \code{\link{subsetDimension}}).
}
\examples{
# Check if the dataset contains missing values (YES):
anyNA(VALUE_Iberia_pr$Data)
getShape(VALUE_Iberia_pr)
na.filtered <- filterNA(VALUE_Iberia_pr)
# Check if the dataset contains missing values (NO):
anyNA(na.filtered$Data)
getShape(na.filtered)
}
\author{
J. Bano-Medina
}
