% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatFiles-methods.R
\docType{methods}
\name{flatFiles}
\alias{flatFiles}
\alias{flatFiles,SummarizedExperiment-method}
\title{Coerce any S4 object to a list}
\usage{
\S4method{flatFiles}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{list}.
}
\description{
Coerce any S4 object to a list
}
\examples{
data(rse, package = "acidtest")

## SummarizedExperiment ====
x <- flatFiles(rse)
class(x)
names(x)
}
\seealso{
\code{\link[=coerceS4ToList]{coerceS4ToList()}}.
}
