% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getCoordinates}
\alias{getCoordinates}
\title{Get geographical coordinates of a climate data object}
\usage{
getCoordinates(obj)
}
\arguments{
\item{obj}{Any object extending the station or grid classes}
}
\value{
A list with x and y components
}
\description{
Returns the coordinates of a climate data object, either stations or grid
}
\examples{
# Regular grid
data("NCEP_Iberia_hus850")
str(getCoordinates(NCEP_Iberia_hus850))
# Regular rotated grid with correpondence with a non regular grid in the 
# lat/lon domain.
data("CORDEX_Iberia_tas")
str(getCoordinates(CORDEX_Iberia_tas))
# Station data
data("VALUE_Iberia_tas")
str(getCoordinates(VALUE_Iberia_tas))
}
\seealso{
Other get.helpers: \code{\link{getDim}},
  \code{\link{getGridVerticalLevels}},
  \code{\link{getGrid}}, \code{\link{getRefDates}},
  \code{\link{getSeason}}, \code{\link{getShape}},
  \code{\link{getTimeResolution}},
  \code{\link{getVarNames}}, \code{\link{getYearsAsINDEX}},
  \code{\link{typeofGrid}}, \code{\link{which.leap}}
}
\author{
J. Bedia
}
