% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getRefDates}
\alias{getRefDates}
\title{Retrieve reference dates}
\usage{
getRefDates(obj)
}
\arguments{
\item{obj}{A grid or station object}
}
\value{
A character vector of dates
}
\description{
Retrieves the \code{$start} component of the \code{$Dates} element of a grid or station.
 In case of multigrids, the dates from the first grid are returned (i.e.: \code{grid$Dates[[1]]$start}).
}
\details{
This utility function should be used always that reference dates are internally required by any function.
It takes into account the particular case when \code{\link{redim}} is applied to create a singleton \code{"var"}
 dimension: in this situation, the \code{"Dates"} component is not changed accordingly, and thus the criterion to
 distinguish multigrids from redim-ed grids is the shape of the array.
}
\examples{
#Station data:
data("VALUE_Iberia_tas")
str(getRefDates(VALUE_Iberia_tas))
#Regular grid
data("NCEP_Iberia_hus850")
str(getRefDates(NCEP_Iberia_hus850))
}
\seealso{
Other get.helpers: \code{\link{getCoordinates}},
  \code{\link{getDim}}, \code{\link{getGrid}},
  \code{\link{getSeason}}, \code{\link{getShape}},
  \code{\link{getYearsAsINDEX}}, \code{\link{which.leap}}
}
\author{
J. Bedia
}
\keyword{internal}
