% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getSeason}
\alias{getSeason}
\title{Get season}
\usage{
getSeason(obj)
}
\arguments{
\item{obj}{Any object extending the station or grid classes}
}
\value{
An integer vector with the season
}
\description{
Retrieves the season encompassed by a station or grid object
}
\examples{
data("NCEP_Iberia_hus850")
getSeason(NCEP_Iberia_hus850) # Boreal winter (DJF)
}
\seealso{
Other get.helpers: \code{\link{get2DmatCoordinates}},
  \code{\link{getCoordinates}}, \code{\link{getDim}},
  \code{\link{getGridVerticalLevels}},
  \code{\link{getGrid}}, \code{\link{getRefDates}},
  \code{\link{getShape}}, \code{\link{getTimeResolution}},
  \code{\link{getVarNames}}, \code{\link{getYearsAsINDEX}},
  \code{\link{typeofGrid}}, \code{\link{which.leap}}
}
\author{
J. Bedia
}
