% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getStationID}
\alias{getStationID}
\title{Obtain station names and IDs}
\usage{
getStationID(grid)
}
\arguments{
\item{grid}{An input station data grid}
}
\value{
A 2D matrix of name-ID (in this order)
}
\description{
Obtain station names and IDs
}
\seealso{
Other get.helpers: \code{\link{get2DmatCoordinates}},
  \code{\link{getCoordinates}}, \code{\link{getDim}},
  \code{\link{getGridVerticalLevels}},
  \code{\link{getGrid}}, \code{\link{getRefDates}},
  \code{\link{getSeason}}, \code{\link{getShape}},
  \code{\link{getTimeResolution}},
  \code{\link{getVarNames}}, \code{\link{getYearsAsINDEX}},
  \code{\link{typeofGrid}}, \code{\link{which.leap}}
}
\author{
M. Iturbide
}
