% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid2mopa.R
\name{grid2mopa0}
\alias{grid2mopa0}
\title{Transform a grid or multigrid to a set of predictors to be used for modeling}
\usage{
grid2mopa0(grid, clim.fun = list(FUN = "mean", na.rm = TRUE),
  varname = "variable", crs = NA)
}
\arguments{
\item{grid}{Input grid or multigrid without member or runtime dimensions. Grid returned by function 
\code{\link[transformeR]{climatology}} are also supported.}

\item{clim.fun}{Function to compute the climatology (i.e. function for temporal aggregation). 
This is specified as a list, indicating the name 
of the aggregation function in first place (as character), and other optional arguments to be passed 
to the aggregation function. Default to mean (i.e., clim.fun = list(FUN="mean",na.rm = TRUE)).}

\item{varname}{Variable name of the grid}

\item{crs}{Optional (Default NA). Character or object of class CRS. PROJ.4 type description of a Coordinate Reference System (map projection). 
If this argument is missing, and the x coordinates are withing -360 .. 360 and the y coordinates are 
within -90 .. 90, "+proj=longlat +datum=WGS84" is used. Also see under Details if x is a character (filename).}
}
\value{
A raster for grid objects and a RasterStack for multigrids.
}
\description{
Transform a grid or multigrid to collections of raster files for modeling
with the \code{mopa} package.
}
\examples{
\dontrun{
#' # Maximum July surface temp forecast climatology
data("tasmax_forecast")
# Aggregate all members and compute climatology
tx_mean.clim <- climatology(tasmax_forecast, by.member = FALSE)
plotClimatology(tx_mean.clim)
tx_ras <- grid2mopa(tx_mean.clim)
spplot(tx_ras)
}
}
\seealso{
\code{\link[transformeR]{climatology}}; \code{\link[loadeR]{loadGridData}}
}
\author{
M. Iturbide
}
