% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridArithmetics.R
\name{gridArithmetics}
\alias{gridArithmetics}
\title{Basic arithmetic operations between grids and numbers}
\usage{
gridArithmetics(..., operator = "*", template = NULL)
}
\arguments{
\item{...}{Input grids or factors (numbers).}

\item{operator}{Character string of arithmetic simbols ("+", "-", "/", "*").  Variable name of the grid}

\item{template}{grid from which the grid structure is keeped for the output data (Default is the first
grid in "...").}
}
\value{
A grid
}
\description{
Perform basic arithmetic operations between two or more grids (or factors).
}
\examples{
#' # Maximum July surface temp forecast climatology
data("CMIP5_Iberia_pr")
data("EOBS_Iberia_pr")
cmip5 <- interpGrid(CMIP5_Iberia_pr, getGrid(EOBS_Iberia_pr))
relative.delta <- gridArithmetics(cmip5, EOBS_Iberia_pr, 
                                  EOBS_Iberia_pr, 
                                  100, 
                                  operator = c("-", "/", "*"), 
                                  template = NULL)
}
\author{
M. Iturbide
}
