% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_wholenumber.R
\name{is_wholenumber}
\alias{is_wholenumber}
\title{Check whole number}
\usage{
is_wholenumber(x)
}
\arguments{
\item{x}{Numeric vector of values to be tested.}
}
\value{
Logical vector. Returns TRUE or FALSE depending on whether its argument is whole number or not.
}
\description{
Check whole number
}
\examples{
is_wholenumber(1)
}
