% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isRegular}
\alias{isRegular}
\title{Check if the imput grid is regular}
\usage{
isRegular(grid)
}
\arguments{
\item{grid}{a grid or station data, or the objectreturned by function \code{\link{getGrid}}}
}
\value{
Logical.
}
\description{
Check if the coordinates in the data are regular or irregular
}
\examples{
data("EOBS_Iberia_tas")
isRegular(EOBS_Iberia_tas)
isRegular(getGrid(EOBS_Iberia_tas))
data("VALUE_Iberia_tas")
isRegular(VALUE_Iberia_tas)
}
\seealso{
\code{\link{getGrid}}
}
\author{
M Iturbide
}
\keyword{internal}
