% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join-S3methods.R
\name{join}
\alias{join}
\alias{left_join}
\alias{left_join.DataFrame}
\title{Join operations supporting S4 class objects}
\usage{
left_join(x, y, by)

\method{left_join}{DataFrame}(x, y, by)
}
\arguments{
\item{x, y}{\code{DataFrame}.
Data frames, or objects, to be coerced into a single object.}

\item{by}{\code{character}.
Column names to use for merge operation.
Note that this is always required, unlike the dplyr \code{tbl_df} method, where
this is optional.}
}
\description{
Join operations supporting S4 class objects
}
\examples{
DataFrame <- S4Vectors::DataFrame

## DataFrame ====
df1 <- DataFrame(
    id = as.factor(seq(4L)),
    genotype = as.factor(rep(x = c("wt", "ko"), each = 2L))
)
print(df1)
df2 <- DataFrame(
    id = as.factor(seq(4L)),
    treatment = as.factor(rep(x = c("control", "expt"), times = 2L))
)
print(df2)
left_join(df1, df2, by = "id")
}
\seealso{
\itemize{
\item \code{help(topic = "join", package = "dplyr")}.
\item \code{help(topic = "merge", package = "S4Vectors")}.
}
}
