% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-join-S4methods.R
\docType{methods}
\name{join}
\alias{join}
\alias{inner_join}
\alias{left_join}
\alias{right_join}
\alias{full_join}
\alias{semi_join}
\alias{nest_join}
\alias{anti_join}
\alias{inner_join,data.frame-method}
\alias{left_join,data.frame-method}
\alias{right_join,data.frame-method}
\alias{full_join,data.frame-method}
\alias{semi_join,data.frame-method}
\alias{nest_join,data.frame-method}
\alias{anti_join,data.frame-method}
\alias{left_join,DataFrame-method}
\alias{right_join,DataFrame-method}
\title{Join operations supporting S4 class objects}
\usage{
inner_join(x, y, ...)

left_join(x, y, ...)

right_join(x, y, ...)

full_join(x, y, ...)

semi_join(x, y, ...)

nest_join(x, y, ...)

anti_join(x, y, ...)

\S4method{inner_join}{data.frame}(x, y, ...)

\S4method{left_join}{data.frame}(x, y, ...)

\S4method{right_join}{data.frame}(x, y, ...)

\S4method{full_join}{data.frame}(x, y, ...)

\S4method{semi_join}{data.frame}(x, y, ...)

\S4method{nest_join}{data.frame}(x, y, ...)

\S4method{anti_join}{data.frame}(x, y, ...)

\S4method{left_join}{DataFrame}(x, y, by)

\S4method{right_join}{DataFrame}(x, y, by)
}
\arguments{
\item{x, y}{\code{DataFrame}.
Data frames, or objects, to be coerced into a single object.}

\item{by}{\code{character}.
Column names to use for merge operation.
Note that this is always required, unlike the dplyr \code{tbl_df} method, where
this is optional.}

\item{copy, suffix, ...}{Disabled for \code{DataFrame} method.}
}
\description{
Join operations supporting S4 class objects
}
\note{
Updated 2019-08-15.
}
\examples{
data(band_members, band_instruments)
left_join(band_members, band_instruments, by = "name")
}
\seealso{
\itemize{
\item \code{help(topic = "join", package = "dplyr")}.
\item \code{help(topic = "merge", package = "S4Vectors")}.
}
}
