% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/join-methods.R
\docType{methods}
\name{innerJoin}
\alias{innerJoin}
\alias{leftJoin}
\alias{rightJoin}
\alias{fullJoin}
\alias{semiJoin}
\alias{antiJoin}
\alias{join}
\alias{innerJoin,DataFrame,DataFrame-method}
\alias{leftJoin,DataFrame,DataFrame-method}
\alias{rightJoin,DataFrame,DataFrame-method}
\alias{fullJoin,DataFrame,DataFrame-method}
\alias{semiJoin,DataFrame,DataFrame-method}
\alias{antiJoin,DataFrame,DataFrame-method}
\title{Join operations supporting S4 class objects}
\usage{
innerJoin(x, y, ...)

leftJoin(x, y, ...)

rightJoin(x, y, ...)

fullJoin(x, y, ...)

semiJoin(x, y, ...)

antiJoin(x, y, ...)

\S4method{innerJoin}{DataFrame,DataFrame}(x, y, by)

\S4method{leftJoin}{DataFrame,DataFrame}(x, y, by)

\S4method{rightJoin}{DataFrame,DataFrame}(x, y, by)

\S4method{fullJoin}{DataFrame,DataFrame}(x, y, by)

\S4method{semiJoin}{DataFrame,DataFrame}(x, y, by)

\S4method{antiJoin}{DataFrame,DataFrame}(x, y, by)
}
\arguments{
\item{x, y}{\code{DataFrame}.
Data frames, or objects, to be coerced into a single object.}

\item{...}{Additional arguments.}

\item{by}{\code{character}.
Column names to use for merge operation.
Note that this is always required, unlike the dplyr \code{tbl_df} method, where
this is optional.}
}
\value{
Merge of \code{x} and \code{y} objects.
}
\description{
Join operations supporting S4 class objects
}
\note{
Updated 2019-08-15.
}
\section{Ordering}{


The join functions never rearrange rows. To accomplish this, we're currently
setting an internal \code{.idx} column that we can use to reorder the rows after
\code{\link[base:merge]{merge()}} operation.
}

\section{Row names}{


Unlike the S3 methods defined in dplyr, the join methods defined here for
\code{DataFrame} always preserve row names.
}

\section{\code{data.frame} methods}{


Since we are defining S4 methods in this package, we are providing
passthrough support to dplyr for \code{data.frame} class objects.

Refer to \code{help(topic = "join", package = "dplyr")} for details.
}

\examples{
data(band_members, band_instruments)

## DataFrame ====
x <- band_members
print(x)
y <- band_instruments
print(y)
by <- "name"
innerJoin(x = x, y = y, by = by)
leftJoin(x = x, y = y, by = by)
rightJoin(x = x, y = y, by = by)
fullJoin(x = x, y = y, by = by)
semiJoin(x = x, y = y, by = by)
antiJoin(x = x, y = y, by = by)
}
\seealso{
\itemize{
\item \code{help(topic = "join", package = "dplyr")}.
\item \code{help(topic = "merge", package = "S4Vectors")}.
\item https://support.bioconductor.org/p/120277/
}
}
