% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join}
\alias{join}
\alias{inner_join}
\alias{left_join}
\alias{left_join.DataFrame}
\title{Join operations supporting S4 class objects}
\usage{
\method{left_join}{DataFrame}(x, y, by)
}
\arguments{
\item{x, y}{\code{DataFrame}.
Data frames, or objects, to be coerced into a single object.}

\item{by}{\code{character}.
Column names to use for merge operation.
Note that this is always required, unlike the dplyr \code{tbl_df} method, where
this is optional.}
}
\description{
Join operations supporting S4 class objects
}
\seealso{
\itemize{
\item \code{dplyr::left_join()}.
\item \code{S4Vectors::merge()}.
}
}
