% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dplyr-join-S4methods.R
\docType{methods}
\name{inner_join}
\alias{inner_join}
\alias{left_join}
\alias{right_join}
\alias{full_join}
\alias{semi_join}
\alias{anti_join}
\alias{join}
\alias{inner_join,data.frame,data.frame-method}
\alias{left_join,data.frame,data.frame-method}
\alias{right_join,data.frame,data.frame-method}
\alias{full_join,data.frame,data.frame-method}
\alias{semi_join,data.frame,data.frame-method}
\alias{anti_join,data.frame,data.frame-method}
\alias{inner_join,DataFrame,DataFrame-method}
\alias{left_join,DataFrame,DataFrame-method}
\alias{right_join,DataFrame,DataFrame-method}
\alias{full_join,DataFrame,DataFrame-method}
\alias{semi_join,DataFrame,DataFrame-method}
\alias{anti_join,DataFrame,DataFrame-method}
\title{Join operations supporting S4 class objects}
\usage{
inner_join(x, y, ...)

left_join(x, y, ...)

right_join(x, y, ...)

full_join(x, y, ...)

semi_join(x, y, ...)

anti_join(x, y, ...)

\S4method{inner_join}{data.frame,data.frame}(x, y, ...)

\S4method{left_join}{data.frame,data.frame}(x, y, ...)

\S4method{right_join}{data.frame,data.frame}(x, y, ...)

\S4method{full_join}{data.frame,data.frame}(x, y, ...)

\S4method{semi_join}{data.frame,data.frame}(x, y, ...)

\S4method{anti_join}{data.frame,data.frame}(x, y, ...)

\S4method{inner_join}{DataFrame,DataFrame}(x, y, by, rownames = TRUE)

\S4method{left_join}{DataFrame,DataFrame}(x, y, by, rownames = TRUE)

\S4method{right_join}{DataFrame,DataFrame}(x, y, by, rownames = TRUE)

\S4method{full_join}{DataFrame,DataFrame}(x, y, by, rownames = TRUE)

\S4method{semi_join}{DataFrame,DataFrame}(x, y, by, rownames = TRUE)

\S4method{anti_join}{DataFrame,DataFrame}(x, y, by, rownames = TRUE)
}
\arguments{
\item{x, y}{\code{DataFrame}.
Data frames, or objects, to be coerced into a single object.}

\item{by}{\code{character}.
Column names to use for merge operation.
Note that this is always required, unlike the dplyr \code{tbl_df} method, where
this is optional.}

\item{copy, suffix, ...}{Disabled for \code{DataFrame} method.}
}
\description{
Join operations supporting S4 class objects
}
\note{
Updated 2019-08-15.
}
\section{Ordering}{


The join functions never rearrange rows. To accomplish this, we're currently
setting an internal \code{.idx} column that we can use to reorder the rows after
\code{\link[base:merge]{merge()}} operation.
}

\section{Row names}{


Unlike the S3 methods defined in dplyr, the join methods defined here for
\code{DataFrame} always preserve row names.
}

\examples{
data(band_members, band_instruments)
x <- band_members
print(x)
y <- band_instruments
print(y)
by <- "name"

inner_join(x = x, y = y, by = by)
left_join(x = x, y = y, by = by)
right_join(x = x, y = y, by = by)
full_join(x = x, y = y, by = by)
semi_join(x = x, y = y, by = by)
anti_join(x = x, y = y, by = by)
}
\seealso{
\itemize{
\item \code{help(topic = "join", package = "dplyr")}.
\item \code{help(topic = "merge", package = "S4Vectors")}.
\item https://support.bioconductor.org/p/120277/
}
}
