% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambWT.R
\name{lambWT}
\alias{lambWT}
\title{Calculation of Lamb Weather types (WT).}
\usage{
lambWT(grid, center.point = c(-5, 55), typeU = FALSE)
}
\arguments{
\item{grid}{A grid (gridded or station dataset), or multimember grid object of MSLP values.}

\item{center.point}{A two value vector that must include lon and lat from a location that will work as center point for the Lamb WT.
See details.}

\item{typeU}{Logical. Should "Unclassified" type be included in the output. Default to \code{FALSE}, so the closest is assigned.}
}
\value{
The Lamb WT circulation index (and members, if applicable) with:
\itemize{
\item index: vector with the corresponding weather type from each point of the series, that is defined as follows:

\tabular{cccc}{
\tab purely anticyclonic:  \tab 1 \tab "A" \cr 
\tab anticyclonic northeasterly:  \tab  2  \tab "ANE" \cr 
\tab anticyclonic easterly: \tab 3 \tab "AE" \cr
\tab anticyclonic southeasterly: \tab 4 \tab "ASE" \cr
\tab anticyclonic southerly: \tab 5 \tab "AS" \cr
\tab anticyclonic southwesterly: \tab 6 \tab "ASW" \cr
\tab anticyclonic westerly: \tab 7 \tab "AW" \cr
\tab anticyclonic northwesterly: \tab 8 \tab "ANW" \cr
\tab anticyclonic northerly: \tab 9 \tab "AN" \cr
\tab northeasterly: \tab 10 \tab "NE" \cr
\tab easterly: \tab 11 \tab "E" \cr
\tab southeasterly:\tab 12 \tab "SE" \cr
\tab southerly: \tab 13 \tab "S" \cr
\tab southwesterly: \tab 14 \tab "SW" \cr
\tab westerly: \tab 15 \tab "W" \cr
\tab northwesterly: \tab 16 \tab "NW" \cr
\tab northerly: \tab 17 \tab "N" \cr
\tab purely cyclonic:  \tab 18 \tab "C" \cr 
\tab cyclonic northeasterly:  \tab  19  \tab "CNE" \cr 
\tab cyclonic easterly: \tab 20 \tab "CE" \cr
\tab cyclonic southeasterly: \tab 21 \tab "CSE" \cr
\tab cyclonic southerly: \tab 22 \tab "CS" \cr
\tab cyclonic southwesterly: \tab 23 \tab "CSW" \cr
\tab cyclonic westerly: \tab 24 \tab "CW" \cr
\tab cyclonic northwesterly: \tab 25 \tab "CNW" \cr
\tab cyclonic northerly: \tab 26 \tab "CN" \cr 
}

\item pattern: Array with the spatial pattern of the 26 weather types obtained.
\item dates and coordinates.
\item further arguments related to the Lamb WT index.
}
}
\description{
Calculate automated Lamb WT as defined in Trigo and daCamara (2000), Int J Climatol
}
\details{
According to Trigo and daCamara (2000), Int J Climatol, Lamb WT is only applied on North Atlantic domain. 
The input grid units must be Pa, not hPa/mbar. If it is not in Pa, the units must be converted.
A center location point must be specified by the user. Then, the function calculates from left to right and from first to 16st 
the rest of the location point from the grid specified by Trigo and daCamara (2000):
 
  \tabular{ccccccccccccc}{
    \tab  \tab  \tab    \tab  \tab  \tab 01 \tab  \tab  \tab 02 \tab  \tab  \tab    \cr
    \tab  \tab  \tab 03 \tab  \tab  \tab 04 \tab  \tab  \tab 05 \tab  \tab  \tab 06 \cr
    \tab  \tab  \tab 07 \tab  \tab  \tab 08 \tab  \tab  \tab 09 \tab  \tab  \tab 10 \cr
    \tab  \tab  \tab 11 \tab  \tab  \tab 12 \tab  \tab  \tab 13 \tab  \tab  \tab 14 \cr
    \tab  \tab  \tab    \tab  \tab  \tab 15 \tab  \tab  \tab 16 \tab  \tab  \tab    
}  

where the north-south distance is 5º and the west-east distance is 10º. 26 different WTs are defined, 10 pure types (NE, E, SE, S, SW, 
W, NW, N, C and A) and 16 hybrid types (8 for each C and A hybrid).
}
\examples{
\dontrun{
require(climate4R.indices)
data(NCEP_slp_2001_2010)
lamb.wt <- lambWT(grid = NCEP_slp_2001_2010)
}
}
