% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeGrid.R
\name{mergeGrid}
\alias{mergeGrid}
\title{Flexible grid aggregation along selected dimensions}
\usage{
mergeGrid(..., aggr.fun = list(FUN = "mean", na.rm = TRUE))
}
\arguments{
\item{aggr.fun}{Aggregation function to the intersection areas among grids. The default option computes the mean
aggr.fun = list(FUN = "mean", na.rm = TRUE).}

\item{grid}{a grid or multigrid to be aggregated.}

\item{parallel}{Logical. Should parallel execution be used?}

\item{max.ncores}{Integer. Upper bound for user-defined number of cores.}

\item{ncores}{Integer number of cores used in parallel computation. Self-selected number of
cores is used when \code{ncpus = NULL} (the default), or when \code{maxcores} exceeds the default \code{ncores} value.}
}
\value{
A single grid result of merging the input grids.
}
\description{
Merge multiple grids in space (latitude and longitude).
}
\details{
\strong{Aggregation function definition}

The aggregation functions are specified in the form of a named list of the type \code{FUN = "function", ...}, where
\code{...} are further arguments passes to FUN. This allows for a flexible definition of aggregation functions, that are 
internally passes to \code{\link{tapply}}. Note that the name of the function is indicated as a character string.
}
\author{
J. Baño-Medina, M. Iturbide
}
