% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dplyr-mutate-S4methods.R
\docType{methods}
\name{mutate_all}
\alias{mutate_all}
\alias{mutate_at}
\alias{mutate_if}
\alias{mutate}
\alias{mutate_all,data.frame-method}
\alias{mutate_all,DataFrame-method}
\alias{mutate_at,data.frame-method}
\alias{mutate_at,DataFrame-method}
\alias{mutate_if,data.frame-method}
\alias{mutate_if,DataFrame-method}
\title{Mutate multiple columns}
\usage{
mutate_all(.tbl, ...)

mutate_at(.tbl, ...)

mutate_if(.tbl, ...)

\S4method{mutate_all}{data.frame}(.tbl, .funs, ...)

\S4method{mutate_all}{DataFrame}(.tbl, .funs, ...)

\S4method{mutate_at}{data.frame}(.tbl, .vars, .funs, ...)

\S4method{mutate_at}{DataFrame}(.tbl, .vars, .funs, ...)

\S4method{mutate_if}{data.frame}(.tbl, .predicate, .funs, ...)

\S4method{mutate_if}{DataFrame}(.tbl, .predicate, .funs, ...)
}
\value{
Modified object.
}
\description{
The \link{scoped} variants of \code{\link[=mutate]{mutate()}} and \code{\link[=transmute]{transmute()}} make it easy to apply
the same transformation to multiple variables. There are three variants:
\itemize{
\item _all affects every variable
\item _at affects variables selected with a character vector or vars()
\item _if affects variables selected with a predicate function:
}
}
\note{
Updated 2019-08-15.
}
\section{\code{data.frame} methods}{


Since we are defining S4 methods in this package, we are providing
passthrough support to dplyr for \code{data.frame} class objects. Our generic
methods pass through to dplyr mutate functions, which are optimized for
\code{tbl_df} class.

Refer to \code{help(topic = "mutate_all", package = "dplyr")} for details.
}

\examples{
data(mtcars, package = "datasets")

## DataFrame ====
x <- as(mtcars, "DataFrame")
print(x)
x \%>\% mutate_all(log, base = 2L)
x \%>\% mutate_if(is.double, as.integer)
x \%>\% mutate_at(c("mpg", "cyl"), log, base = 2L)
}
