% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overGrid.R
\name{overGrid}
\alias{overGrid}
\title{Spatial overlay for grids and Spatial* objects}
\usage{
overGrid(grid, layer, subset = FALSE)
}
\arguments{
\item{grid}{Input grid. The first element must be a grid, the rest could
be either grids or numbers.}

\item{layer}{Spatial* object from which the geometries or attributes are 
queried (type \code{help(package = "sp")}.}

\item{subset}{Logical (default is FALSE). If TRUE, spatial subsetting is performed over
the imput grid using the bounding coordinates of argument 'layer'.}
}
\value{
A grid
}
\description{
Application of function \code{over} from \pkg{sp} over grids.
}
\examples{
library(visualizeR)
data("PRUDENCEregions")
library(loadeR)
grid <- loadGridData(
      "http://opendap.knmi.nl/knmi/thredds/dodsC/e-obs_0.25regular/tx_0.25deg_reg_v16.0.nc",
      var = "tx", years = 1983, season = 12)

iberia <- overGrid(grid, PRUDENCEregions[2, ])
iberia2 <- overGrid(grid, PRUDENCEregions[2,], subset = TRUE)
allregions <- overGrid(grid, PRUDENCEregions, subset = TRUE)

spatialPlot(climatology(grid),
            sp.layout = list(list(PRUDENCEregions, first = FALSE)))
spatialPlot(climatology(iberia),
            sp.layout = list(list(PRUDENCEregions, first = FALSE)))
spatialPlot(climatology(iberia2),
            sp.layout = list(list(PRUDENCEregions, first = FALSE)))
spatialPlot(climatology(allregions),
            sp.layout = list(list(PRUDENCEregions, first = FALSE)))
}
\author{
M. Iturbide
}
