% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClimatology.R
\name{plotClimatology}
\alias{plotClimatology}
\title{Lattice plot methods for climatological grids}
\usage{
plotClimatology(grid, backdrop.theme = "none", set.min = NULL,
  set.max = NULL, lonCenter = NULL, ...)
}
\arguments{
\item{grid}{Input grid}

\item{backdrop.theme}{Reference geographical lines to be added to the plot. See Details.}

\item{set.min}{Numeric value indicating an absolute minimum value (default to \code{NULL}). All grid values below this are mapped to \code{set.min}. See details.}

\item{set.max}{Same as \code{set.min} argument, but to force a ceiling.}

\item{lonCenter}{Value of the longitude to be centered in the plot.}

\item{...}{Further arguments passed to \code{spplot}}
}
\value{
As spplot, \code{plotClimatology} returns a lattice plot of class \dQuote{trellis}. 
 If you fail to \dQuote{see} it, explicitly call \code{print(plotClimatology(...))}.
}
\description{
A wrapper for the lattice (trellis) plot methods for spatial data in \code{sp::spplot}
}
\details{
The function applies the \code{\link[sp]{spplot}} method after conversion of the climatological map(s) to a
 \code{SpatialGridDataFrame}.
 
 The \code{set.min} and \code{set.max} options are useful in order to preserve adequate ranges for map representation,
  avoiding the influence of extreme values. Note that this is different than setting a range of values with an 
  interval using the \code{at} argument. The latter choice, that overrides \code{set.min} and \code{set.max},
   leaves blank grid points for outlying values.
 
 \strong{Multigrids}
 
 Multigrids of climatologies can be created using \code{makeMultiGrid} 
 for trellis visualization of different variables, or for instance, for the comparison of
 raw and corrected/downscaled scenarios side to side. In case of multimember multigrids, 
 the function will internally compute the ensemble mean of each variable in the multigrid
  for representation (with a message).
 
 \strong{Backdrop theme}
 
 Current implemented options are \code{"none"} and \code{"coastline"}, which contains
 a simplied vector theme delineating the world coastlines. Any other themes can be introduced
 by the user using the \code{sp.layout} options in \code{spplot}.
 
 \strong{Controlling graphical parameters}
 
 Many different aspects of the map can be controlled passing the relevant arguments to 
 \code{spplot}. Fine control of graphical parameters for the trellis display can
 be also controlled using \code{\link[lattice]{trellis.par.set}}.
}
\examples{
data("CFS_Iberia_tas")
}
\references{
\itemize{
\item Bivand, R.S., Pebesma, E.J., Gomez-Rubio, V., 2013. Applied Spatial Data Analysis with R, 2nd ed, useR! Springer, NY.
\item For some graticulate customization examples, visit the \emph{sp Gallery}: \url{https://edzer.github.io/sp/}
}
}
\seealso{
\code{\link{climatology}} for details on climatology calculation.
 \code{\link{map.stippling}}, for adding a custom point layer on top of the map.
 \code{\link{map.lines}}, to add lines and polygons to climatological maps
Also see \code{\link[sp]{spplot}} in package \pkg{sp} for further information on plotting capabilities and options
}
\author{
J. Bedia
}
