% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redim.R
\name{redim}
\alias{redim}
\title{Complete missing dimensions of Grid objects}
\usage{
redim(
  grid,
  member = TRUE,
  runtime = FALSE,
  var = FALSE,
  time = TRUE,
  loc = FALSE,
  drop = FALSE
)
}
\arguments{
\item{member}{logical. Add 'member' dimension (default = TRUE)}

\item{runtime}{logical. Add 'runtime' dimension (default = FALSE)}

\item{var}{logical. Add 'var' dimension (default = FALSE)}

\item{loc}{logical. Only if \code{obj} is a stations dataset. If TRUE, the \code{"loc"} 
dimension is not replaced by fake \code{"lat"} and \code{"lon"} dimensions (default is FALSE).}

\item{drop}{logical. Drop dimensions of length = 1 (default = FALSE)}

\item{obj}{A regular or irregular (e.g. station data) grid}
}
\value{
The same object with all the dimensions (i.e. member, time, loc)
}
\description{
Add missing dimensions to the 'Data' array as singletons. For internal usage mainly.
}
\details{
The function does not handle multigrids (i.e. \code{"var"} dimension).
 Thus, multigrids need to be subsetted along \code{"var"} prior to redimensioning via \code{\link{subsetGrid}}.
}
\examples{
\donttest{
require(climate4R.datasets)
data("EOBS_Iberia_tas")
getShape(EOBS_Iberia_tas)
a <- redim(EOBS_Iberia_tas)
getShape(a)
b <- redim(EOBS_Iberia_tas, var = TRUE)
getShape(b)
# This one is probably never needed, but for illustration:
y <- redim(EOBS_Iberia_tas, member = FALSE, var = TRUE)
getShape(y)
# 'drop = TRUE' performs the reverse operation:
z <- redim(b, drop = TRUE)
getShape(z)

# Example with station data
data("VALUE_Iberia_pr")
getShape(VALUE_Iberia_pr)
# By default, the internal data array is converted to the time.lat-lon structure.
# This is a conveninet behaviour for internal usage:
getShape(redim(VALUE_Iberia_pr, member = TRUE))
# Use the argument 'loc=TRUE' to avoid this behaviour:
getShape(redim(VALUE_Iberia_pr, member = TRUE, loc = TRUE))
}
}
\seealso{
Other internal.helpers: 
\code{\link{array3Dto2Dmat.stations}()},
\code{\link{array3Dto2Dmat}()},
\code{\link{bindGrid.member}()},
\code{\link{bindGrid.spatial}()},
\code{\link{bindGrid.time}()},
\code{\link{grid3Dto2Dmatrix}()},
\code{\link{mat2Dto3Darray.stations}()},
\code{\link{sortDim.spatial}()},
\code{\link{sortDim.time}()}
}
\author{
M. Iturbide, J. Bedia
}
\concept{internal.helpers}
\keyword{internal}
