% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-tbl_df.R, R/flatFiles-methods.R
\docType{import}
\name{reexports}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bioverbs}{\code{\link[bioverbs]{flatFiles}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

