% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame-s4.R, R/flatFiles-methods.R,
%   R/tbl_df-s3.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tibble}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{BiocGenerics}{\code{\link[BiocGenerics]{as.data.frame}}}

  \item{bioverbs}{\code{\link[bioverbs]{flatFiles}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

