% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tbl_df.R, R/atomize-methods.R,
%   R/decode-methods.R, R/encode-methods.R, R/factorize-methods.R,
%   R/flatFiles-methods.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{DataFrame}
\alias{colData}
\alias{rowRanges}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{bioverbs}{\code{\link[bioverbs]{atomize}}, \code{\link[bioverbs]{encode}}, \code{\link[bioverbs]{factorize}}, \code{\link[bioverbs]{flatFiles}}}

  \item{S4Vectors}{\code{\link[S4Vectors]{decode}}, \code{\link[S4Vectors]{DataFrame}}}

  \item{SummarizedExperiment}{\code{\link[SummarizedExperiment]{colData}}, \code{\link[SummarizedExperiment]{rowRanges}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}
}}

