% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel.R
\name{relevel}
\alias{relevel}
\alias{relevel.DataFrame}
\alias{relevel.GRanges}
\title{Reorder factor levels}
\usage{
relevel(x, ref, ...)

\method{relevel}{DataFrame}(x, ref = NULL, ...)

\method{relevel}{GRanges}(x, ref = NULL, ...)
}
\arguments{
\item{x}{Object.}

\item{ref}{\code{NULL}.
Unsupported argument that is part of S3 generic.}

\item{...}{Additional arguments}
}
\value{
Modified object.
Factor levels will be readjusted (i.e. superfluous levels are dropped).
}
\description{
Reorder factor levels
}
\examples{
data(rse, package = "acidtest")

## DataFrame
df <- SummarizedExperiment::rowData(rse)
x <- relevel(df)
summary(x)

## GRanges
gr <- SummarizedExperiment::rowRanges(rse)
x <- relevel(gr)
summary(x)
}
