% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel-S3methods.R
\name{relevel}
\alias{relevel}
\alias{relevel.DataFrame}
\alias{relevel.Ranges}
\alias{relevel.SummarizedExperiment}
\title{Reorder factor levels}
\usage{
relevel(x, ref, ...)

\method{relevel}{DataFrame}(x, ref = NULL, ...)

\method{relevel}{Ranges}(x, ref = NULL, ...)

\method{relevel}{SummarizedExperiment}(x, ref = NULL, ...)
}
\arguments{
\item{x}{Object.}

\item{ref}{\code{NULL}.
Unsupported argument that is part of S3 generic.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Factor levels will be readjusted (i.e. superfluous levels are dropped).
}
\description{
Reorder factor levels
}
\examples{
library(SummarizedExperiment)
data(RangedSummarizedExperiment, package = "acidtest")

## DataFrame ====
df <- rowData(RangedSummarizedExperiment)
x <- relevel(df)
summary(x)

## GenomicRanges ====
gr <- rowRanges(RangedSummarizedExperiment)
x <- relevel(gr)
summary(x)
}
