% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/select-methods.R
\docType{methods}
\name{selectIf}
\alias{selectIf}
\alias{select}
\alias{selectIf,DataFrame-method}
\title{Select multiple columns}
\usage{
selectIf(object, ...)

\S4method{selectIf}{DataFrame}(object, .predicate)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}

\item{.predicate}{\code{function} or \code{logical}.
A predicate function to be applied to the columns, or a logical vector
matching the number of columns.}
}
\value{
Modified object.
}
\description{
Select multiple columns
}
\note{
Updated 2019-08-24.
}
\examples{
data(iris, package = "datasets")

## DataFrame ====
x <- as(iris, "DataFrame")
selectIf(x, .predicate = is.factor)
}
\seealso{
These functions are inspired by dplyr. However, they are designed to only
work on \code{DataFrame} class, and use base R code internally.\preformatted{`help(topic = "select_all", package = "dplyr")`
}
}
