% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dplyr-select-S4methods.R
\docType{methods}
\name{select_all}
\alias{select_all}
\alias{select_at}
\alias{select_if}
\alias{select}
\alias{select_all,data.frame-method}
\alias{select_all,DataFrame-method}
\alias{select_at,data.frame-method}
\alias{select_at,DataFrame-method}
\alias{select_if,data.frame-method}
\alias{select_if,DataFrame-method}
\title{Select and rename a selection of variables}
\usage{
select_all(.tbl, ...)

select_at(.tbl, ...)

select_if(.tbl, ...)

\S4method{select_all}{data.frame}(.tbl, .funs = list(), ...)

\S4method{select_all}{DataFrame}(.tbl, .funs = list(), ...)

\S4method{select_at}{data.frame}(.tbl, .vars, .funs = list(), ...)

\S4method{select_at}{DataFrame}(.tbl, .vars, .funs = list(), ...)

\S4method{select_if}{data.frame}(.tbl, .predicate, .funs = list(), ...)

\S4method{select_if}{DataFrame}(.tbl, .predicate, .funs = list(), ...)
}
\arguments{
\item{.tbl}{Object.}

\item{...}{Additional argument.}

\item{.funs}{\code{function}.
Refer to dplyr documentation for details.}

\item{.vars}{\code{character}.
Column names.}

\item{.predicate}{\code{function}.
Refer to dplyr documentation for details.}
}
\value{
Modified object.
}
\description{
Select and rename a selection of variables
}
\note{
Updated 2019-08-15.
}
\section{\code{data.frame} methods}{


Since we are defining S4 methods in this package, we are providing
passthrough support to dplyr for \code{data.frame} class objects. Our generic
methods pass through to dplyr select functions, which are optimized for
\code{tbl_df} class.

Refer to \code{help(topic = "select_all", package = "dplyr")} for details.
}

\examples{
data(mtcars, package = "datasets")

## DataFrame ====
x <- as(mtcars, "DataFrame")
select_all(x, .funs = toupper)
select_at(x, .vars = c("mpg", "cyl"))
select_if(x, .predicate = is.double)
}
