% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/select-methods.R
\docType{methods}
\name{selectAll}
\alias{selectAll}
\alias{selectAt}
\alias{selectIf}
\alias{select}
\alias{selectAll,DataFrame-method}
\alias{selectAt,DataFrame-method}
\alias{selectIf,DataFrame-method}
\title{Select and rename a selection of variables}
\usage{
selectAll(object, ...)

selectAt(object, ...)

selectIf(object, ...)

\S4method{selectAll}{DataFrame}(object, funs = list(), ...)

\S4method{selectAt}{DataFrame}(object, .vars, funs = list(), ...)

\S4method{selectIf}{DataFrame}(object, predicate, funs = list(), ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional argument.}

\item{funs}{\code{function}.
Refer to dplyr documentation for details.}

\item{predicate}{\code{function}.
Refer to dplyr documentation for details.}
}
\value{
Modified object.
}
\description{
Select and rename a selection of variables
}
\note{
Updated 2019-08-15.
}
\section{\code{data.frame} methods}{


Since we are defining S4 methods in this package, we are providing
passthrough support to dplyr for \code{data.frame} class objects. Our generic
methods pass through to dplyr select functions, which are optimized for
\code{tbl_df} class.

Refer to \code{help(topic = "select_all", package = "dplyr")} for details.
}

\examples{
data(mtcars, package = "datasets")

## DataFrame ====
x <- as(mtcars, "DataFrame")
selectAll(x, funs = toupper)
selectAt(x, .vars = c("mpg", "cyl"))
selectIf(x, predicate = is.double)
}
