% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-methods.R
\docType{methods}
\name{select}
\alias{select}
\alias{selectIf,DataFrame-method}
\title{Select multiple columns}
\usage{
\S4method{selectIf}{DataFrame}(object, .predicate)
}
\arguments{
\item{object}{Object.}
}
\value{
Modified object.
}
\description{
Select multiple columns
}
\note{
Updated 2019-08-26.
}
\examples{
data(iris, package = "datasets")

## DataFrame ====
x <- as(iris, "DataFrame")
selectIf(x, .predicate = is.factor)
}
\seealso{
These functions are inspired by dplyr. However, they are designed to only
work on \code{DataFrame} class, and use base R code internally.\preformatted{`help(topic = "select_all", package = "dplyr")`
}
}
